### 2025-05-02 #8152 @editorialbot
[PRE REVIEW]: pytau: A Python package for streamlined changepoint model analysis in neuroscience
**Submitting author:** @abzuarmahmood (<a href="http://orcid.org/0000-0002-2448-4818">Abuzar Mahmood</a>)
**Repository:** https://github.com/abuzarmahmood/pytau
**Branch with paper.md** (empty if default branch): master
**Version:** v0.1
