### 2025-05-07 7f3a20f
feat: implement t-distribution based changepoint model and tests
