### 2025-05-07 85592df
fix: close parenthesis in pytest parametrize decorator
