+++ tests/test_changepoint_model.py
@@ -80,7 +80,7 @@ def test_gen_test_array():
 def test_model_initialization(model_class, data_shape, n_states, extra_args):
     """Test that models can be initialized and generate a model."""
     # Generate test data
-    data_type = (
+    data_type=(
         "normal"
         if model_class
         in [
@@ -90,24 +90,24 @@ def test_model_initialization(model_class, data_shape, n_states, extra_args):
         ]
         else "poisson"
     )
-    test_data = gen_test_array(
+    test_data=gen_test_array(
         data_shape, n_states=3 if n_states is None else n_states, type=data_type
     )
 
     # Initialize model
     if n_states is None:
-        model_instance = model_class(data_array=test_data, **extra_args)
+        model_instance=model_class(data_array=test_data, **extra_args)
     else:
-        model_instance = model_class(
+        model_instance=model_class(
             data_array=test_data, n_states=n_states, **extra_args)
 
     # Check that model can be generated
-    model = model_instance.generate_model()
+    model=model_instance.generate_model()
     assert model is not None
 
 
 # Test the run_all_tests function
-@pytest.mark.slow
+@ pytest.mark.slow
 def test_run_all_tests():
     """Test that run_all_tests can be imported and executed."""
     try:
@@ -130,8 +130,8 @@ def test_extract_inferred_values():
     # Create a mock trace
     class MockTrace:
         def __init__(self):
-            self.varnames = ["tau", "lambda"]
-            self._data = {
+            self.varnames=["tau", "lambda"]
+            self._data={
                 "tau": np.random.rand(10, 5, 2),
                 "lambda": np.random.rand(10, 5, 3),
             }
@@ -139,8 +139,8 @@ def test_extract_inferred_values():
         def __getitem__(self, key):
             return self._data[key]
 
-    trace = MockTrace()
-    result = extract_inferred_values(trace)
+    trace=MockTrace()
+    result=extract_inferred_values(trace)
 
     assert "tau_samples" in result
     assert "lambda_stack" in result
@@ -148,7 +148,7 @@ def test_extract_inferred_values():
     assert result["lambda_stack"].shape == (5, 10, 3)  # Swapped axes
 
 
-@pytest.mark.slow
+@ pytest.mark.slow
 def test_advi_fit():
     """Test the advi_fit function."""
     # This is a more complex test that would require mocking PyMC3
