### 2025-05-17 1845e18
docs: add CONTRIBUTING.md and update README.md with contribution instructions
