+++ CONTRIBUTING.md
@@ -1,0 +1,23 @@
+# Contributing to PyTau
+
+Thank you for considering contributing to PyTau! Here are some guidelines to help you get started:
+
+## Code Style
+- Follow PEP 8 guidelines for Python code.
+- Ensure code is well-documented with comments and docstrings.
+
+## Pull Request Process
+1. Fork the repository.
+2. Create a new branch for your feature or bugfix.
+3. Write clear, concise commit messages.
+4. Submit a pull request with a description of your changes.
+
+## Reporting Issues
+- Use the issue tracker to report bugs or request features.
+- Provide as much detail as possible to help us understand and address the issue.
+
+## Testing
+- Ensure that all tests pass before submitting a pull request.
+- Add new tests for any new features or bug fixes.
+
+We appreciate your contributions and look forward to working with you!
