### 2025-05-22 a968e54
feat: implement TDistributionChangepointModel and add corresponding test
