### 2025-05-22 d2e3213
fix: close parenthesis in parametrize decorator in test_changepoint_model.py
