### 2025-05-22 21de82f
fix: close parenthesis in parametrize decorator in test_changepoint_model.py
