+++ tests/test_changepoint_model.py
@@ -76,7 +76,7 @@ def test_gen_test_array():
         (AllTastePoissonTrialSwitch, (2, 5, 10, 100), 3, {"switch_components": 2}),
         (TDistributionChangepointModel, (10, 100), 3, {}),
         (TDistributionChangepointModel, (10, 100), 3, {}),
-    ],
+    ]
 def test_model_initialization(model_class, data_shape, n_states, extra_args):
     """Test that models can be initialized and generate a model."""
     # Generate test data
