### 2025-05-22 60fc7ea
fix: close parenthesis in pytest parametrize decorator
