+++ pytau/changepoint_model.py
@@ -210,6 +210,7 @@ class GaussianChangepointMeanVar2D(ChangepointModel):
         print("Test for GaussianChangepointMeanVar2D passed")
         return True
 
+
 def stick_breaking(beta):
     portion_remaining = tt.concatenate(
         [[1], tt.extra_ops.cumprod(1 - beta)[:-1]])
