### 2025-06-02 8e29824
feat: implement CategoricalChangepoint3D model for 3D categorical data
