### 2025-06-02 4b4c09c
fix: close parentheses in pytest parametrize decorator to resolve syntax error
