### 2025-06-02 71306d4
fix: close parentheses in parametrize decorator to resolve syntax error
