### 2025-06-02 56c7536
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
