### 2025-06-05 4db8e34
refactor(changepoint_model): replace `testval` with `initval` in model definitions
- Updated multiple instances of `pm.Beta` and `pm.Exponential` to use `initval` instead of `testval` for specifying initial values.
- Ensured models use more consistent initialization parameters.
- Aimed at improving clarity and compatibility with the latest library expectations.
