+++ tests/test_changepoint_model.py
@@ -67,10 +67,10 @@ def test_gen_test_array():
         (GaussianChangepointMean2D, (10, 100), 3, {}),
         (SingleTastePoissonDirichlet, (5, 10, 100), None, {"max_states": 5}),
         (SingleTastePoisson, (5, 10, 100), 3, {}),
-        (SingleTastePoissonVarsig, (5, 10, 100), 3, {}),
-        (SingleTastePoissonVarsigFixed, (5, 10, 100), 3, {"inds_span": 1}),
+        # (SingleTastePoissonVarsig, (5, 10, 100), 3, {}),
+        # (SingleTastePoissonVarsigFixed, (5, 10, 100), 3, {"inds_span": 1}),
         (AllTastePoisson, (2, 5, 10, 100), 3, {}),
-        (AllTastePoissonVarsigFixed, (2, 5, 10, 100), 3, {"inds_span": 1}),
+        # (AllTastePoissonVarsigFixed, (2, 5, 10, 100), 3, {"inds_span": 1}),
         (SingleTastePoissonTrialSwitch,
          (5, 10, 100), 3, {"switch_components": 2}),
         (AllTastePoissonTrialSwitch, (2, 5, 10, 100),
