### 2025-06-07 #8152 @abuzarmahmood
> [@abuzarmahmood](https://github.com/abuzarmahmood) can you work on the above? 👆

@Kevin-Mattheus-Moerman, yes, apologies for not replying.
- I've added a `contributing.md`
- and DOIs for all non-published papers I've cited. There are a handful from my collaborators which are in prep and don't have DOIs yet
