### 2025-06-18 431c01b
feat(changepoint_model): implement 3D categorical changepoint model (WIP)
- Introduce `CategoricalChangepoint3D` class for 3D categorical data modeling.
- Add preliminary input validation for category mapping to continuous integers.
- Integrate PyMC3 with `Dirichlet` and `Categorical` distributions for model setup.
- Preliminary implementation of changepoint inference using ADVI.
- Begin incorporating visualization with `matplotlib` and PyMC3's trace plotting.

Note: This commit is a backup for an incomplete model with nonfunctional features that require further development.
