### 2025-06-18 a42e115
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
