+++ pytau/changepoint_model.py
@@ -1726,14 +1726,15 @@ class AllTastePoissonTrialSwitch(ChangepointModel):
         print("Test for AllTastePoissonTrialSwitch passed")
         return True
 
-cat_p = np.random.rand(5,100)
+
+cat_p = np.random.rand(5, 100)
 cat_p = cat_p / cat_p.sum(axis=0)  # Normalize to sum to 1
 
 with pm.Model() as model:
     category = pm.Categorical(
         "category",
-        p = cat_p, 
-        shape = (10, 5, 100),
+        p=cat_p,
+        shape=(10, 5, 100),
         # shape = (10,100),
         )
 
@@ -1771,7 +1772,8 @@ class CategoricalChangepoint3D(ChangepointModel):
 
         # Make sure data array is int
         if not np.issubdtype(data_array.dtype, np.integer):
-            raise ValueError("Data array must contain integer category values.")
+            raise ValueError(
+                "Data array must contain integer category values.")
 
         # If features in data_array are not continuous integer values, map them
         feature_set = np.unique(data_array)
@@ -1784,8 +1786,8 @@ class CategoricalChangepoint3D(ChangepointModel):
 
         with pm.Model() as model:
             p = pm.Dirichlet(
-                "p", 
-                a=np.ones((n_states, features)), 
+                "p",
+                a=np.ones((n_states, features)),
                 shape=(n_states, features)
             )
 
@@ -1794,9 +1796,9 @@ class CategoricalChangepoint3D(ChangepointModel):
             b_tau = pm.HalfCauchy("b_tau", 3.0, shape=n_states - 1)
             # Shape: trials x changepoints
             tau_latent = pm.Beta(
-                    "tau_latent", 
-                    a_tau, 
-                    b_tau, 
+                    "tau_latent",
+                    a_tau,
+                    b_tau,
                     shape=(trials, n_states - 1)).sort(
                 axis=-1
             )
@@ -1819,7 +1821,7 @@ class CategoricalChangepoint3D(ChangepointModel):
 
             # shape: trials x features x length
             lambda_ = tt.tensordot(weight_stack, p, [
-                                   1, 0])# .swapaxes(1, 2)
+                                   1, 0])  # .swapaxes(1, 2)
 
             flat_lambda = lambda_.reshape(
                 (trials * length, features))
@@ -1829,12 +1831,13 @@ class CategoricalChangepoint3D(ChangepointModel):
             # Use categorical likelihood
             # data_array = trials x length
             # category = pm.Categorical("category", p=lambda_, observed=data_array)
-            category = pm.Categorical("category", p=flat_lambda, observed=flat_data_array)
+            category = pm.Categorical(
+                "category", p=flat_lambda, observed=flat_data_array)
 
         with model:
             inference = pm.ADVI("full-rank")
-            approx = pm.fit(n = 100000, method=inference)
-            trace = approx.sample(draws = 10000)
+            approx = pm.fit(n=100000, method=inference)
+            trace = approx.sample(draws=10000)
 
         # Visualize the model
         pm.plot_trace(trace)
@@ -1856,27 +1859,28 @@ class CategoricalChangepoint3D(ChangepointModel):
 
             # Ensure p is a 3D tensor with shape (trials, features, length)
             if p.ndim != 3 or p.shape[0] != x.shape[0]:
[…]
