### 2025-06-18 efd53fe
feat(changepoint_model): augment 3D categorical model with posterior predictive checks
- Introduce new data visualization for `data_array` with random changes
- Refactor `flat_data_array` initialization within the `ChangepointModel` class
- Enable posterior predictive sampling to evaluate model predictions
- Visualize results by plotting a sample of the posterior predictive category
- Note: Commented-out sections remain unchanged and require further review

This commit serves as a backup for ongoing testing of the categorical model, acknowledging certain parts are incomplete or nonfunctional.
