### 2025-06-18 1697c62
refactor(changepoint-model): simplify and clean up CategoricalChangepoint2D model
- Remove unused code in the CategoricalChangepoint3D class to reduce clutter.
- Rename CategoricalChangepoint3D to CategoricalChangepoint2D for clarity on 2D data array requirement.
- Add validation to ensure data arrays are integers and strictly 2D, enhancing model reliability.
- Streamline the likelihood setup by discarding unnecessary transformations, leading to clearer examples.

This commit focuses on cleaning up the categorical model to improve maintainability.
