### 2025-06-18 a20c303
fix(tests): update test cases for CategoricalChangepoint2D model
- Replaced instances of `CategoricalChangepoint3D` with `CategoricalChangepoint2D` in test cases.
- Adjusted model initialization and comments to reflect the 2D model changes.
