+++ tests/test_changepoint_model.py
@@ -18,6 +18,7 @@ from pytau.changepoint_model import (
     SingleTastePoissonTrialSwitch,
     SingleTastePoissonVarsig,
     SingleTastePoissonVarsigFixed,
+    CategoricalChangepoint2D,
     advi_fit,
     extract_inferred_values,
     gen_test_array,
@@ -75,7 +76,7 @@ def test_gen_test_array():
          (5, 10, 100), 3, {"switch_components": 2}),
         (AllTastePoissonTrialSwitch, (2, 5, 10, 100),
          3, {"switch_components": 2}),
-        (CategoricalChangepoint3D, (5, 10, 100), 3, {}),
+        (CategoricalChangepoint2D, (5, 10, 100), 3, {}),
     ]
 def test_model_initialization(model_class, data_shape, n_states, extra_args):
     """Test that models can be initialized and generate a model."""
@@ -107,9 +108,9 @@ def test_model_initialization(model_class, data_shape, n_states, extra_args):
 
 
 def test_categorical_changepoint_3d():
-    """Test the CategoricalChangepoint3D model."""
+    """Test the CategoricalChangepoint2D model."""
     data=np.random.randint(0, 3, size=(5, 10, 100))
-    model_instance=CategoricalChangepoint3D(data_array=data, n_states=3)
+    model_instance=CategoricalChangepoint2D(data_array=data, n_states=3)
     model=model_instance.generate_model()
     with model:
         inference=pm.ADVI()
