+++ pytau/changepoint_model.py
@@ -1726,6 +1726,7 @@ class AllTastePoissonTrialSwitch(ChangepointModel):
         print("Test for AllTastePoissonTrialSwitch passed")
         return True
 
+
 class CategoricalChangepoint2D(ChangepointModel):
     """Model for categorical data changepoint detection on 2D arrays."""
 
@@ -1742,7 +1743,8 @@ class CategoricalChangepoint2D(ChangepointModel):
         super().__init__(**kwargs)
         # Make sure data array is int
         if not np.issubdtype(data_array.dtype, np.integer):
-            raise ValueError("Data array must contain integer category values.")
+            raise ValueError(
+                "Data array must contain integer category values.")
         # Check that data_array is 2D
         if data_array.ndim != 2:
             raise ValueError("Data array must be 2D (trials x length).")
@@ -1766,21 +1768,14 @@ class CategoricalChangepoint2D(ChangepointModel):
         flat_data_array = data_array.reshape((trials * length,))
 
         with pm.Model() as model:
-            p = pm.Dirichlet(
-                "p",
-                a=np.ones((n_states, features)),
-                shape=(n_states, features)
-            )
+            p = pm.Dirichlet("p", a=np.ones(
+                (n_states, features)), shape=(n_states, features))
 
             # Infer changepoint locations
             a_tau = pm.HalfCauchy("a_tau", 3.0, shape=n_states - 1)
             b_tau = pm.HalfCauchy("b_tau", 3.0, shape=n_states - 1)
             # Shape: trials x changepoints
-            tau_latent = pm.Beta(
-                    "tau_latent",
-                    a_tau,
-                    b_tau,
-                    shape=(trials, n_states - 1)).sort(
+            tau_latent = pm.Beta("tau_latent", a_tau, b_tau, shape=(trials, n_states - 1)).sort(
                 axis=-1
             )
 
@@ -1800,25 +1795,24 @@ class CategoricalChangepoint2D(ChangepointModel):
             #   - weight_stack: trials x states x length
             #   - p : states x features
 
-            # shape: trials x length x features 
-            lambda_ = tt.tensordot(weight_stack, p, [
-                                   1, 0])
+            # shape: trials x length x features
+            lambda_ = tt.tensordot(weight_stack, p, [1, 0])
 
-            flat_lambda = lambda_.reshape(
-                (trials * length, features))
+            flat_lambda = lambda_.reshape((trials * length, features))
 
             # Use categorical likelihood
             # data_array = trials x length
-            category = pm.Categorical("category", p=flat_lambda, observed=flat_data_array)
+            category = pm.Categorical(
+                "category", p=flat_lambda, observed=flat_data_array)
 
     def test(self):
-        test_data=np.random.randint(0, self.n_states, size=(5, 10, 100))
-        test_model=CategoricalChangepoint3D(test_data, self.n_states)
-        model=test_model.generate_model()
+        test_data = np.random.randint(0, self.n_states, size=(5, 10, 100))
+        test_model = CategoricalChangepoint3D(test_data, self.n_states)
+        model = test_model.generate_model()
         with model:
-            inference=pm.ADVI()
-            approx=pm.fit(n=10, method=inference)
-            trace=approx.sample(draws=10)
+            inference = pm.ADVI()
[…]
