### 2025-06-19 fc510a6
fix(changepoint_model): add missing return statement in CategoricalChangepoint2D class
- Added a return statement to exit the context manager with the `model` object in the `CategoricalChangepoint2D` class.
- Ensures the correct functionality and flow within the method.
