### 2025-06-23 #8152 @fboehm
@rly, @neuromusic, @djsaunde - would you be willing to review this submission for JOSS? Thank you.
