### 2025-06-28 #8509 @editorialbot
[REVIEW]: pytau: A Python package for streamlined changepoint model analysis in neuroscience
**Submitting author:** @abzuarmahmood (<a href="http://orcid.org/0000-0002-2448-4818">Abuzar Mahmood</a>)
**Repository:** https://github.com/abuzarmahmood/pytau
**Branch with paper.md** (empty if default branch): master
**Version:** 0.4.1
