### 2025-06-30 b740177
feat: introduce base classes for changepoint, emission, and likelihood logic
Co-authored-by: aider (gpt-4o) <aider@aider.chat>
