### 2025-06-30 2cab92d
fix: resolve undefined name and syntax error in changepoint model tests
Co-authored-by: aider (gpt-4o) <aider@aider.chat>
