### 2025-06-30 abb0649
fix: correct indentation in pytest parametrize decorator
Co-authored-by: aider (gpt-4o) <aider@aider.chat>
