### 2025-06-30 db11759
fix: close parentheses in pytest parametrize decorator
Co-authored-by: aider (gpt-4o) <aider@aider.chat>
