### 2025-06-30 08252d5
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
