### 2025-06-30 52a9c6d
refactor: Introduce base classes for changepoint, emission, and likelihood
Co-authored-by: aider (gpt-4o) <aider@aider.chat>
