### 2025-06-30 09649fb
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
