### 2025-07-15 5faa119
Update README.md
with joss badge
