### 2025-07-17 33162ab
docs: Enhance README with improved structure, visuals, and content
