+++ README.md
@@ -1,138 +1,124 @@
-[![status](https://joss.theoj.org/papers/e3e3d9ce5b59166cef17ee7e9bb9f53c/status.svg)](https://joss.theoj.org/papers/e3e3d9ce5b59166cef17ee7e9bb9f53c)
-[![pre-commit.ci status](https://results.pre-commit.ci/badge/github/abuzarmahmood/pytau/master.svg)](https://results.pre-commit.ci/latest/github/abuzarmahmood/pytau/master)
-[![Pytest](https://github.com/abuzarmahmood/pytau/actions/workflows/pytest_workflow.yml/badge.svg)](https://github.com/abuzarmahmood/pytau/actions/workflows/pytest_workflow.yml)
+<div align="center">
+  <img src="https://raw.githubusercontent.com/abuzarmahmood/pytau/master/docs/logo.png" alt="PyTau Logo" width="200"/>
+  <h1>PyTau</h1>
+  <p><strong>Powerful Changepoint Detection for Neural Data</strong></p>
+  
+  [![status](https://joss.theoj.org/papers/e3e3d9ce5b59166cef17ee7e9bb9f53c/status.svg)](https://joss.theoj.org/papers/e3e3d9ce5b59166cef17ee7e9bb9f53c)
+  [![pre-commit.ci status](https://results.pre-commit.ci/badge/github/abuzarmahmood/pytau/master.svg)](https://results.pre-commit.ci/latest/github/abuzarmahmood/pytau/master)
+  [![Pytest](https://github.com/abuzarmahmood/pytau/actions/workflows/pytest_workflow.yml/badge.svg)](https://github.com/abuzarmahmood/pytau/actions/workflows/pytest_workflow.yml)
+</div>
 
-# PyTau
+## 🚀 What is PyTau?
 
-## API Documentation
-https://abuzarmahmood.github.io/pytau/
+PyTau is a specialized Python package for detecting state changes in neural data using Bayesian changepoint models. It provides:
 
-## Outline
+- **Streamlined batch inference** on PyMC3-based changepoint models
+- **Robust detection** of state transitions in neural firing patterns
+- **Inter-region correlation analysis** for taste-evoked neural responses
+- **Comprehensive visualization tools** for model results
 
-- Perform inter-region transition correlations on models fit to each taste
-    individually
+## 📚 Documentation
 
-## Installation
-```
+[**Full API Documentation**](https://abuzarmahmood.github.io/pytau/)
 
+## ⚡ Quick Start
+
+```bash
 # Create and activate conda environment
 conda create -n "pytau_env" python=3.6.13 ipython notebook -y
 conda activate pytau_env
 
 # Clone repository
-cd ~/Desktop
 git clone https://github.com/abuzarmahmood/pytau.git
 
-# Install requirements from specified file
+# Install requirements
 cd pytau
 pip install -r requirements.txt
 
-# Test everything is working by running notebook
+# Download test data and run example notebook
 cd pytau/how_to
 bash scripts/download_test_data.sh
 cd notebooks
 jupyter notebook
-# Run a notebook
 ```
 
-## Data Organization
+## 🧠 Key Features
 
-- Models stored in CENTRAL LOCATION and accessed by indexing an info file,
-    which also contains model parameters and metadata
-- Metadata also stored in model file to be able to recreate info file in case
-    something happens
+- **Multiple Model Types**:
+  - Single-taste Poisson models
+  - All-taste hierarchical models
+  - Dirichlet process models for automatic state detection
+  - Variable sigmoid models for flexible transition shapes
 
-Database for models fit:
+- **Flexible Data Handling**:
+  - Support for shuffled, simulated, and actual neural data
[…]
