### 2025-07-17 d93e69e
chore: update .gitignore and refine README.md
- Added `.aider*` to `.gitignore` to exclude internal development files.
- Removed detailed information on "Inter-region correlation analysis" and "Model Comparison" from README for a more streamlined presentation.
