### 2025-07-17 74f1169
docs: update README logo path to use local repo image
