### 2025-07-17 c89f746
ci(github): update Python version and OS for pytest workflow
- Updated the OS version from Ubuntu 20.04 to Ubuntu 22.04.
- Changed Python version from 3.6.10 to 3.8.15 in the GitHub workflow configuration.
