+++ .github/workflows/pytest_workflow.yml
@@ -8,10 +8,10 @@ on:
 
 jobs:
   test:
-    runs-on: ubuntu-20.04
+    runs-on: ubuntu-22.04
     strategy:
       matrix:
-        python-version: ['3.6.10']
+        python-version: ['3.8.15']
 
     steps:
     - uses: actions/checkout@v2
