### 2025-07-17 6dfad43
fix(requirements): update Python version compatibility
- Remove specific Python version pinning from requirements.txt
- Extend PYTHON_MAX_VERSION to 3.9 in setup.py to allow compatibility with newer versions of Python 3.
