+++ setup.py
@@ -25,7 +25,7 @@ VERSION = "0.1.1"
 ISRELEASED = False
 
 PYTHON_MIN_VERSION = "3.6.10"
-PYTHON_MAX_VERSION = "3.6.10"
+PYTHON_MAX_VERSION = "3.9"
 PYTHON_REQUIRES = f">={PYTHON_MIN_VERSION}, <={PYTHON_MAX_VERSION}"
 
 INSTALL_REQUIRES = ["pymc3", "numpy", "theano", "tqdm"]
