+++ requirements.txt
@@ -1,4 +1,3 @@
-#python==3.6.13
 arviz==0.11.1
 pymc3==3.9.3
 theano==1.0.5
