### 2025-07-17 47630c9
fix(tests): correct syntax error in test file and remove whitespace
- Fixed a misplaced parenthesis in `test_changepoint_model.py`
- Removed unnecessary whitespace in `README.md`
