### 2025-07-17 3604c76
fix(dependencies): constrain numpy version in requirements
- Changed numpy version constraint from `>=1.19.5` to `<=1.19.5` to possibly address compatibility issues or bugs with newer versions.
