### 2025-07-17 dd8aa43
chore: Update dependencies and bump version to 0.1.2
