### 2025-07-17 8ee7ba4
feat: Add pyproject.toml and update README for modern Python packaging
