+++ .github/workflows/pytest_workflow.yml
@@ -24,7 +24,7 @@ jobs:
         python -m pip install --upgrade pip
         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
         pip install -r requirements-dev.txt
-        pip install -e .
+        pip install -e ../pytau['dev']
     - name: Run tests
       run: |
         pytest --cov=pytau tests/
