+++ .github/workflows/pytest_workflow.yml
@@ -22,8 +22,6 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
-        pip install -r requirements-dev.txt
         pip install -e ../pytau['dev']
     - name: Run tests
       run: |
