### 2025-07-17 590ef80
fix: Resolve 2D/3D data handling in CategoricalChangepoint2D tests
