### 2025-07-17 170cea5
fix: update pre-commit config to ensure hooks allow file modifications
Co-authored-by: aider (gpt-4o) <aider@aider.chat>
