+++ .pre-commit-config.yaml
@@ -14,7 +14,7 @@ repos:
     -   id: debug-statements
     -   id: pretty-format-json
         args:
-            - "--autofix"
+            - "--autofix"  # Ensure this allows automatic fixing
             - "--indent=4"
             - "--no-sort-keys"
 
@@ -36,4 +36,4 @@ repos:
     hooks:
     -   id: autopep8
         args:
-            - "--in-place"
+            - "--in-place"  # Ensure this allows in-place modifications
