### 2025-07-17 f0528a1
Update .pre-commit-config.yaml
remove black
