+++ .pre-commit-config.yaml
@@ -25,12 +25,6 @@ repos:
     -   id: isort
         args: ['--profile', 'black']
 
--   repo: https://github.com/psf/black
-    rev: 25.1.0
-    hooks:
-    -   id: black
-        args: ['--line-length=100']
-
 -   repo: https://github.com/hhatto/autopep8
     rev: v2.3.2  # select the tag or revision you want, or run `pre-commit autoupdate`
     hooks:
