### 2025-08-04 52ffbd4
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
