### 2025-08-04 ebb8d0f
refactor(pymc): update references from pymc3 to pymc
- Changed module reference from `pymc3` to `pymc` in changepoint_io.py
- Updated docstrings and function comments in changepoint_model.py to reflect the change from `pymc3` to `pymc`
- Improved code consistency and accuracy by aligning with the updated `pymc` library naming conventions
