### 2025-08-04 0230d12
chore: update project dependencies and Python version to 3.12
Co-authored-by: aider (anthropic/claude-sonnet-4-20250514) <aider@aider.chat>
