+++ pyproject.toml
@@ -21,7 +21,7 @@ classifiers = [
     "Framework :: tox",
     "Framework :: Pytest",
 ]
-requires-python = "==3.10.13"
+requires-python = ">=3.10.13"
 dependencies = [
     "pymc==5.16.2",
     "numpyro==0.15.1",
