### 2025-08-04 8a142c1
chore(dependencies): update Python version and remove setup.py
- Updated the Python version in the GitHub Actions workflow to 3.10.13.
- Changed `requires-python` in `pyproject.toml` to fix the version to 3.10.13.
- Removed the `setup.py` file as it is no longer needed with `pyproject.toml` being used for package configuration.
- Removed the test dependencies from `pyproject.toml` indicating they should be managed differently.
