+++ .github/workflows/pytest_workflow.yml
@@ -12,7 +12,7 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest, macos-latest, windows-latest]
-        python-version: ['3.8.15', '3.10.13', '3.11', '3.15']
+        python-version: ['3.10.13', '3.11', '3.15']
       fail-fast: false
     steps:
     - uses: actions/checkout@v2
