+++ pyproject.toml
@@ -21,7 +21,7 @@ classifiers = [
     "Framework :: tox",
     "Framework :: Pytest",
 ]
-requires-python = ">=3.12.4"
+requires-python = "==3.10.13"
 dependencies = [
     "pymc==5.16.2",
     "numpyro==0.15.1",
@@ -30,10 +30,6 @@ dependencies = [
 ]
 
 [project.optional-dependencies]
-test = [
-    "pytest",
-    "pytest-cov",
-]
 dev = [
     "pre-commit",
     "pytest",
