### 2025-08-04 9619c1b
fix(changepoint_model): update sigmoid function import for compatibility
- Changed from `tt.nnet.sigmoid` to `tt.math.sigmoid` to ensure compatibility and correct functionality.
