### 2025-08-04 fa245ab
feat: Add comprehensive testing for ADVI, MCMC, and DPP samplers
Co-authored-by: aider (anthropic/claude-sonnet-4-20250514) <aider@aider.chat>
