+++ tests/test_changepoint_model.py
@@ -161,13 +161,247 @@ def test_extract_inferred_values():
 
 @pytest.mark.slow
 def test_advi_fit():
-    """Test the advi_fit function."""
-    # This is a more complex test that would require mocking PyMC3
-    # For now, we'll just check that the function exists and is callable
-    assert callable(advi_fit)
+    """Test the advi_fit function with a simple model."""
+    from pytau.changepoint_model import SingleTastePoisson
+    
+    # Generate test data
+    test_data = gen_test_array((5, 10, 100), n_states=3, type="poisson")
+    
+    # Create a simple model
+    model_instance = SingleTastePoisson(test_data, n_states=3)
+    model = model_instance.generate_model()
+    
+    # Test ADVI fit with minimal iterations
+    result = advi_fit(model, fit=10, samples=5)
+    
+    # Check that we get the expected return values
+    assert len(result) == 5  # model, trace, lambda_stack, tau_samples, observations
+    model_out, trace, lambda_stack, tau_samples, observations = result
+    
+    # Check shapes and types
+    assert tau_samples.shape[0] == 5  # samples
+    assert lambda_stack.shape[1] == 5  # samples
+    assert observations.shape == test_data.shape
 
-    # Skip actual testing as it would require a full PyMC3 model
-    pytest.skip("Full testing of advi_fit requires a PyMC3 model")
+
+@pytest.mark.slow
+def test_mcmc_fit():
+    """Test the mcmc_fit function with a simple model."""
+    from pytau.changepoint_model import mcmc_fit, SingleTastePoisson
+    
+    # Generate test data
+    test_data = gen_test_array((3, 5, 50), n_states=2, type="poisson")  # Smaller for speed
+    
+    # Create a simple model
+    model_instance = SingleTastePoisson(test_data, n_states=2)
+    model = model_instance.generate_model()
+    
+    # Test MCMC fit with minimal samples
+    result = mcmc_fit(model, samples=10)
+    
+    # Check that we get the expected return values
+    assert len(result) == 5  # model, trace, lambda_stack, tau_samples, observations
+    model_out, trace, lambda_stack, tau_samples, observations = result
+    
+    # Check that we have samples
+    assert tau_samples.shape[-1] > 0  # Should have some samples
+    assert lambda_stack.shape[-1] > 0  # Should have some samples
+    assert observations.shape == test_data.shape
+
+
+@pytest.mark.slow
+def test_dpp_fit():
+    """Test the dpp_fit function with a simple model."""
+    from pytau.changepoint_model import dpp_fit, SingleTastePoisson
+    
+    # Generate test data
+    test_data = gen_test_array((3, 5, 50), n_states=2, type="poisson")  # Smaller for speed
+    
+    # Create a simple model
+    model_instance = SingleTastePoisson(test_data, n_states=2)
+    model = model_instance.generate_model()
+    
+    # Test DPP fit with minimal parameters
+    trace = dpp_fit(model, n_chains=2, n_cores=1, tune=10, draws=10)
+    
+    # Check that we get a trace with expected variables
+    assert "tau" in trace.varnames
+    assert "lambda" in trace.varnames
+    
+    # Check that we have the expected number of samples
+    assert trace["tau"].shape[0] == 20  # 2 chains * 10 draws
+
+
+@pytest.mark.slow 
+def test_sampler_comparison():
+    """Test that different samplers produce reasonable results on the same model."""
+    from pytau.changepoint_model import SingleTastePoisson, advi_fit, mcmc_fit
+    
+    # Generate test data with known structure
[…]
