### 2025-08-04 58f1d87
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
