### 2025-08-08 abuzarmahmood/pytau#112 @rly
Support newer versions of Python and numpy
Pytau requires Python version between 3.6.10 and 3.9:
https://github.com/abuzarmahmood/pytau/blob/ea4524c04113588e557359e6e3d96c155987d46c/setup.py#L27-L29

(note: Python 3.9.23 is not compatible with these constraints)

and Pytau requires Numpy version <= 1.19.5:
https://github.com/abuzarmahmood/pytau/blob/ea4524c04113588e557359e6e3d96c155987d46c/setup.py#L38

Python 3.9 will reach end-of-life in October 2025: https://devguide.python.org/versions/
Already, newer versions of scientific python packages have dropped support for Python 3.9.

To maximize compatibility of pytau with other python packages that users may have installed in their environment and want to use in conjunction with pytau, I suggest that support is added for Python 3.9.1-3.13 and newer versions of Numpy (latest is 2.3.2).
