### 2025-08-08 abuzarmahmood/pytau#112 @rly
I am running a Mac with an M1 (arm64) chip. I cannot install the recommended Python 3.6.13 because it is not available for my platform, so I am using a conda environment with the latest supported version of Python allowed, Python 3.9.0.

Unfortunately, I encountered errors building pytables, numpy, and pandas when running `pip install -e .` to install Pytau on my Mac M1, so I installed those packages via conda first. Then, running `pip install -e .` resulted in these pip dependency conflicts:
```
Installing collected packages: typing-extensions, easygui, tqdm, scipy, pyparsing, pillow, patsy, kiwisolver, h5py, fonttools, fastprogress, cycler, contourpy, cftime, theano, netcdf4, matplotlib, xarray, arviz, pymc3, pytau
  Attempting uninstall: typing-extensions
    Found existing installation: typing_extensions 4.14.1
    Uninstalling typing_extensions-4.14.1:
      Successfully uninstalled typing_extensions-4.14.1
ERROR: pip's dependency resolver does not currently take into account all the packages that are installed. This behaviour is the source of the following dependency conflicts.
anyio 4.10.0 requires typing_extensions>=4.5; python_version < "3.13", but you have typing-extensions 3.10.0.2 which is incompatible.
async-lru 2.0.5 requires typing_extensions>=4.0.0; python_version < "3.11", but you have typing-extensions 3.10.0.2 which is incompatible.
beautifulsoup4 4.13.4 requires typing-extensions>=4.0.0, but you have typing-extensions 3.10.0.2 which is incompatible.
exceptiongroup 1.3.0 requires typing-extensions>=4.6.0; python_version < "3.13", but you have typing-extensions 3.10.0.2 which is incompatible.
referencing 0.36.2 requires typing-extensions>=4.4.0; python_version < "3.13", but you have typing-extensions 3.10.0.2 which is incompatible.
Successfully installed arviz-0.11.1 cftime-1.6.4.post1 contourpy-1.1.1 cycler-0.12.1 easygui-0.98.1 fastprogress-1.0.3 fonttools-4.59.0 h5py-3.14.0 kiwisolver-1.4.7 matplotlib-3.6.3 netcdf4-1.7.2 patsy-1.0.1 pillow-11.3.0 pymc3-3.9.3 pyparsing-3.2.3 pytau-0.1.2 scipy-1.10.1 theano-1.0.5 tqdm-4.67.1 typing-extensions-3.10.0.2 xarray-2022.9.0
```

Ignoring these errors and running `jupyter notebook` failed with an error from `jupyter` -> `jsonschema` -> `referencing`:

```
Traceback (most recent call last):
  File "/Users/rly/miniconda3/envs/pytau_env/bin/jupyter-notebook", line 6, in <module>
    from notebook.app import main
  File "/Users/rly/miniconda3/envs/pytau_env/lib/python3.9/site-packages/notebook/app.py", line 18, in <module>
    from jupyter_server.serverapp import flags
  File "/Users/rly/miniconda3/envs/pytau_env/lib/python3.9/site-packages/jupyter_server/serverapp.py", line 40, in <module>
    from jupyter_events.logger import EventLogger
  File "/Users/rly/miniconda3/envs/pytau_env/lib/python3.9/site-packages/jupyter_events/__init__.py", line 3, in <module>
    from .logger import EVENTS_METADATA_VERSION, EventLogger
  File "/Users/rly/miniconda3/envs/pytau_env/lib/python3.9/site-packages/jupyter_events/logger.py", line 15, in <module>
    from jsonschema import ValidationError
  File "/Users/rly/miniconda3/envs/pytau_env/lib/python3.9/site-packages/jsonschema/__init__.py", line 13, in <module>
    from jsonschema._format import FormatChecker
  File "/Users/rly/miniconda3/envs/pytau_env/lib/python3.9/site-packages/jsonschema/_format.py", line 11, in <module>
    from jsonschema.exceptions import FormatError
  File "/Users/rly/miniconda3/envs/pytau_env/lib/python3.9/site-packages/jsonschema/exceptions.py", line 15, in <module>
    from referencing.exceptions import Unresolvable as _Unresolvable
  File "/Users/rly/miniconda3/envs/pytau_env/lib/python3.9/site-packages/referencing/__init__.py", line 5, in <module>
    from referencing._core import Anchor, Registry, Resource, Specification
  File "/Users/rly/miniconda3/envs/pytau_env/lib/python3.9/site-packages/referencing/_core.py", line 18, in <module>
    from referencing.typing import URI, Anchor as AnchorType, D, Mapping, Retrieve
  File "/Users/rly/miniconda3/envs/pytau_env/lib/python3.9/site-packages/referencing/typing.py", line 22, in <module>
    D = TypeVar("D", default=Any)
TypeError: __init__() got an unexpected keyword argument 'default'
```

It looks like installing `typing-extensions=3.10.0.2` is incompatible with `referencing`: `referencing=0.36.0` was installed but was [buggy](https://github.com/python-jsonschema/referencing/pull/211). The fix in 0.36.2 requires `typing-extensions>=4.4.0` but `arviz==0.11.1` requires `typing-extensions<4,>=3.7.4.3`.

Surprisingly, installing `arviz==0.11.1` and `typing-extensions==4.14.4` from conda works but then running `pip install -e .` seems to trigger a resolution of the dependencies of `arviz==0.11.1` and this downgrades `typing-extensions` once again:
```
Obtaining file:///Users/rly/Documents/Paper%20Reviews/JOSS/pytau/pytau
  Installing build dependencies ... done
  Checking if build backend supports build_editable ... done
  Getting requirements to build editable ... done
  Preparing editable metadata (pyproject.toml) ... done
Requirement already satisfied: arviz==0.11.1 in /Users/rly/miniconda3/envs/pytau_env/lib/python3.9/site-packages (from pytau==0.1.2) (0.11.1)
Collecting pymc3==3.9.3 (from pytau==0.1.2)
  Using cached pymc3-3.9.3-py3-none-any.whl.metadata (14 kB)
Collecting theano==1.0.5 (from pytau==0.1.2)
  Using cached theano-1.0.5-py3-none-any.whl
Collecting easygui==0.98.1 (from pytau==0.1.2)
  Using cached easygui-0.98.1-py2.py3-none-any.whl.metadata (4.1 kB)
Requirement already satisfied: tables==3.6.1 in /Users/rly/miniconda3/envs/pytau_env/lib/python3.9/site-packages (from pytau==0.1.2) (3.6.1)
Collecting tqdm>=4.62.3 (from pytau==0.1.2)
  Using cached tqdm-4.67.1-py3-none-any.whl.metadata (57 kB)
Requirement already satisfied: numpy<=1.19.5 in /Users/rly/miniconda3/envs/pytau_env/lib/python3.9/site-packages (from pytau==0.1.2) (1.19.5)
Requirement already satisfied: pandas>=1.1.5 in /Users/rly/miniconda3/envs/pytau_env/lib/python3.9/site-packages (from pytau==0.1.2) (1.3.3)
Requirement already satisfied: matplotlib>=3.3.4 in /Users/rly/miniconda3/envs/pytau_env/lib/python3.9/site-packages (from pytau==0.1.2) (3.5.3)
Requirement already satisfied: setuptools>=38.4 in /Users/rly/miniconda3/envs/pytau_env/lib/python3.9/site-packages (from arviz==0.11.1->pytau==0.1.2) (59.8.0)
Requirement already satisfied: scipy>=0.19 in /Users/rly/miniconda3/envs/pytau_env/lib/python3.9/site-packages (from arviz==0.11.1->pytau==0.1.2) (1.9.1)
Requirement already satisfied: packaging in /Users/rly/miniconda3/envs/pytau_env/lib/python3.9/site-packages (from arviz==0.11.1->pytau==0.1.2) (25.0)
Requirement already satisfied: xarray>=0.16.1 in /Users/rly/miniconda3/envs/pytau_env/lib/python3.9/site-packages (from arviz==0.11.1->pytau==0.1.2) (2022.9.0)
Requirement already satisfied: netcdf4 in /Users/rly/miniconda3/envs/pytau_env/lib/python3.9/site-packages (from arviz==0.11.1->pytau==0.1.2) (1.6.0)
Collecting typing-extensions<4,>=3.7.4.3 (from arviz==0.11.1->pytau==0.1.2)
  Using cached typing_extensions-3.10.0.2-py3-none-any.whl.metadata (2.1 kB)
Collecting patsy>=0.5.1 (from pymc3==3.9.3->pytau==0.1.2)
  Using cached patsy-1.0.1-py2.py3-none-any.whl.metadata (3.3 kB)
Collecting fastprogress>=0.2.0 (from pymc3==3.9.3->pytau==0.1.2)
  Using cached fastprogress-1.0.3-py3-none-any.whl.metadata (5.6 kB)
Collecting h5py>=2.7.0 (from pymc3==3.9.3->pytau==0.1.2)
  Using cached h5py-3.14.0-cp39-cp39-macosx_11_0_arm64.whl.metadata (2.7 kB)
Requirement already satisfied: numexpr>=2.6.2 in /Users/rly/miniconda3/envs/pytau_env/lib/python3.9/site-packages (from tables==3.6.1->pytau==0.1.2) (2.10.0)
Requirement already satisfied: six>=1.9.0 in /Users/rly/.local/lib/python3.9/site-packages (from theano==1.0.5->pytau==0.1.2) (1.16.0)
Requirement already satisfied: cycler>=0.10 in /Users/rly/miniconda3/envs/pytau_env/lib/python3.9/site-packages (from matplotlib>=3.3.4->pytau==0.1.2) (0.12.1)
Requirement already satisfied: fonttools>=4.22.0 in /Users/rly/miniconda3/envs/pytau_env/lib/python3.9/site-packages (from matplotlib>=3.3.4->pytau==0.1.2) (4.59.0)
Requirement already satisfied: kiwisolver>=1.0.1 in /Users/rly/miniconda3/envs/pytau_env/lib/python3.9/site-packages (from matplotlib>=3.3.4->pytau==0.1.2) (1.4.7)
Requirement already satisfied: pillow>=6.2.0 in /Users/rly/miniconda3/envs/pytau_env/lib/python3.9/site-packages (from matplotlib>=3.3.4->pytau==0.1.2) (9.4.0)
Requirement already satisfied: pyparsing>=2.2.1 in /Users/rly/miniconda3/envs/pytau_env/lib/python3.9/site-packages (from matplotlib>=3.3.4->pytau==0.1.2) (3.2.3)
Requirement already satisfied: python-dateutil>=2.7 in /Users/rly/.local/lib/python3.9/site-packages (from matplotlib>=3.3.4->pytau==0.1.2) (2.8.2)
Requirement already satisfied: pytz>=2017.3 in /Users/rly/miniconda3/envs/pytau_env/lib/python3.9/site-packages (from pandas>=1.1.5->pytau==0.1.2) (2025.2)
Requirement already satisfied: cftime in /Users/rly/miniconda3/envs/pytau_env/lib/python3.9/site-packages (from netcdf4->arviz==0.11.1->pytau==0.1.2) (1.6.4)
Using cached easygui-0.98.1-py2.py3-none-any.whl (90 kB)
Using cached pymc3-3.9.3-py3-none-any.whl (1.9 MB)
Using cached typing_extensions-3.10.0.2-py3-none-any.whl (26 kB)
Using cached fastprogress-1.0.3-py3-none-any.whl (12 kB)
Using cached h5py-3.14.0-cp39-cp39-macosx_11_0_arm64.whl (2.8 MB)
Using cached patsy-1.0.1-py2.py3-none-any.whl (232 kB)
Using cached tqdm-4.67.1-py3-none-any.whl (78 kB)
Building wheels for collected packages: pytau
  Building editable for pytau (pyproject.toml) ... done
  Created wheel for pytau: filename=pytau-0.1.2-0.editable-py3-none-any.whl size=5874 sha256=4eebe9fa21ead07ca8165b839a7e7e5fe01702b813e8cb4d49d6256e657f57da
  Stored in directory: /private/var/folders/cp/nh4x2lls1qq03jh58crvh8t00000gp/T/pip-ephem-wheel-cache-kasdbaek/wheels/8a/2f/9a/4d8dc310f22f66602255b4650f4ab2e761d9c721ecdf4f6a84
Successfully built pytau
Installing collected packages: typing-extensions, easygui, tqdm, patsy, h5py, fastprogress, theano, pymc3, pytau
  Attempting uninstall: typing-extensions
    Found existing installation: typing_extensions 4.14.1
    Uninstalling typing_extensions-4.14.1:
      Successfully uninstalled typing_extensions-4.14.1
ERROR: pip's dependency resolver does not currently take into account all the packages that are installed. This behaviour is the source of the following dependency conflicts.
anyio 4.10.0 requires typing_extensions>=4.5; python_version < "3.13", but you have typing-extensions 3.10.0.2 which is incompatible.
async-lru 2.0.5 requires typing_extensions>=4.0.0; python_version < "3.11", but you have typing-extensions 3.10.0.2 which is incompatible.
beautifulsoup4 4.13.4 requires typing-extensions>=4.0.0, but you have typing-extensions 3.10.0.2 which is incompatible.
exceptiongroup 1.3.0 requires typing-extensions>=4.6.0; python_version < "3.13", but you have typing-extensions 3.10.0.2 which is incompatible.
referencing 0.36.2 requires typing-extensions>=4.4.0; python_version < "3.13", but you have typing-extensions 3.10.0.2 which is incompatible.
Successfully installed easygui-0.98.1 fastprogress-1.0.3 h5py-3.14.0 patsy-1.0.1 pymc3-3.9.3 pytau-0.1.2 theano-1.0.5 tqdm-4.67.1 typing-extensions-3.10.0.2
```

Forcing an upgrade of `typing-extensions` results in:
```
Requirement already satisfied: typing-extensions in /Users/rly/miniconda3/envs/pytau_env/lib/python3.9/site-packages (3.10.0.2)
Collecting typing-extensions
  Using cached typing_extensions-4.14.1-py3-none-any.whl.metadata (3.0 kB)
Using cached typing_extensions-4.14.1-py3-none-any.whl (43 kB)
Installing collected packages: typing-extensions
  Attempting uninstall: typing-extensions
    Found existing installation: typing-extensions 3.10.0.2
    Uninstalling typing-extensions-3.10.0.2:
      Successfully uninstalled typing-extensions-3.10.0.2
ERROR: pip's dependency resolver does not currently take into account all the packages that are installed. This behaviour is the source of the following dependency conflicts.
arviz 0.11.1 requires typing-extensions<4,>=3.7.4.3, but you have typing-extensions 4.14.1 which is incompatible.
Successfully installed typing-extensions-4.14.1
```

With this, and ignoring the arviz & typing-extensions incompatibility error, I can finally run `jupyter notebook` successfully.

But then when running the notebook `PyTau_Walkthrough_no_handlers.ipynb`, on cell 3 (imports), I get this error:
```
WARNING (theano.configdefaults): install mkl with `conda install mkl-service`: No module named 'mkl'
Cell In[3], line 10
      8 sys.path.append(pytau_base_dir)
      9 from pytau.utils import plotting
---> 10 from pytau.changepoint_model import single_taste_poisson, advi_fit
     11 from pytau.changepoint_preprocess import preprocess_single_taste

File ~/Documents/Paper Reviews/JOSS/pytau/pytau/pytau/changepoint_model.py:14
      9 import numpy as np
     11 ########################################
     12 # Import
     13 ########################################
---> 14 import pymc3 as pm
     15 import theano
     16 import theano.tensor as tt

File ~/miniconda3/envs/pytau_env/lib/python3.9/site-packages/pymc3/__init__.py:41
     38 __set_compiler_flags()
     40 from .blocking import *
---> 41 from .distributions import *
     42 from .distributions import transforms
     43 from .glm import *

File ~/miniconda3/envs/pytau_env/lib/python3.9/site-packages/pymc3/distributions/__init__.py:15
      1 #   Copyright 2020 The PyMC Developers
      2 #
      3 #   Licensed under the Apache License, Version 2.0 (the "License");
   (...)
     12 #   See the License for the specific language governing permissions and
     13 #   limitations under the License.
---> 15 from . import timeseries
     16 from . import transforms
     17 from . import shape_utils

File ~/miniconda3/envs/pytau_env/lib/python3.9/site-packages/pymc3/distributions/timeseries.py:23
     20 import numpy as np
     22 from pymc3.util import get_variable_name
---> 23 from .continuous import get_tau_sigma, Normal, Flat
     24 from .shape_utils import to_tuple
     25 from . import multivariate

File ~/miniconda3/envs/pytau_env/lib/python3.9/site-packages/pymc3/distributions/continuous.py:32
     30 from .special import log_i0
     31 from ..math import invlogit, logit, logdiffexp
---> 32 from .dist_math import (
     33     alltrue_elemwise, betaln, bound, gammaln, i0e, incomplete_beta, logpow,
     34     normal_lccdf, normal_lcdf, SplineWrapper, std_cdf, zvalue,
     35     clipped_beta_rvs,
     36 )
     37 from .distribution import (Continuous, draw_values, generate_samples)
     39 __all__ = ['Uniform', 'Flat', 'HalfFlat', 'Normal', 'TruncatedNormal', 'Beta',
     40            'Kumaraswamy', 'Exponential', 'Laplace', 'StudentT', 'Cauchy',
     41            'HalfCauchy', 'Gamma', 'Weibull', 'HalfStudentT', 'Lognormal',
     42            'ChiSquared', 'HalfNormal', 'Wald', 'Pareto', 'InverseGamma',
     43            'ExGaussian', 'VonMises', 'SkewNormal', 'Triangular', 'Gumbel',
     44            'Logistic', 'LogitNormal', 'Interpolated', 'Rice', 'Moyal']

File ~/miniconda3/envs/pytau_env/lib/python3.9/site-packages/pymc3/distributions/dist_math.py:44
     38 _beta_clip_values = {
     39     dtype: (np.nextafter(0, 1, dtype=dtype), np.nextafter(1, 0, dtype=dtype))
     40     for dtype in ["float16", "float32", "float64"]
     41 }
     42 if platform.system() in ["Linux", "Darwin"]:
     43     _beta_clip_values["float128"] = (
---> 44         np.nextafter(0, 1, dtype="float128"),
     45         np.nextafter(1, 0, dtype="float128")
     46     )
     50 def bound(logp, *conditions, **kwargs):
     51     """
     52     Bounds a log probability density with several conditions.
     53
   (...)
     67     logp with elements set to -inf where any condition is False
     68     """

TypeError: data type 'float128' not understood
```

So it looks like the pinned version of `pymc3==3.9.3` is not compatible with Mac arm64 platforms. See https://github.com/pymc-devs/pymc-resources/issues/90 . I could downgrade pymc to 3.8 as suggested in that issue ticket but I don't know if that would break other things.

Unfortunately I have been troubleshooting installation issues for an hour and am giving up and cannot continue my JOSS review... Could you please provide guidance on installing pytau on a Mac with an Apple Silicon chip, or look into supporting newer versions of Python and other dependencies? Thank you!
