### 2025-08-15 c5ca46e
ci: update GitHub workflow to use matrix for OS and Python versions
Co-authored-by: aider (gpt-4o) <aider@aider.chat>
