### 2025-08-15 0304166
test: add unit tests for functions in changepoint_preprocess.py
Co-authored-by: aider (gpt-4o) <aider@aider.chat>
