### 2025-08-15 2565743
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
