### 2025-08-15 1bda2f5
test: add unit tests for FitHandler and DatabaseHandler classes
Co-authored-by: aider (gpt-4o) <aider@aider.chat>
