### 2025-08-15 5a0a613
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
