### 2025-08-15 d6138e8
test: Add unit tests for changepoint_analysis.py module functions and classes
Co-authored-by: aider (gpt-4o) <aider@aider.chat>
