### 2025-08-15 ee498f3
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
