### 2025-08-15 #8509 @rly
I am working with the author to resolve issues with installing pytau on my computer here: https://github.com/abuzarmahmood/pytau/issues/112 . Unfortunately I cannot proceed with my review until then.
