### 2025-09-18 feed2f9
refactor(changepoint): improve model fitting and remove unused functions
- Removed unnecessary `theano_lock_present` and `compile_wait` functions.
- Added an optional `convergence_tol` parameter in `find_best_states` and `advi_fit` to enable convergence checks.
- Integrated the `CheckParametersConvergence` callback from pymc for model fitting with a specified tolerance.
- Fixed data type issue in `model_generator` by explicitly converting `n_states` to an integer.
- Simplified the `advi_fit` return structure and removed commented-out sample extraction logic.
