### 2025-09-18 5156f12
fix: replace deprecated np.int with astype(int) for NumPy 2.x compatibility
- Replace np.vectorize(np.int) with spike_binned.astype(int) in both functions
- Fixes AttributeError: module 'numpy' has no attribute 'int'
- Maintains same functionality while supporting modern NumPy versions

Co-authored-by: Ona <no-reply@ona.com>
