### 2025-09-18 f0b7c95
feat: implement comprehensive tests for changepoint_preprocess.py
Expand basic shape tests into comprehensive test suite with 22 test cases:

- TestPreprocessSingleTaste: 9 tests covering shape validation, data preservation,
  transform functionality, parameter variations, and error handling
- TestPreprocessAllTaste: 10 tests with similar coverage plus consistency checks
- TestEdgeCases: 3 tests for edge cases and reproducibility

Key improvements:
- Test all transform types (trial_shuffled, spike_shuffled, simulated, None)
- Validate data integrity and binning correctness
- Test different tim
[…]
