### 2025-09-18 4ffff23
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
