+++ tests/test_changepoint_preprocess.py
@@ -16,9 +16,10 @@ class TestPreprocessSingleTaste:
 
     def test_shape_output(self):
         """Test that output shape is correct for all transforms."""
-        transforms = ['trial_shuffled', 'spike_shuffled', 'simulated', None, 'None']
+        transforms = ['trial_shuffled',
+                      'spike_shuffled', 'simulated', None, 'None']
         expected_shape = (10, 5, 10)
-        
+
         for transform in transforms:
             result = preprocess_single_taste(
                 self.spike_array, self.time_lims, self.bin_width, transform)
@@ -30,10 +31,10 @@ class TestPreprocessSingleTaste:
             self.spike_array, self.time_lims, self.bin_width, None)
         result_none_str = preprocess_single_taste(
             self.spike_array, self.time_lims, self.bin_width, 'None')
-        
+
         # Both None and 'None' should give same result
         np.testing.assert_array_equal(result_none, result_none_str)
-        
+
         # Check that binning is correct
         expected = np.sum(
             self.spike_array[..., self.time_lims[0]:self.time_lims[1]].reshape(
@@ -47,11 +48,11 @@ class TestPreprocessSingleTaste:
         np.random.seed(42)
         original = preprocess_single_taste(
             self.spike_array, self.time_lims, self.bin_width, None)
-        
+
         np.random.seed(123)  # Different seed
         shuffled = preprocess_single_taste(
             self.spike_array, self.time_lims, self.bin_width, 'trial_shuffled')
-        
+
         # Should be different (with high probability)
         assert not np.array_equal(original, shuffled)
 
@@ -60,11 +61,11 @@ class TestPreprocessSingleTaste:
         np.random.seed(42)
         original = preprocess_single_taste(
             self.spike_array, self.time_lims, self.bin_width, None)
-        
+
         np.random.seed(123)  # Different seed
         shuffled = preprocess_single_taste(
             self.spike_array, self.time_lims, self.bin_width, 'spike_shuffled')
-        
+
         # Should be different (with high probability)
         assert not np.array_equal(original, shuffled)
 
@@ -73,10 +74,10 @@ class TestPreprocessSingleTaste:
         # Use data with clear firing patterns
         spike_array = np.zeros((5, 3, 100))
         spike_array[:, 0, 10:20] = 1  # High firing in specific window
-        
+
         result = preprocess_single_taste(
             spike_array, self.time_lims, self.bin_width, 'simulated')
-        
+
         # Should be non-negative integers (binned spike counts)
         assert np.all(result >= 0)
         assert np.issubdtype(result.dtype, np.integer)
@@ -92,7 +93,7 @@ class TestPreprocessSingleTaste:
         time_lims = [10, 90]
         result = preprocess_single_taste(
             self.spike_array, time_lims, self.bin_width, None)
-        
+
         expected_shape = (10, 5, 8)  # (90-10)/10 = 8 bins
         assert result.shape == expected_shape
 
@@ -100,7 +101,7 @@ class TestPreprocessSingleTaste:
         """Test with different bin widths."""
         bin_widths = [5, 20, 25]
         expected_bins = [20, 5, 4]  # 100/bin_width
-        
+
         for bin_width, expected_bin_count in zip(bin_widths, expected_bins):
             result = preprocess_single_taste(
                 self.spike_array, self.time_lims, bin_width, None)
@@ -125,9 +126,10 @@ class TestPreprocessAllTaste:
 
     def test_shape_output(self):
         """Test that output shape is correct for all transforms."""
-        transforms = ['trial_shuffled', 'spike_shuffled', 'simulated', None, 'None']
+        transforms = ['trial_shuffled',
[…]
