### 2025-09-18 38346e9
test: Add comprehensive tests for changepoint_preprocess.py and fix NumPy compatibility
Co-authored-by: aider (gpt-4o) <aider@aider.chat>
