+++ tests/test_changepoint_preprocess.py
@@ -1,284 +1,31 @@
-import numpy as np
 import pytest
-
-from pytau.changepoint_preprocess import preprocess_all_taste, preprocess_single_taste
-
+import numpy as np
+from pytau.changepoint_preprocess import preprocess_single_taste, preprocess_all_taste
 
 class TestPreprocessSingleTaste:
-    """Test suite for preprocess_single_taste function."""
-
     def setup_method(self):
-        """Set up test fixtures."""
-        np.random.seed(42)  # For reproducible tests
-        self.spike_array = np.random.randint(0, 2, (10, 5, 100))
-        self.time_lims = [0, 100]
-        self.bin_width = 10
+        # Setup code here
+        self.spike_array = np.random.randint(0, 2, (10, 5, 1000))
+        self.time_lims = [200, 800]
+        self.bin_width = 50
 
-    def test_shape_output(self):
-        """Test that output shape is correct for all transforms."""
-        transforms = ['trial_shuffled',
-                      'spike_shuffled', 'simulated', None, 'None']
-        expected_shape = (10, 5, 10)
-
-        for transform in transforms:
-            result = preprocess_single_taste(
-                self.spike_array, self.time_lims, self.bin_width, transform)
-            assert result.shape == expected_shape, f"Failed for transform: {transform}"
-
-    def test_none_transform_preserves_data(self):
-        """Test that None transform preserves original data after binning."""
-        result_none = preprocess_single_taste(
-            self.spike_array, self.time_lims, self.bin_width, None)
-        result_none_str = preprocess_single_taste(
-            self.spike_array, self.time_lims, self.bin_width, 'None')
-
-        # Both None and 'None' should give same result
-        np.testing.assert_array_equal(result_none, result_none_str)
-
-        # Check that binning is correct
-        expected = np.sum(
-            self.spike_array[..., self.time_lims[0]:self.time_lims[1]].reshape(
-                *self.spike_array.shape[:-1], -1, self.bin_width
-            ), axis=-1
-        )
-        np.testing.assert_array_equal(result_none, expected)
-
-    def test_trial_shuffled_different_from_original(self):
-        """Test that trial shuffled data is different from original."""
-        np.random.seed(42)
-        original = preprocess_single_taste(
-            self.spike_array, self.time_lims, self.bin_width, None)
-
-        np.random.seed(123)  # Different seed
-        shuffled = preprocess_single_taste(
-            self.spike_array, self.time_lims, self.bin_width, 'trial_shuffled')
-
-        # Should be different (with high probability)
-        assert not np.array_equal(original, shuffled)
-
-    def test_spike_shuffled_different_from_original(self):
-        """Test that spike shuffled data is different from original."""
-        np.random.seed(42)
-        original = preprocess_single_taste(
-            self.spike_array, self.time_lims, self.bin_width, None)
-
-        np.random.seed(123)  # Different seed
-        shuffled = preprocess_single_taste(
-            self.spike_array, self.time_lims, self.bin_width, 'spike_shuffled')
-
-        # Should be different (with high probability)
-        assert not np.array_equal(original, shuffled)
-
-    def test_simulated_data_properties(self):
-        """Test that simulated data has reasonable properties."""
-        # Use data with clear firing patterns
-        spike_array = np.zeros((5, 3, 100))
-        spike_array[:, 0, 10:20] = 1  # High firing in specific window
-
-        result = preprocess_single_taste(
-            spike_array, self.time_lims, self.bin_width, 'simulated')
-
-        # Should be non-negative integers (binned spike counts)
-        assert np.all(result >= 0)
-        assert np.issubdtype(result.dtype, np.integer)
-
[…]
