### 2025-09-18 22635c3
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
