### 2025-09-18 0a52866
fix(ci): resolve CI/CD failures by updating Python version constraints
- Remove Python 3.13 from test matrix (dependency compatibility issues)
- Fix installation command syntax: '../pytau[dev]' -> '.[dev]'
- Update Python version constraints to exclude 3.13 (>=3.10,<3.13)
- Add explicit Python version classifiers for 3.10, 3.11, 3.12

Fixes:
- PyTensor version gap (2.25.x missing for Python 3.13)
- NumPy compilation issues on Python 3.13
- Incorrect pip install path in workflow

Tested locally with Python 3.12.3 - all 22 tests pass with 100% coverage

Co-authored-by: Ona <no-reply@ona.com>
