+++ .github/workflows/pytest_workflow.yml
@@ -12,7 +12,7 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest, macos-latest, windows-latest]
-        python-version: ['3.10.13', '3.11', '3.13']
+        python-version: ['3.10', '3.11', '3.12']
       fail-fast: false
     steps:
     - uses: actions/checkout@v2
@@ -23,7 +23,7 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install -e ../pytau['dev']
+        pip install -e .[dev]
     - name: Run tests
       run: |
         pytest --cov=pytau tests/
