+++ pyproject.toml
@@ -13,6 +13,8 @@ authors = [
 license = {text = "MIT"}
 classifiers = [
     "License :: OSI Approved",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
     "Intended Audience :: Science/Research",
     "Topic :: Scientific/Engineering",
@@ -21,7 +23,7 @@ classifiers = [
     "Framework :: tox",
     "Framework :: Pytest",
 ]
-requires-python = ">=3.10.13"
+requires-python = ">=3.10,<3.13"
 dependencies = [
     "pymc==5.16.2",
     "numpyro==0.15.1",
