+++ .github/workflows/pytest_workflow.yml
@@ -26,4 +26,4 @@ jobs:
         pip install -e .[dev]
     - name: Run tests
       run: |
-        pytest --cov=pytau tests/
+        pytest --cov=pytau tests/test_changepoint_preprocess.py tests/test_changepoint_model.py -v
