### 2025-09-18 f28fde3
build: update CI/CD workflow and project configuration for compatibility
Co-authored-by: aider (gpt-4o) <aider@aider.chat>
