+++ pyproject.toml
@@ -23,7 +23,7 @@ classifiers = [
     "Framework :: tox",
     "Framework :: Pytest",
 ]
-requires-python = ">=3.10,<3.13"
+requires-python = ">=3.10,<3.12"
 dependencies = [
     "pymc==5.16.2",
     "numpyro==0.15.1",
