+++ .github/workflows/pytest_workflow.yml
@@ -12,7 +12,7 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest, macos-latest, windows-latest]
-        python-version: ['3.10', '3.11', '3.12']
+        python-version: ['3.10', '3.11']
       fail-fast: false
     steps:
     - uses: actions/checkout@v2
