### 2025-09-18 5386d49
Add changepoint_io test file to pytest workflow
