### 2025-09-18 d1dc21d
Add easygui to requirements
