+++ requirements.txt
@@ -3,3 +3,4 @@ pymc==5.16.2
 numpyro==0.15.1
 arviz==0.18.0
 graphviz==0.20.3
+easygui
