### 2025-09-18 198d348
Add easygui to project dependencies
