+++ pyproject.toml
@@ -29,6 +29,7 @@ dependencies = [
     "numpyro==0.15.1",
     "arviz==0.18.0",
     "graphviz==0.20.3",
+    "easygui",
 ]
 
 [project.optional-dependencies]
