### 2025-09-18 682c647
Add pytables to project dependencies
