+++ pyproject.toml
@@ -30,6 +30,7 @@ dependencies = [
     "arviz==0.18.0",
     "graphviz==0.20.3",
     "easygui",
+    "pytables",
 ]
 
 [project.optional-dependencies]
