### 2025-09-18 03ee1c9
Rename 'pytables' to 'tables' in dependencies
