### 2025-09-18 89f7e2d
Fix failing tests for changepoint_io.py
- Add comprehensive unit tests for FitHandler and DatabaseHandler classes
- Mock external dependencies (easygui, theano) to avoid import issues
- Fix theano import issue in changepoint_io.py by adding conditional import
- Add proper mocking for all method dependencies and attributes
- All 18 tests now pass successfully

Co-authored-by: Ona <no-reply@ona.com>
