+++ tests/test_changepoint_io.py
@@ -1,77 +1,376 @@
-import os
-import unittest
+"""
+Tests for the changepoint_io module.
+"""
 
+import os
+import sys
+import tempfile
+import unittest
+from unittest.mock import MagicMock, Mock, patch
+
+import numpy as np
 import pandas as pd
+import pytest
+
+# Mock easygui and theano before importing pytau modules
+sys.modules['easygui'] = Mock()
+mock_theano = Mock()
+mock_theano.__version__ = '1.0.0'
+mock_theano.config = Mock()
+mock_theano.config.compiledir = '/tmp/theano'
+sys.modules['theano'] = mock_theano
 
 from pytau.changepoint_io import DatabaseHandler, FitHandler
 
 
 class TestFitHandler(unittest.TestCase):
-    def setUp(self):
-        self.handler = FitHandler(
+    """Test cases for FitHandler class."""
+
+    def test_initialization_missing_experiment_name(self):
+        """Test that FitHandler raises exception when experiment_name is missing."""
+        with self.assertRaises(Exception) as context:
+            FitHandler(data_dir='path/to/data', taste_num=1, region_name='region')
+        self.assertIn("experiment name", str(context.exception))
+
+    def test_initialization_invalid_laser_type(self):
+        """Test that FitHandler raises exception for invalid laser_type."""
+        with self.assertRaises(Exception) as context:
+            FitHandler(
+                data_dir='path/to/data',
+                taste_num=1,
+                region_name='region',
+                experiment_name='exp',
+                laser_type='invalid'
+            )
+        self.assertIn("laser_type", str(context.exception))
+
+    def test_initialization_invalid_taste_num(self):
+        """Test that FitHandler raises exception for invalid taste_num."""
+        with self.assertRaises(Exception) as context:
+            FitHandler(
+                data_dir='path/to/data',
+                taste_num='invalid',
+                region_name='region',
+                experiment_name='exp'
+            )
+        self.assertIn("taste_num", str(context.exception))
+
+    @patch('pytau.changepoint_io.EphysData')
+    def test_initialization_success(self, mock_ephys_data):
+        """Test successful FitHandler initialization."""
+        mock_ephys_data.return_value = Mock()
+        
+        handler = FitHandler(
             data_dir='path/to/data',
             taste_num=1,
             region_name='region',
             experiment_name='exp'
         )
+        
+        self.assertEqual(handler.data_dir, 'path/to/data')
+        self.assertEqual(handler.taste_num, 1)
+        self.assertEqual(handler.region_name, 'region')
+        self.assertEqual(handler.experiment_name, 'exp')
+        mock_ephys_data.assert_called_once_with('path/to/data')
 
-    def test_initialization(self):
-        with self.assertRaises(Exception):
-            FitHandler(data_dir='path/to/data',
-                       taste_num=1, region_name='region')
+    @patch('pytau.changepoint_io.EphysData')
+    def test_set_preprocess_params(self, mock_ephys_data):
+        """Test setting preprocessing parameters."""
+        mock_ephys_data.return_value = Mock()
+        
+        handler = FitHandler(
+            data_dir='path/to/data',
+            taste_num=1,
+            region_name='region',
+            experiment_name='exp'
+        )
+        
+        handler.set_preprocess_params(
+            time_lims=(0, 100),
+            bin_width=10,
+            data_transform='log'
+        )
+        
+        self.assertEqual(handler.preprocess_params['bin_width'], 10)
+        self.assertEqual(handler.preprocess_params['time_lims'], (0, 100))
+        self.assertEqual(handler.preprocess_params['data_transform'], 'log')
 
[…]
