### 2025-09-18 b887ae5
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
