+++ tests/test_changepoint_io.py
@@ -2,6 +2,7 @@
 Tests for the changepoint_io module.
 """
 
+from pytau.changepoint_io import DatabaseHandler, FitHandler
 import os
 import sys
 import tempfile
@@ -20,8 +21,6 @@ mock_theano.config = Mock()
 mock_theano.config.compiledir = '/tmp/theano'
 sys.modules['theano'] = mock_theano
 
-from pytau.changepoint_io import DatabaseHandler, FitHandler
-
 
 class TestFitHandler(unittest.TestCase):
     """Test cases for FitHandler class."""
@@ -29,7 +28,8 @@ class TestFitHandler(unittest.TestCase):
     def test_initialization_missing_experiment_name(self):
         """Test that FitHandler raises exception when experiment_name is missing."""
         with self.assertRaises(Exception) as context:
-            FitHandler(data_dir='path/to/data', taste_num=1, region_name='region')
+            FitHandler(data_dir='path/to/data',
+                       taste_num=1, region_name='region')
         self.assertIn("experiment name", str(context.exception))
 
     def test_initialization_invalid_laser_type(self):
@@ -59,14 +59,14 @@ class TestFitHandler(unittest.TestCase):
     def test_initialization_success(self, mock_ephys_data):
         """Test successful FitHandler initialization."""
         mock_ephys_data.return_value = Mock()
-        
+
         handler = FitHandler(
             data_dir='path/to/data',
             taste_num=1,
             region_name='region',
             experiment_name='exp'
         )
-        
+
         self.assertEqual(handler.data_dir, 'path/to/data')
         self.assertEqual(handler.taste_num, 1)
         self.assertEqual(handler.region_name, 'region')
@@ -77,20 +77,20 @@ class TestFitHandler(unittest.TestCase):
     def test_set_preprocess_params(self, mock_ephys_data):
         """Test setting preprocessing parameters."""
         mock_ephys_data.return_value = Mock()
-        
+
         handler = FitHandler(
             data_dir='path/to/data',
             taste_num=1,
             region_name='region',
             experiment_name='exp'
         )
-        
+
         handler.set_preprocess_params(
             time_lims=(0, 100),
             bin_width=10,
             data_transform='log'
         )
-        
+
         self.assertEqual(handler.preprocess_params['bin_width'], 10)
         self.assertEqual(handler.preprocess_params['time_lims'], (0, 100))
         self.assertEqual(handler.preprocess_params['data_transform'], 'log')
@@ -102,14 +102,14 @@ class TestFitHandler(unittest.TestCase):
         test_data = np.random.poisson(1, (5, 10, 100))
         mock_ephys_instance.return_region_spikes.return_value = test_data
         mock_ephys_data.return_value = mock_ephys_instance
-        
+
         handler = FitHandler(
             data_dir='path/to/data',
             taste_num=1,
             region_name='region',
             experiment_name='exp'
         )
-        
+
         handler.load_spike_trains()
         self.assertIsNotNone(handler.data)
         mock_ephys_instance.return_region_spikes.assert_called_once()
@@ -118,7 +118,7 @@ class TestFitHandler(unittest.TestCase):
     def test_preprocess_data(self, mock_ephys_data):
         """Test data preprocessing."""
         mock_ephys_data.return_value = Mock()
-        
+
         handler = FitHandler(
             data_dir='path/to/data',
             taste_num=1,
@@ -126,14 +126,14 @@ class TestFitHandler(unittest.TestCase):
             experiment_name='exp'
         )
         handler.data = np.random.poisson(1, (5, 10, 100))
-        
+
         # Mock the preprocessor
         mock_preprocessor = Mock()
         mock_preprocessor.return_value = np.random.poisson(1, (5, 50))
[…]
