### 2025-09-18 85cc0bc
Remove obsolete compile_wait calls from changepoint_io.py
- Remove calls to compile_wait() which was removed in recent refactor
- Update tests to remove compile_wait patches
- Fixes CI/CD test failures in GitHub Actions
- All 18 tests now pass successfully

Co-authored-by: Ona <no-reply@ona.com>
