+++ tests/test_changepoint_io.py
@@ -2,7 +2,6 @@
 Tests for the changepoint_io module.
 """
 
-from pytau.changepoint_io import DatabaseHandler, FitHandler
 import os
 import sys
 import tempfile
@@ -20,6 +19,8 @@ mock_theano.__version__ = '1.0.0'
 mock_theano.config = Mock()
 mock_theano.config.compiledir = '/tmp/theano'
 sys.modules['theano'] = mock_theano
+
+from pytau.changepoint_io import DatabaseHandler, FitHandler
 
 
 class TestFitHandler(unittest.TestCase):
@@ -139,12 +140,10 @@ class TestFitHandler(unittest.TestCase):
         mock_preprocessor.assert_called_once()
 
     @patch('pytau.changepoint_io.EphysData')
-    @patch('pytau.changepoint_model.compile_wait')
-    def test_create_model(self, mock_compile_wait, mock_ephys_data):
+    def test_create_model(self, mock_ephys_data):
         """Test model creation."""
         mock_ephys_data.return_value = Mock()
-        mock_compile_wait.return_value = None
-
+        
         handler = FitHandler(
             data_dir='path/to/data',
             taste_num=1,
@@ -169,12 +168,10 @@ class TestFitHandler(unittest.TestCase):
         mock_model_template.assert_called_once()
 
     @patch('pytau.changepoint_io.EphysData')
-    @patch('pytau.changepoint_model.compile_wait')
-    def test_run_inference(self, mock_compile_wait, mock_ephys_data):
+    def test_run_inference(self, mock_ephys_data):
         """Test running inference."""
         mock_ephys_data.return_value = Mock()
-        mock_compile_wait.return_value = None
-
+        
         handler = FitHandler(
             data_dir='path/to/data',
             taste_num=1,
