### 2025-09-18 d6a2b40
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
