### 2025-09-18 110b20f
Simplify pytest command in workflow
Updated pytest command to run all tests with coverage.
