### 2025-09-18 3c0234d
fix: replace deprecated np.int with int and enhance test coverage
- Fix deprecated np.int usage in changepoint_analysis.py (lines 71, 187, 191)
- Add comprehensive unit tests for all functions and classes:
  - get_transition_snips() with edge cases and error handling
  - get_state_firing() with different input shapes
  - calc_significant_neurons_firing() with various p-values
  - calc_significant_neurons_snippets() with statistical validation
  - _tau class initialization and attribute testing
  - _firing class structure and method signature testing
  - PklHandler class basic functionality tes
[…]
