+++ tests/test_changepoint_analysis.py
@@ -1,6 +1,9 @@
 import os
+import tempfile
+from unittest.mock import patch
 
 import numpy as np
+import pytest
 
 from pytau.changepoint_analysis import (
     PklHandler,
@@ -14,59 +17,217 @@ from pytau.changepoint_analysis import (
 
 
 def test_get_transition_snips():
+    """Test get_transition_snips function with valid inputs."""
+    np.random.seed(42)  # For reproducible tests
     spike_array = np.random.rand(5, 10, 1000)  # Example spike array
     tau_array = np.array(
         [[100, 300], [200, 400], [150, 350], [250, 450], [300, 500]])
     window_radius = 50
     result = get_transition_snips(spike_array, tau_array, window_radius)
     assert result.shape == (5, 10, 2 * window_radius, 2)
+    assert isinstance(result, np.ndarray)
+    assert result.dtype == spike_array.dtype
 
 
 def test_get_state_firing():
+    """Test get_state_firing function with valid inputs."""
+    np.random.seed(42)  # For reproducible tests
     spike_array = np.random.rand(5, 10, 1000)
     tau_array = np.array(
         [[100, 300], [200, 400], [150, 350], [250, 450], [300, 500]])
     result = get_state_firing(spike_array, tau_array)
     assert result.shape == (5, 3, 10)
+    assert isinstance(result, np.ndarray)
+    # Check that all values are non-negative (firing rates should be >= 0)
+    assert np.all(result >= 0)
 
 
 def test_calc_significant_neurons_firing():
+    """Test calc_significant_neurons_firing function with valid inputs."""
+    np.random.seed(42)  # For reproducible tests
     state_firing = np.random.rand(5, 3, 10)
     p_val = 0.05
     anova_p_val_array, anova_sig_neurons = calc_significant_neurons_firing(
         state_firing, p_val)
     assert len(anova_p_val_array) == 10
+    assert isinstance(anova_p_val_array, np.ndarray)
+    assert isinstance(anova_sig_neurons, np.ndarray)
+    # P-values should be between 0 and 1
+    assert np.all(anova_p_val_array >= 0)
+    assert np.all(anova_p_val_array <= 1)
 
 
 def test_calc_significant_neurons_snippets():
+    """Test calc_significant_neurons_snippets function with valid inputs."""
+    np.random.seed(42)  # For reproducible tests
     transition_snips = np.random.rand(5, 10, 100, 2)
     p_val = 0.05
     pairwise_p_val_array, pairwise_sig_neurons = calc_significant_neurons_snippets(
         transition_snips, p_val)
     assert pairwise_p_val_array.shape == (10, 2)
+    assert isinstance(pairwise_p_val_array, np.ndarray)
+    assert isinstance(pairwise_sig_neurons, np.ndarray)
+    # P-values should be between 0 and 1
+    assert np.all(pairwise_p_val_array >= 0)
+    assert np.all(pairwise_p_val_array <= 1)
 
 
 def test_firing_class():
-    tau_instance = _tau(np.random.rand(5, 2), {"preprocess": {
+    """Test _firing class basic attributes without full initialization."""
+    np.random.seed(42)  # For reproducible tests
+    # Create tau values that are reasonable integers within the spike array range
+    tau_array = np.array([[100, 300], [200, 400], [150, 350], [250, 450], [300, 500]], dtype=float)
+    tau_instance = _tau(tau_array, {"preprocess": {
                         "time_lims": [0, 1000], "bin_width": 1}})
     processed_spikes = np.random.rand(5, 10, 1000)
-    metadata = {"data": {"data_dir": "/path/to/data",
+    metadata = {"data": {"data_dir": "/tmp",
                          "region_name": "region", "taste_num": "all"}}
[…]
