+++ tests/test_changepoint_analysis.py
@@ -76,19 +76,21 @@ def test_firing_class():
     """Test _firing class basic attributes without full initialization."""
     np.random.seed(42)  # For reproducible tests
     # Create tau values that are reasonable integers within the spike array range
-    tau_array = np.array([[100, 300], [200, 400], [150, 350], [250, 450], [300, 500]], dtype=float)
+    tau_array = np.array([[100, 300], [200, 400], [150, 350], [
+                         250, 450], [300, 500]], dtype=float)
     tau_instance = _tau(tau_array, {"preprocess": {
                         "time_lims": [0, 1000], "bin_width": 1}})
     processed_spikes = np.random.rand(5, 10, 1000)
     metadata = {"data": {"data_dir": "/tmp",
                          "region_name": "region", "taste_num": "all"}}
-    
+
     # Test that we can create a _firing instance and set basic attributes
-    firing_instance = _firing.__new__(_firing)  # Create without calling __init__
+    # Create without calling __init__
+    firing_instance = _firing.__new__(_firing)
     firing_instance.tau = tau_instance
     firing_instance.processed_spikes = processed_spikes
     firing_instance.metadata = metadata
-    
+
     # Test basic attribute access
     assert hasattr(firing_instance, 'tau')
     assert hasattr(firing_instance, 'processed_spikes')
@@ -100,7 +102,8 @@ def test_tau_class():
     """Test _tau class initialization and basic functionality."""
     np.random.seed(42)  # For reproducible tests
     # Create reasonable tau values
-    tau_array = np.array([[100, 300], [200, 400], [150, 350], [250, 450], [300, 500]], dtype=float)
+    tau_array = np.array([[100, 300], [200, 400], [150, 350], [
+                         250, 450], [300, 500]], dtype=float)
     metadata = {"preprocess": {"time_lims": [0, 1000], "bin_width": 1}}
     tau_instance = _tau(tau_array, metadata)
     assert tau_instance.raw_tau.shape == (5, 2)
@@ -115,20 +118,20 @@ def test_pkl_handler():
     """Test PklHandler class basic functionality without full initialization."""
     # Test that the PklHandler class has the expected structure
     assert hasattr(PklHandler, '__init__')
-    
+
     # Test that we can inspect the class without instantiating it
     import inspect
     init_signature = inspect.signature(PklHandler.__init__)
     expected_params = ['self', 'file_path']
     actual_params = list(init_signature.parameters.keys())
     assert actual_params == expected_params
-    
+
     # Test basic file path handling
     test_path = "/path/to/test_file.pkl"
     handler = PklHandler.__new__(PklHandler)  # Create without calling __init__
     handler.dir_name = os.path.dirname(test_path)
     handler.file_name_base = os.path.splitext(os.path.basename(test_path))[0]
-    
+
     assert handler.dir_name == "/path/to"
     assert handler.file_name_base == "test_file"
 
@@ -138,12 +141,12 @@ def test_get_transition_snips_edge_cases():
     np.random.seed(42)
     spike_array = np.random.rand(3, 5, 500)
     tau_array = np.array([[100, 200], [150, 250], [120, 220]])
-    
+
     # Test with small window radius
     window_radius = 10
     result = get_transition_snips(spike_array, tau_array, window_radius)
     assert result.shape == (3, 5, 2 * window_radius, 2)
-    
+
     # Test that function raises error when window extends outside bounds
     with pytest.raises(ValueError, match="Transition window extends outside data bounds"):
[…]
