### 2025-09-18 c0bbe68
remove jupyter notebooks from precommmit checks
