+++ .pre-commit-config.yaml
@@ -8,6 +8,7 @@ repos:
     -   id: check-added-large-files
     -   id: check-ast
     -   id: check-json
+        exclude_types: [jupyter]
     -   id: check-merge-conflict
     -   id: detect-private-key
     -   id: check-toml
