### 2025-09-19 82948b3
Add scikit-learn style API for direct numpy array input
- Add ChangepointDetector class with fit/predict methods
- Add fit_changepoint_model convenience function
- Update __init__.py to expose new API while maintaining backward compatibility
- Add comprehensive example notebook demonstrating new API
- Support all existing model types (SingleTastePoisson, etc.)
- Remove dependency on HDF5 files for basic usage
- Follow scikit-learn conventions (get_params, set_params, score)

Addresses issue #10

Co-authored-by: Ona <no-reply@ona.com>
