+++ pytau/changepoint_api.py
@@ -1,0 +1,431 @@
+"""
+Scikit-learn style API for PyTau changepoint detection models.
+
+This module provides a simplified interface for changepoint detection
+that accepts numpy arrays directly, similar to scikit-learn estimators.
+"""
+
+import numpy as np
+import pymc as pm
+from typing import Optional, Union, Dict, Any
+import warnings
+
+from . import changepoint_model
+from . import changepoint_preprocess
+
+
+class ChangepointDetector:
+    """
+    Scikit-learn style interface for changepoint detection in neural data.
+    
+    This class provides a simplified API that accepts spike train arrays directly
+    without requiring HDF5 files, following scikit-learn conventions.
+    
+    Parameters
+    ----------
+    model_type : str, default='single_taste_poisson'
+        Type of changepoint model to use. Options:
+        - 'single_taste_poisson': Standard Poisson changepoint model
+        - 'single_taste_poisson_varsig': Variable sigmoid slope model
+        - 'single_taste_poisson_dirichlet': Dirichlet process model
+        - 'gaussian_changepoint_mean_var_2d': Gaussian model for 2D data
+        
+    n_states : int, default=2
+        Number of states/changepoints to detect
+        
+    bin_width : int, default=1
+        Width for binning spike data (in time units)
+        
+    data_transform : str, default=None
+        Data transformation to apply. Options:
+        - None: Use actual data
+        - 'trial_shuffled': Shuffle trials
+        - 'spike_shuffled': Shuffle spikes within trials
+        - 'simulated': Generate simulated data
+        
+    inference_method : str, default='advi'
+        Inference method to use. Options:
+        - 'advi': Automatic Differentiation Variational Inference
+        - 'nuts': No-U-Turn Sampler (MCMC)
+        
+    n_iterations : int, default=10000
+        Number of iterations for inference
+        
+    random_state : int, optional
+        Random seed for reproducibility
+        
+    Attributes
+    ----------
+    model_ : pymc.Model
+        The fitted PyMC model
+        
+    trace_ : pymc.backends.base.MultiTrace
+        Posterior samples from inference
+        
+    is_fitted_ : bool
+        Whether the model has been fitted
+        
+    Examples
+    --------
+    >>> import numpy as np
+    >>> from pytau import ChangepointDetector
+    >>> 
+    >>> # Generate synthetic spike data (trials x neurons x time)
+    >>> spike_data = np.random.poisson(2, size=(20, 5, 100))
+    >>> 
+    >>> # Create and fit model
+    >>> detector = ChangepointDetector(n_states=3)
+    >>> detector.fit(spike_data)
+    >>> 
+    >>> # Get changepoint predictions
+    >>> changepoints = detector.predict(spike_data)
+    """
+    
+    def __init__(
+        self,
+        model_type: str = 'single_taste_poisson',
+        n_states: int = 2,
+        bin_width: int = 1,
+        data_transform: Optional[str] = None,
+        inference_method: str = 'advi',
+        n_iterations: int = 10000,
+        random_state: Optional[int] = None
+    ):
+        self.model_type = model_type
+        self.n_states = n_states
+        self.bin_width = bin_width
+        self.data_transform = data_transform
+        self.inference_method = inference_method
+        self.n_iterations = n_iterations
+        self.random_state = random_state
+        
+        # Initialize attributes that will be set during fitting
+        self.model_ = None
+        self.trace_ = None
+        self.is_fitted_ = False
+        self._processed_data = None
+        
+        # Validate parameters
+        self._validate_parameters()
+    
+    def _validate_parameters(self):
+        """Validate initialization parameters."""
+        valid_models = [
+            'single_taste_poisson',
[…]
