### 2025-09-19 45ac3fd
Add global verbose setting with comprehensive control
- Add config.py with global verbose configuration system
- Support environment variable PYTAU_VERBOSE for default setting
- Add verbose_context context manager for temporary settings
- Update ChangepointDetector with instance-level verbose parameter
- Add verbose parameter to fit_changepoint_model convenience function
- Update all modules to respect verbose settings with detailed progress messages
- Add comprehensive verbose functionality example notebook
- Update main API documentation and README

Features:
- Global verbose control: set_ve
[…]
