+++ pytau/changepoint_api.py
@@ -12,6 +12,7 @@ import warnings
 
 from . import changepoint_model
 from . import changepoint_preprocess
+from .config import verbose_print, verbose_context
 
 
 class ChangepointDetector:
@@ -54,6 +55,10 @@ class ChangepointDetector:
     random_state : int, optional
         Random seed for reproducibility
         
+    verbose : bool, optional
+        If True, enable verbose output for this instance only.
+        If None, uses global verbose setting.
+        
     Attributes
     ----------
     model_ : pymc.Model
@@ -89,8 +94,10 @@ class ChangepointDetector:
         data_transform: Optional[str] = None,
         inference_method: str = 'advi',
         n_iterations: int = 10000,
-        random_state: Optional[int] = None
+        random_state: Optional[int] = None,
+        verbose: Optional[bool] = None
     ):
+        # Set attributes first so _verbose_print can access them
         self.model_type = model_type
         self.n_states = n_states
         self.bin_width = bin_width
@@ -98,6 +105,18 @@ class ChangepointDetector:
         self.inference_method = inference_method
         self.n_iterations = n_iterations
         self.random_state = random_state
+        self.verbose = verbose
+        
+        # Now we can use verbose printing
+        self._verbose_print(f"🔧 Initializing ChangepointDetector with:")
+        self._verbose_print(f"   - Model type: {model_type}")
+        self._verbose_print(f"   - Number of states: {n_states}")
+        self._verbose_print(f"   - Bin width: {bin_width}")
+        self._verbose_print(f"   - Data transform: {data_transform}")
+        self._verbose_print(f"   - Inference method: {inference_method}")
+        self._verbose_print(f"   - Iterations: {n_iterations}")
+        self._verbose_print(f"   - Random state: {random_state}")
+        self._verbose_print(f"   - Local verbose: {verbose}")
         
         # Initialize attributes that will be set during fitting
         self.model_ = None
@@ -106,7 +125,9 @@ class ChangepointDetector:
         self._processed_data = None
         
         # Validate parameters
+        self._verbose_print("🔍 Validating parameters...")
         self._validate_parameters()
+        self._verbose_print("✅ ChangepointDetector initialized successfully")
     
     def _validate_parameters(self):
         """Validate initialization parameters."""
@@ -133,6 +154,16 @@ class ChangepointDetector:
         if self.inference_method not in valid_inference:
             raise ValueError(f"inference_method must be one of {valid_inference}")
     
+    def _verbose_print(self, *args, **kwargs):
+        """Print function that respects instance or global verbose setting."""
+        if self.verbose is not None:
+            # Use instance-specific verbose setting
+            if self.verbose:
+                print(*args, **kwargs)
+        else:
+            # Use global verbose setting
+            verbose_print(*args, **kwargs)
+    
     def _preprocess_data(self, X: np.ndarray, time_lims: Optional[tuple] = None) -> np.ndarray:
         """
         Preprocess spike data for model fitting.
@@ -149,22 +180,32 @@ class ChangepointDetector:
         processed_data : ndarray
             Preprocessed spike data
         """
+        self._verbose_print("📊 Preprocessing spike data...")
         X = np.asarray(X)
         
         if X.ndim != 3:
             raise ValueError("Input data must be 3D array (trials x neurons x time)")
+        
+        self._verbose_print(f"   - Input shape: {X.shape} (trials × neurons × time)")
             
         if time_lims is None:
             time_lims = (0, X.shape[-1])
+        
+        self._verbose_print(f"   - Time limits: {time_lims}")
+        self._verbose_print(f"   - Bin width: {self.bin_width}")
             
         # Apply preprocessing if specified
         if self.data_transform is not None:
+            self._verbose_print(f"   - Applying data transform: {self.data_transform}")
[…]
