### 2025-09-19 c2ec3e9
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
