+++ pytau/config.py
@@ -12,25 +12,25 @@ from typing import Union
 class PyTauConfig:
     """
     Global configuration class for PyTau settings.
-    
+
     This class manages global settings that affect the behavior of PyTau
     across all modules, including verbosity levels and logging preferences.
-    
+
     Attributes
     ----------
     verbose : bool
         Global verbosity setting. When True, enables informative print
         statements throughout PyTau operations.
     """
-    
+
     def __init__(self):
         # Initialize with default settings
         self._verbose = self._get_default_verbose()
-    
+
     def _get_default_verbose(self) -> bool:
         """
         Get default verbose setting from environment variable or default.
-        
+
         Returns
         -------
         bool
@@ -45,17 +45,17 @@ class PyTauConfig:
         else:
             # Default to False for quiet operation
             return False
-    
+
     @property
     def verbose(self) -> bool:
         """Get current verbose setting."""
         return self._verbose
-    
+
     @verbose.setter
     def verbose(self, value: Union[bool, int, str]):
         """
         Set verbose setting.
-        
+
         Parameters
         ----------
         value : bool, int, or str
@@ -77,30 +77,31 @@ class PyTauConfig:
             else:
                 raise ValueError(f"Invalid verbose string value: {value}")
         else:
-            raise TypeError(f"Verbose setting must be bool, int, or str, got {type(value)}")
-    
+            raise TypeError(
+                f"Verbose setting must be bool, int, or str, got {type(value)}")
+
     def set_verbose(self, value: Union[bool, int, str]) -> None:
         """
         Set verbose setting (alternative method).
-        
+
         Parameters
         ----------
         value : bool, int, or str
             Verbose setting
         """
         self.verbose = value
-    
+
     def get_verbose(self) -> bool:
         """
         Get current verbose setting (alternative method).
-        
+
         Returns
         -------
         bool
             Current verbose setting
         """
         return self.verbose
-    
+
     def reset_to_defaults(self) -> None:
         """Reset all settings to their default values."""
         self._verbose = self._get_default_verbose()
@@ -113,27 +114,27 @@ _config = PyTauConfig()
 def set_verbose(value: Union[bool, int, str]) -> None:
     """
     Set global verbose setting for PyTau.
-    
+
     When verbose is True, PyTau will print informative messages about:
     - Model initialization and configuration
     - Method activation and progress
     - Data preprocessing steps
     - Inference progress and completion
-    
+
     Parameters
     ----------
     value : bool, int, or str
         Verbose setting. Accepts:
         - bool: True/False
-        - int: 0 (False) or non-zero (True)  
+        - int: 0 (False) or non-zero (True)
         - str: 'true'/'false', '1'/'0', 'yes'/'no', 'on'/'off'
-    
+
     Examples
     --------
     >>> import pytau
     >>> pytau.set_verbose(True)
     >>> # Now all PyTau operations will print informative messages
-    >>> 
+    >>>
     >>> pytau.set_verbose('off')
     >>> # Disable verbose output
     """
@@ -143,12 +144,12 @@ def set_verbose(value: Union[bool, int, str]) -> None:
 def get_verbose() -> bool:
     """
     Get current global verbose setting.
-    
+
     Returns
     -------
     bool
         Current verbose setting
-    
+
     Examples
     --------
     >>> import pytau
@@ -164,16 +165,16 @@ def get_verbose() -> bool:
 def verbose_print(*args, **kwargs) -> None:
     """
     Print function that respects global verbose setting.
-    
+
     Only prints if global verbose setting is True. Otherwise, does nothing.
-    
+
     Parameters
     ----------
     *args
         Arguments passed to print()
[…]
