### 2025-09-23 c46b25a
Add comprehensive model demonstration notebooks
- Created demo notebooks for 6 core PyTau model types:
  * GaussianChangepointMeanVar2D - mean and variance changes
  * GaussianChangepointMean2D - mean-only changes
  * GaussianChangepointMeanDirichlet - automatic state selection
  * SingleTastePoisson - single condition spike data
  * AllTastePoisson - multi-condition hierarchical model
  * CategoricalChangepoint2D - discrete/categorical data

- Each notebook includes:
  * Synthetic data generation with known changepoints
  * Comprehensive data visualization
  * Model fitting with variational i
[…]
