### 2025-09-23 f4c7524
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
