### 2025-09-23 bc05c28
Add OS-specific test badges to README
- Add individual badges for Ubuntu, macOS, and Windows test status
- Badges show test results for each operating system in the CI matrix
- Resolves issue #130

Co-authored-by: Ona <no-reply@ona.com>
