+++ pytau/changepoint_api.py
@@ -5,23 +5,23 @@ This module provides a simplified interface for changepoint detection
 that accepts numpy arrays directly, similar to scikit-learn estimators.
 """
 
+import warnings
+from typing import Any, Dict, Optional, Union
+
 import numpy as np
 import pymc as pm
-from typing import Optional, Union, Dict, Any
-import warnings
 
-from . import changepoint_model
-from . import changepoint_preprocess
-from .config import verbose_print, verbose_context
+from . import changepoint_model, changepoint_preprocess
+from .config import verbose_context, verbose_print
 
 
 class ChangepointDetector:
     """
     Scikit-learn style interface for changepoint detection in neural data.
-    
+
     This class provides a simplified API that accepts spike train arrays directly
     without requiring HDF5 files, following scikit-learn conventions.
-    
+
     Parameters
     ----------
     model_type : str, default='single_taste_poisson'
@@ -30,62 +30,62 @@ class ChangepointDetector:
         - 'single_taste_poisson_varsig': Variable sigmoid slope model
         - 'single_taste_poisson_dirichlet': Dirichlet process model
         - 'gaussian_changepoint_mean_var_2d': Gaussian model for 2D data
-        
+
     n_states : int, default=2
         Number of states/changepoints to detect
-        
+
     bin_width : int, default=1
         Width for binning spike data (in time units)
-        
+
     data_transform : str, default=None
         Data transformation to apply. Options:
         - None: Use actual data
         - 'trial_shuffled': Shuffle trials
         - 'spike_shuffled': Shuffle spikes within trials
         - 'simulated': Generate simulated data
-        
+
     inference_method : str, default='advi'
         Inference method to use. Options:
         - 'advi': Automatic Differentiation Variational Inference
         - 'nuts': No-U-Turn Sampler (MCMC)
-        
+
     n_iterations : int, default=10000
         Number of iterations for inference
-        
+
     random_state : int, optional
         Random seed for reproducibility
-        
+
     verbose : bool, optional
         If True, enable verbose output for this instance only.
         If None, uses global verbose setting.
-        
+
     Attributes
     ----------
     model_ : pymc.Model
         The fitted PyMC model
-        
+
     trace_ : pymc.backends.base.MultiTrace
         Posterior samples from inference
-        
+
     is_fitted_ : bool
         Whether the model has been fitted
-        
+
     Examples
     --------
     >>> import numpy as np
     >>> from pytau import ChangepointDetector
-    >>> 
+    >>>
     >>> # Generate synthetic spike data (trials x neurons x time)
     >>> spike_data = np.random.poisson(2, size=(20, 5, 100))
-    >>> 
+    >>>
     >>> # Create and fit model
     >>> detector = ChangepointDetector(n_states=3)
     >>> detector.fit(spike_data)
-    >>> 
+    >>>
     >>> # Get changepoint predictions
     >>> changepoints = detector.predict(spike_data)
     """
-    
+
     def __init__(
         self,
         model_type: str = 'single_taste_poisson',
@@ -106,7 +106,7 @@ class ChangepointDetector:
         self.n_iterations = n_iterations
         self.random_state = random_state
         self.verbose = verbose
-        
+
         # Now we can use verbose printing
         self._verbose_print(f"🔧 Initializing ChangepointDetector with:")
         self._verbose_print(f"   - Model type: {model_type}")
@@ -117,43 +117,46 @@ class ChangepointDetector:
         self._verbose_print(f"   - Iterations: {n_iterations}")
         self._verbose_print(f"   - Random state: {random_state}")
         self._verbose_print(f"   - Local verbose: {verbose}")
-        
+
         # Initialize attributes that will be set during fitting
         self.model_ = None
[…]
