### 2025-09-23 dc17fbc
Add comprehensive notebook testing with nbmake
- Add nbmake dependency for automated Jupyter notebook testing
- Create dedicated test suite for notebooks in tests/test_notebooks.py
- Configure pytest with custom markers and notebook testing support
- Add separate tox environment for notebook testing (tox -e notebooks)
- Update CI/CD workflow to include notebook testing for key notebooks
- Document 7 working notebooks and 4 with known API issues
- Create comprehensive documentation in docs/notebook_testing.md
- Update README with testing section and usage instructions

Addresses issue #128 by e
[…]
