+++ pytest.ini
@@ -1,0 +1,10 @@
+[tool:pytest]
+testpaths = tests
+python_files = test_*.py
+python_functions = test_*
+markers =
+    slow: marks tests as slow (deselect with '-m "not slow"')
+    notebook: marks tests as notebook tests
+addopts = 
+    -v
+    --tb=short
