+++ tests/test_notebooks.py
@@ -6,9 +6,10 @@ without errors. It uses nbmake to run notebooks and verify they complete
 successfully.
 """
 
-import pytest
 import os
 from pathlib import Path
+
+import pytest
 
 # Get the project root directory
 PROJECT_ROOT = Path(__file__).parent.parent
@@ -16,14 +17,15 @@ PROJECT_ROOT = Path(__file__).parent.parent
 # Define notebook directories
 NOTEBOOK_DIRS = [
     PROJECT_ROOT / "pytau" / "how_to" / "notebooks",
-    PROJECT_ROOT / "pytau" / "how_to" / "examples", 
+    PROJECT_ROOT / "pytau" / "how_to" / "examples",
     PROJECT_ROOT / "pytau" / "how_to" / "model_demos",
 ]
+
 
 def collect_notebooks():
     """Collect all notebook files from the specified directories."""
     notebooks = []
-    
+
     # Known problematic notebooks to skip (due to API issues)
     skip_notebooks = {
         "PyTau_New_API_Example.ipynb",  # Has AttributeError with trace_.varnames
@@ -31,7 +33,7 @@ def collect_notebooks():
         "PyTau_Walkthrough_no_handlers.ipynb",  # May have similar issues
         "PyTau_Walkthrough_with_handlers.ipynb",  # May have similar issues
     }
-    
+
     for notebook_dir in NOTEBOOK_DIRS:
         if notebook_dir.exists():
             for notebook in notebook_dir.glob("*.ipynb"):
@@ -39,18 +41,20 @@ def collect_notebooks():
                     notebooks.append(notebook)
     return notebooks
 
+
 # Collect all notebooks
 NOTEBOOKS = collect_notebooks()
+
 
 @pytest.mark.notebook
 @pytest.mark.parametrize("notebook_path", NOTEBOOKS, ids=lambda x: x.name)
 def test_notebook_execution(notebook_path):
     """
     Test that a notebook can be executed without errors.
-    
+
     This test uses nbmake's pytest integration to execute notebooks
     and verify they complete successfully.
-    
+
     Args:
         notebook_path: Path to the notebook file to test
     """
@@ -59,42 +63,46 @@ def test_notebook_execution(notebook_path):
     assert notebook_path.exists(), f"Notebook {notebook_path} does not exist"
     assert notebook_path.suffix == ".ipynb", f"File {notebook_path} is not a notebook"
 
+
 @pytest.mark.notebook
 def test_notebooks_exist():
     """Test that we have notebooks to test."""
     assert len(NOTEBOOKS) > 0, "No notebooks found to test"
-    
+
     # Print found notebooks for debugging
     print(f"\nFound {len(NOTEBOOKS)} notebooks:")
     for notebook in NOTEBOOKS:
         print(f"  - {notebook.relative_to(PROJECT_ROOT)}")
 
+
 @pytest.mark.notebook
 def test_known_problematic_notebooks():
     """
     Document notebooks that are currently skipped due to known issues.
-    
+
     This test serves as documentation and will fail if the problematic
     notebooks are fixed and should be re-enabled.
     """
     skip_notebooks = {
         "PyTau_New_API_Example.ipynb",
-        "PyTau_Verbose_Example.ipynb", 
+        "PyTau_Verbose_Example.ipynb",
         "PyTau_Walkthrough_no_handlers.ipynb",
         "PyTau_Walkthrough_with_handlers.ipynb",
     }
-    
+
     # Find all notebooks that exist but are being skipped
     all_notebooks = []
     for notebook_dir in NOTEBOOK_DIRS:
         if notebook_dir.exists():
             all_notebooks.extend(notebook_dir.glob("*.ipynb"))
-    
[…]
