+++ pytest.ini
@@ -5,6 +5,6 @@ python_functions = test_*
 markers =
     slow: marks tests as slow (deselect with '-m "not slow"')
     notebook: marks tests as notebook tests
-addopts = 
+addopts =
     -v
     --tb=short
