### 2025-09-23 54e9dad
Add model notebooks for each model type with dummy data and nbmake testing
- Created notebooks under pytau/how_to/model_notebooks/ for 7 key model types:
  - GaussianChangepointMeanVar2D
  - GaussianChangepointMeanDirichlet
  - GaussianChangepointMean2D
  - SingleTastePoissonDirichlet
  - SingleTastePoisson
  - AllTastePoisson
  - CategoricalChangepoint2D

- Each notebook demonstrates:
  - Generating appropriate dummy data for the model type
  - Initializing the model with correct parameters
  - Running model tests
  - Visualizing input data with plots

- Added nbmake dependency for notebook t
[…]
