+++ tox.ini
@@ -10,8 +10,12 @@ deps =
     arviz==0.18.0
     graphviz==0.20.3
     pytest-cov
+    nbmake
+    matplotlib
+    jupyter
 commands =
     pytest {posargs:tests/} --cov=pytau --cov-report=term-missing
+    pytest --nbmake pytau/how_to/model_notebooks/
 
 [pytest]
 testpaths = tests
