+++ pyproject.toml
@@ -38,6 +38,9 @@ dev = [
     "pre-commit",
     "pytest",
     "pytest-cov",
+    "nbmake",
+    "matplotlib",
+    "jupyter",
 ]
 
 [tool.setuptools]
