### 2025-09-23 980ca13
Add notebook testing to GitHub Actions workflow
- Added separate step to run nbmake tests on model notebooks
- Ensures notebooks are tested across all OS and Python versions in CI
- Notebooks will be validated on every PR and push to main

Co-authored-by: Ona <no-reply@ona.com>
