### 2025-09-23 e80c9b8
Add notebook execution with outputs to GitHub Actions workflow
WORKFLOW ENHANCEMENTS:
✅ Added 'execute-notebooks' job that runs after all tests pass
✅ Executes all model notebooks and saves outputs using jupyter nbconvert
✅ Uploads executed notebooks as artifacts for every run
✅ Auto-commits executed notebooks to main branch (with [skip ci])

DEPENDENCY UPDATES:
✅ Added nbconvert to pyproject.toml dev dependencies
✅ Added nbconvert to requirements-dev.txt
✅ Added nbconvert to tox.ini dependencies

FEATURES:
🔹 Notebooks executed with 300-second timeout for complex models
🔹 Artifacts retained fo
[…]
