+++ tox.ini
@@ -13,6 +13,7 @@ deps =
     nbmake
     matplotlib
     jupyter
+    nbconvert
 commands =
     pytest {posargs:tests/} --cov=pytau --cov-report=term-missing
     pytest --nbmake pytau/how_to/model_notebooks/
