### 2025-09-23 05832eb
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
