+++ .github/workflows/pytest_workflow.yml
@@ -72,11 +72,11 @@ jobs:
         else
           echo "Committing updated notebooks with execution outputs"
           git commit -m "Auto-update model notebooks with execution outputs
-          
+
           - Executed all model notebooks with latest code
           - Preserved outputs for documentation and review
           - Generated by GitHub Actions workflow
-          
+
           [skip ci]"
           git push
         fi
