### 2025-09-23 00372e9
Fix GitHub Actions workflow: Update deprecated action versions
FIXES:
❌ Fixed: actions/upload-artifact@v3 → v4 (was causing workflow failure)
🔄 Updated: actions/checkout@v2 → v4 (for better compatibility)
🔄 Updated: actions/setup-python@v2 → v4 (for better compatibility)

ERROR RESOLVED:
The execute-notebooks job was failing due to deprecated v3 artifact action.
GitHub automatically fails workflows using deprecated actions as of April 2024.

This should resolve the workflow failure and allow notebook execution to proceed.

Co-authored-by: Ona <no-reply@ona.com>
