### 2025-09-24 a90ff4b
Add 1D Poisson changepoint model
- Implement PoissonChangepoint1D class for 1D time series changepoint detection
- Extend gen_test_array function to handle 1D arrays (integer input)
- Add comprehensive tests for the new model
- Update test imports to include the new model class

Fixes #137

Co-authored-by: Ona <no-reply@ona.com>
