+++ tests/test_changepoint_model.py
@@ -14,6 +14,7 @@ from pytau.changepoint_model import (
     GaussianChangepointMean2D,
     GaussianChangepointMeanDirichlet,
     GaussianChangepointMeanVar2D,
+    PoissonChangepoint1D,
     SingleTastePoisson,
     SingleTastePoissonDirichlet,
     SingleTastePoissonTrialSwitch,
@@ -168,6 +169,26 @@ def test_advi_fit():
 
     # Skip actual testing as it would require a full PyMC3 model
     pytest.skip("Full testing of advi_fit requires a PyMC3 model")
+
+
+@pytest.mark.slow
+def test_poisson_changepoint_1d():
+    """Test the PoissonChangepoint1D model."""
+    # Generate 1D test data
+    test_data = gen_test_array(100, n_states=3, type="poisson")
+    
+    # Test model creation
+    model_class = PoissonChangepoint1D(test_data, 3)
+    assert model_class.data_array.ndim == 1
+    assert model_class.n_states == 3
+    
+    # Test model generation
+    model = model_class.generate_model()
+    assert model is not None
+    
+    # Test that it raises error for non-1D data
+    with pytest.raises(ValueError):
+        PoissonChangepoint1D(np.random.poisson(2, (10, 100)), 3)
 
 
 def test_module_import():
