+++ pytau/changepoint_model.py
@@ -49,12 +49,46 @@ def gen_test_array(array_size, n_states, type="poisson"):
     Time will always be last dimension
 
     Args:
-        array_size (tuple): Size of array to generate
+        array_size (tuple or int): Size of array to generate. If int, generates 1D array.
         n_states (int): Number of states to generate
         type (str): Type of data to generate
             - normal
             - poisson
     """
+    # Handle 1D case
+    if isinstance(array_size, int):
+        assert array_size > n_states, "Array too small for states"
+        assert type in [
+            "normal", "poisson"], "Invalid type, please use normal or poisson"
+        
+        # Generate transition times for 1D case
+        transition_times = np.random.random(n_states)
+        transition_times = np.cumsum(transition_times)
+        transition_times = transition_times / transition_times.max()
+        transition_times *= array_size
+        transition_times = transition_times.astype(int)
+        
+        # Generate state bounds
+        state_bounds = np.zeros(n_states + 1, dtype=int)
+        state_bounds[1:] = transition_times
+        state_bounds[-1] = array_size
+        
+        # Generate state rates
+        lambda_vals = np.random.exponential(2.0, n_states) + 0.5
+        
+        # Generate 1D array
+        rate_array = np.zeros(array_size)
+        for i in range(n_states):
+            start_idx = state_bounds[i]
+            end_idx = state_bounds[i + 1]
+            rate_array[start_idx:end_idx] = lambda_vals[i]
+        
+        if type == "poisson":
+            return np.random.poisson(rate_array)
+        else:
+            return np.random.normal(loc=rate_array, scale=0.1)
+    
+    # Handle multi-dimensional case (existing code)
     assert array_size[-1] > n_states, "Array too small for states"
     assert type in [
         "normal", "poisson"], "Invalid type, please use normal or poisson"
@@ -1826,12 +1860,14 @@ def all_taste_poisson_trial_switch(data_array, switch_components, n_states, **kw
 def run_all_tests():
     """Run tests for all model classes"""
     # Create test data
+    test_data_1d = gen_test_array(100, n_states=3, type="poisson")
     test_data_2d = gen_test_array((10, 100), n_states=3, type="normal")
     test_data_3d = gen_test_array((5, 10, 100), n_states=3, type="poisson")
     test_data_4d = gen_test_array((2, 5, 10, 100), n_states=3, type="poisson")
 
     # Test each model class
     models_to_test = [
+        PoissonChangepoint1D(test_data_1d, 3),
         GaussianChangepointMeanVar2D(test_data_2d, 3),
         GaussianChangepointMeanDirichlet(test_data_2d, 5),
         GaussianChangepointMean2D(test_data_2d, 3),
@@ -1858,6 +1894,126 @@ def run_all_tests():
     print("All tests completed")
     if failed_tests:
         print("Failed tests:", failed_tests)
+
+
+############################################################
+# 1D Poisson Changepoint Model
+############################################################
+
+
+class PoissonChangepoint1D(ChangepointModel):
+    """Model for changepoint detection in 1D Poisson time series
+    
+    This model detects changepoints in 1D time series data using a Poisson likelihood.
+    It assumes the data follows a Poisson distribution with different rates in different
+    segments separated by changepoints.
+    """
+
+    def __init__(self, data_array, n_states, **kwargs):
+        """
+        Args:
+            data_array (1D Numpy array): Time series data
+            n_states (int): Number of states to model
+            **kwargs: Additional arguments
+        """
+        super().__init__(**kwargs)
+        self.data_array = np.asarray(data_array)
+        if self.data_array.ndim != 1:
+            raise ValueError("data_array must be 1-dimensional")
[…]
