### 2025-09-24 349a637
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
