### 2025-09-24 d77b9e1
refactor(changepoint_model): eliminate duplicate code to remove redundancy
- Removed duplicated test and model code to simplify module structure.
- Consolidated similar functionalities under existing model implementations to enhance maintainability.
