+++ pytau/changepoint_model.py
@@ -2174,185 +2174,3 @@ def mcmc_fit(model, samples):
         sigma_stack = trace["sigma"].swapaxes(0, 1)
         return model, trace, mu_stack, sigma_stack, tau_samples, model.obs.observations
 
-
-######################################################################
-# Run Inference
-######################################################################
-
-
-def run_all_tests():
-    """Run tests for all model classes"""
-    # Create test data
-    test_data_1d = gen_test_array(100, n_states=3, type="poisson")
-    test_data_2d = gen_test_array((10, 100), n_states=3, type="normal")
-    test_data_3d = gen_test_array((5, 10, 100), n_states=3, type="poisson")
-    test_data_4d = gen_test_array((2, 5, 10, 100), n_states=3, type="poisson")
-
-    # Test each model class
-    models_to_test = [
-        PoissonChangepoint1D(test_data_1d, 3),
-        GaussianChangepointMeanVar2D(test_data_2d, 3),
-        GaussianChangepointMeanDirichlet(test_data_2d, 5),
-        GaussianChangepointMean2D(test_data_2d, 3),
-        SingleTastePoissonDirichlet(test_data_3d, 5),
-        SingleTastePoisson(test_data_3d, 3),
-        SingleTastePoissonVarsig(test_data_3d, 3),
-        SingleTastePoissonVarsigFixed(test_data_3d, 3, 1),
-        SingleTastePoissonTrialSwitch(test_data_3d, 2, 3),
-        AllTastePoisson(test_data_4d, 3),
-        AllTastePoissonVarsigFixed(test_data_4d, 3, 1),
-        AllTastePoissonTrialSwitch(test_data_4d, 2, 3),
-    ]
-
-    failed_tests = []
-    pbar = tqdm(models_to_test, total=len(models_to_test))
-    for model in pbar:
-        try:
-            model.test()
-            pbar.set_description(f"Test passed for {model.__class__.__name__}")
-        except Exception as e:
-            failed_tests.append(model.__class__.__name__)
-            print(f"Test failed for {model.__class__.__name__}: {str(e)}")
-
-    print("All tests completed")
-    if failed_tests:
-        print("Failed tests:", failed_tests)
-
-
-############################################################
-# 1D Poisson Changepoint Model
-############################################################
-
-
-class PoissonChangepoint1D(ChangepointModel):
-    """Model for changepoint detection in 1D Poisson time series
-
-    This model detects changepoints in 1D time series data using a Poisson likelihood.
-    It assumes the data follows a Poisson distribution with different rates in different
-    segments separated by changepoints.
-    """
-
-    def __init__(self, data_array, n_states, **kwargs):
-        """
-        Args:
-            data_array (1D Numpy array): Time series data
-            n_states (int): Number of states to model
-            **kwargs: Additional arguments
-        """
-        super().__init__(**kwargs)
-        self.data_array = np.asarray(data_array)
-        if self.data_array.ndim != 1:
-            raise ValueError("data_array must be 1-dimensional")
-        self.n_states = n_states
-
-    def generate_model(self):
-        """
-        Returns:
-            pymc model: Model class containing graph to run inference on
-        """
-        data_array = self.data_array
-        n_states = self.n_states
-
-        # Calculate initial lambda values by splitting data into segments
-        mean_vals = np.array([
-            np.mean(x) for x in np.array_split(data_array, n_states)
-        ])
-        mean_vals += 0.01  # To avoid zero starting prob
-
-        idx = np.arange(len(data_array))
-        length = len(data_array)
-
-        with pm.Model() as model:
-            # Lambda parameters for each state (Poisson rates)
-            lambda_latent = pm.Exponential(
[…]
