### 2025-09-24 453edb6
docs: Update README with comprehensive model types and inference methods
Co-authored-by: aider (anthropic/claude-sonnet-4-20250514) <aider@aider.chat>
