+++ README.md
@@ -51,6 +51,8 @@ jupyter notebook
   - All-taste hierarchical models
   - Dirichlet process models for automatic state detection
   - Variable sigmoid models for flexible transition shapes
+  - Gaussian changepoint models for continuous data
+  - Categorical changepoint models for discrete data
 
 - **Flexible Data Handling**:
   - Support for shuffled, simulated, and actual neural data
@@ -61,6 +63,89 @@ jupyter notebook
   - State transition detection
   - Cross-region correlation analysis
   - Statistical significance testing
+
+## 📊 Available Models
+
+### Poisson Models (for spike count data)
+- **`PoissonChangepoint1D`**: 1D time series changepoint detection
+- **`SingleTastePoisson`**: Basic single-taste model with fixed sigmoid transitions
+- **`SingleTastePoissonVarsig`**: Variable sigmoid slope inferred from data
+- **`SingleTastePoissonVarsigFixed`**: Fixed sigmoid slope with configurable sharpness
+- **`SingleTastePoissonDirichlet`**: Automatic state number detection using Dirichlet process
+- **`SingleTastePoissonTrialSwitch`**: Models trial-to-trial emission changes
+- **`AllTastePoisson`**: Hierarchical model across multiple tastes
+- **`AllTastePoissonVarsigFixed`**: All-taste model with fixed sigmoid transitions
+- **`AllTastePoissonTrialSwitch`**: All-taste model with trial switching
+
+### Gaussian Models (for continuous data)
+- **`GaussianChangepointMean2D`**: Detects changes in mean only
+- **`GaussianChangepointMeanVar2D`**: Detects changes in both mean and variance
+- **`GaussianChangepointMeanDirichlet`**: Automatic state detection for Gaussian data
+
+### Categorical Models (for discrete categorical data)
+- **`CategoricalChangepoint2D`**: Changepoint detection for categorical time series
+
+## 🔧 Inference Methods
+
+### Variational Inference (ADVI)
+- **Fast approximate inference** using Automatic Differentiation Variational Inference
+- **Convergence monitoring** with customizable tolerance
+- **Full-rank ADVI** for better approximation quality
+
+```python
+from pytau.changepoint_model import advi_fit
+
+# Fit model using ADVI
+model, approx = advi_fit(model, fit=10000, samples=5000, convergence_tol=0.01)
+trace = approx.sample(draws=5000)
+```
+
+### MCMC Sampling
+- **High-quality posterior samples** using NUTS sampler
+- **Multi-chain sampling** for convergence diagnostics
+- **NumPyro backend support** for improved performance
+
+```python
+from pytau.changepoint_model import mcmc_fit, dpp_fit
+
+# Standard MCMC
+model, trace, lambda_stack, tau_samples, observations = mcmc_fit(model, samples=1000)
+
+# Dirichlet Process Prior models
+dpp_trace = dpp_fit(model, n_chains=4, tune=500, draws=500, use_numpyro=True)
+```
+
+### Model Selection
+- **Automatic state number detection** using ELBO comparison
+- **Cross-validation support** for model comparison
+
+```python
+from pytau.changepoint_model import find_best_states
+
+# Find optimal number of states
+best_model, model_list, elbo_values = find_best_states(
+    data, model_generator, n_fit=5000, n_samples=1000, 
+    min_states=2, max_states=8
+)
+```
+
+## 📈 Data Formats
+
+### Input Data Shapes
+- **1D**: `(time,)` - Single time series
+- **2D**: `(trials, time)` or `(neurons, time)` - Multiple trials or neurons
+- **3D**: `(trials, neurons, time)` - Single taste experiments
[…]
