### 2025-09-24 e367a9a
Add comprehensive testing for how_to notebooks and scripts
- Add test_how_to_notebooks.py for notebook structure validation and execution testing
- Add test_how_to_scripts.py for Python and shell script syntax validation
- Update tox.ini to include nbmake testing for all how_to directories
- Update GitHub Actions workflow to test notebooks and scripts
- Add comprehensive documentation in README_how_to_testing.md
- Tests validate file existence, syntax, and structure without requiring full dependencies
- Separate nbmake execution tests available for environments with full dependencies

Addresse
[…]
