### 2025-09-24 06b8285
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
