+++ tests/test_how_to_scripts.py
@@ -2,41 +2,42 @@
 Tests for how_to scripts to ensure they can be imported and have basic functionality.
 """
 
-import pytest
-import sys
-import subprocess
-from pathlib import Path
 import importlib.util
+import subprocess
+import sys
+from pathlib import Path
 
+import pytest
 
 # Get the path to the how_to scripts directory
-HOW_TO_SCRIPTS_DIR = Path(__file__).parent.parent / "pytau" / "how_to" / "scripts"
+HOW_TO_SCRIPTS_DIR = Path(__file__).parent.parent / \
+    "pytau" / "how_to" / "scripts"
 EXAMPLES_DIR = Path(__file__).parent.parent / "pytau" / "how_to" / "examples"
 
 
 def get_python_script_files():
     """Get all Python script files from how_to directories."""
     script_files = []
-    
+
     # Get scripts from scripts directory
     if HOW_TO_SCRIPTS_DIR.exists():
         script_files.extend(list(HOW_TO_SCRIPTS_DIR.glob("*.py")))
-    
+
     # Get scripts from examples directory
     if EXAMPLES_DIR.exists():
         script_files.extend(list(EXAMPLES_DIR.glob("*.py")))
-    
+
     return script_files
 
 
 def get_shell_script_files():
     """Get all shell script files from how_to directories."""
     script_files = []
-    
+
     # Get shell scripts from scripts directory
     if HOW_TO_SCRIPTS_DIR.exists():
         script_files.extend(list(HOW_TO_SCRIPTS_DIR.glob("*.sh")))
-    
+
     return script_files
 
 
@@ -45,11 +46,11 @@ def test_python_script_syntax(script_path):
     """Test that Python scripts have valid syntax and can be imported."""
     assert script_path.exists(), f"Script {script_path} does not exist"
     assert script_path.suffix == ".py", f"File {script_path} is not a Python script"
-    
+
     # Test syntax by attempting to compile
     with open(script_path, 'r') as f:
         source_code = f.read()
-    
+
     try:
         compile(source_code, str(script_path), 'exec')
     except SyntaxError as e:
@@ -61,11 +62,11 @@ def test_shell_script_syntax(script_path):
     """Test that shell scripts have basic syntax validity."""
     assert script_path.exists(), f"Script {script_path} does not exist"
     assert script_path.suffix == ".sh", f"File {script_path} is not a shell script"
-    
+
     # Basic check - ensure file is not empty and has shebang or content
     with open(script_path, 'r') as f:
         content = f.read().strip()
-    
+
     assert len(content) > 0, f"Shell script {script_path} is empty"
 
 
@@ -77,20 +78,20 @@ def test_expected_scripts_exist():
         "changepoint_fit_test.py",
         "database_handler_test.py"
     ]
-    
+
     expected_shell_scripts = [
         "download_test_data.sh"
     ]
-    
+
     all_python_scripts = get_python_script_files()
     python_script_names = [script.name for script in all_python_scripts]
-    
+
     all_shell_scripts = get_shell_script_files()
     shell_script_names = [script.name for script in all_shell_scripts]
-    
+
     for expected in expected_python_scripts:
         assert expected in python_script_names, f"Expected Python script {expected} not found"
-    
+
     for expected in expected_shell_scripts:
         assert expected in shell_script_names, f"Expected shell script {expected} not found"
 
@@ -108,21 +109,23 @@ def test_python_script_imports(script_path):
     # and might require specific data or setup
     skip_import_test = [
         "fit_manually.py",
-        "fit_with_fit_handler.py", 
+        "fit_with_fit_handler.py",
         "changepoint_fit_test.py",
         "database_handler_test.py"
     ]
-    
+
     if script_path.name in skip_import_test:
[…]
