+++ .github/workflows/pytest_workflow.yml
@@ -33,7 +33,7 @@ jobs:
         pytest tests/test_how_to_notebooks.py -v
         # Try to run notebooks with nbmake (may fail if dependencies missing)
         pytest --nbmake pytau/how_to/model_notebooks/ -v --tb=short || echo "Model notebooks failed - likely missing dependencies"
-        pytest --nbmake pytau/how_to/notebooks/ -v --tb=short || echo "How-to notebooks failed - likely missing dependencies" 
+        pytest --nbmake pytau/how_to/notebooks/ -v --tb=short || echo "How-to notebooks failed - likely missing dependencies"
         pytest --nbmake pytau/how_to/examples/ -v --tb=short || echo "Example notebooks failed - likely missing dependencies"
 
   execute-notebooks:
