+++ tox.ini
@@ -19,7 +19,7 @@ commands =
     # Note: nbmake tests require full dependencies and may fail in minimal environments
     # Run with: tox -- --nbmake-timeout=300 to increase timeout for slow notebooks
     - pytest --nbmake pytau/how_to/model_notebooks/ --nbmake-timeout=300
-    - pytest --nbmake pytau/how_to/notebooks/ --nbmake-timeout=300  
+    - pytest --nbmake pytau/how_to/notebooks/ --nbmake-timeout=300
     - pytest --nbmake pytau/how_to/examples/ --nbmake-timeout=300
 
 [pytest]
