### 2025-09-24 181cfc5
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
