### 2025-09-25 3fbab8a
Simplify notebook testing in pytest workflow
Removed conditional checks for notebook tests and streamlined the execution of pytest for all notebooks.
