+++ .github/workflows/pytest_workflow.yml
@@ -34,19 +34,13 @@ jobs:
 
         # Try to run notebooks with nbmake (may fail if dependencies missing or path issues)
         echo "Testing model notebooks..."
-        if ! pytest --nbmake pytau/how_to/model_notebooks/ -v --tb=short; then
-          echo "Model notebooks failed - likely missing dependencies or path issues"
-        fi
+        pytest --nbmake pytau/how_to/model_notebooks/ -v
 
         echo "Testing how-to notebooks..."
-        if ! pytest --nbmake pytau/how_to/notebooks/ -v --tb=short; then
-          echo "How-to notebooks failed - likely missing dependencies or path issues"
-        fi
+        pytest --nbmake pytau/how_to/notebooks/ -v
 
         echo "Testing example notebooks..."
-        if ! pytest --nbmake pytau/how_to/examples/ -v --tb=short; then
-          echo "Example notebooks failed - likely missing dependencies or path issues"
-        fi
+        pytest --nbmake pytau/how_to/examples/ -v
 
         echo "Notebook testing completed"
       shell: bash
