### 2025-09-25 7f6cf2d
Fix matplotlib dependency issues in model notebooks
- Replace hard matplotlib imports with try-except blocks
- Create dummy plt object when matplotlib is not available
- All plotting calls now work gracefully without matplotlib
- Notebooks will only fail on core dependencies (pymc, etc.)
- Simplify GitHub Actions workflow since notebooks handle dependencies

This fixes the CI failures where notebooks were failing on matplotlib
imports in environments without full plotting dependencies.

Co-authored-by: Ona <no-reply@ona.com>
