### 2025-09-25 #8509 @abuzarmahmood
@fboehm thanks for checking in! I'm almost done. I've added testing across a couple versions of Python for Windows, mac-os and Linux. I just need till the end of the weekend to confirm everything is working well. Thanks again for your patience
