### 2025-09-26 3fa561c
Fix advi_fit function to return expected 5 values
- Restore trace sampling and variable extraction
- Return model, approx, lambda_stack, tau_samples, fit_data as expected by notebooks
- Add fallback return for edge cases
- Fixes ValueError: not enough values to unpack (expected 5, got 2)

Co-authored-by: Ona <no-reply@ona.com>
