+++ pytau/changepoint_model.py
@@ -2124,24 +2124,24 @@ def advi_fit(model, fit, samples, convergence_tol=None):
     with model:
         inference = pm.ADVI("full-rank")
         approx = pm.fit(n=fit, method=inference, callbacks=callbacks)
-        # trace = approx.sample(draws=samples)
+        trace = approx.sample(draws=samples)
 
-    return model, approx  # , trace
+    # Check if tau exists in trace
+    if "tau" not in trace.varnames:
+        raise KeyError(f"'tau' not found in trace. Available variables: {list(trace.varnames)}")
 
-    # # Check if tau exists in trace
-    # if "tau" not in trace.varnames:
-    #     raise KeyError(f"'tau' not found in trace. Available variables: {list(trace.varnames)}")
-    #
-    # # Extract relevant variables from trace
-    # tau_samples = trace["tau"]
-    # if "lambda" in trace.varnames:
-    #     lambda_stack = trace["lambda"].swapaxes(0, 1)
-    #     return model, trace, lambda_stack, tau_samples, model.obs.observations
-    # if "mu" in trace.varnames:
-    #     mu_stack = trace["mu"].swapaxes(0, 1)
-    #     sigma_stack = trace["sigma"].swapaxes(0, 1)
-    #     return model, trace, mu_stack, sigma_stack, tau_samples, model.obs.observations
-    #
+    # Extract relevant variables from trace
+    tau_samples = trace["tau"]
+    if "lambda" in trace.varnames:
+        lambda_stack = trace["lambda"].swapaxes(0, 1)
+        return model, approx, lambda_stack, tau_samples, model.obs.observations
+    if "mu" in trace.varnames:
+        mu_stack = trace["mu"].swapaxes(0, 1)
+        sigma_stack = trace["sigma"].swapaxes(0, 1)
+        return model, approx, mu_stack, sigma_stack, tau_samples, model.obs.observations
+    
+    # Fallback - return what we can
+    return model, approx, None, tau_samples, model.obs.observations
 
 
 def mcmc_fit(model, samples):
