### 2025-09-26 8bccdb4
Fix PyMC5 compatibility issues in advi_fit and mcmc_fit
- Replace trace.varnames with idata.posterior.data_vars for PyMC5
- Replace trace['var'] with idata.posterior['var'].values
- Update both advi_fit and mcmc_fit functions
- Fixes AttributeError: 'InferenceData' object has no attribute 'varnames'

Co-authored-by: Ona <no-reply@ona.com>
