### 2025-09-26 eb9c3a2
Fix observed data access in PyMC5 for advi_fit and mcmc_fit
- Replace model.obs.observations with PyMC5-compatible access
- Use model.observed_RVs[0].data.eval() to get observed data
- Add error handling for cases where observed data can't be accessed
- Fixes AttributeError: 'TensorVariable' object has no attribute 'observations'

Co-authored-by: Ona <no-reply@ona.com>
