### 2025-09-26 fbee058
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
