### 2025-09-26 bde39b1
Add defensive error handling for tensor dimension issues
- Add try/catch blocks around tensor extraction
- Handle potential dimension mismatches in PyMC5
- Add debug print statements to identify issues
- Reshape tensors if they have unexpected dimensions

Co-authored-by: Ona <no-reply@ona.com>
