### 2025-09-26 39d95cf
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
