### 2025-09-26 be574e6
Fix NumPy deprecation and PyMC5 pickling issues
- Replace deprecated np.int with int in notebook
- Fix pickling error by excluding model object from pickle dump
- Model object contains unpicklable local functions in PyMC5
- Fixes AttributeError: module 'numpy' has no attribute 'int'
- Fixes AttributeError: Can't pickle local object '_make_nice_attr_error.<locals>.fn'

Co-authored-by: Ona <no-reply@ona.com>
