### 2025-09-26 4f17a27
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
