### 2025-09-26 ca4e03b
Fix tensor dimension issues in plotting functions
- Fix notebook to create proper 2D tau array for plotting functions
- Modify _tau class to handle n_trials parameter and create 2D scaled_mode_tau
- Update PklHandler to pass number of trials from processed_spikes
- Ensure mode_result is always at least 1D before tiling

This fixes IndexError: tuple index out of range in plotting functions
that expect tau with shape (n_trials, n_changepoints).

Co-authored-by: Ona <no-reply@ona.com>
