### 2025-09-26 8825271
Fix remaining pickling error in FitHandler
Add comprehensive error handling for PyMC5 objects that contain
unpicklable local functions like '_make_nice_attr_error.<locals>.fn'.

When full pickling fails, save metadata-only version with essential
data (tau_array, processed_spikes) while skipping problematic objects.

This fixes the AttributeError in PyTau_Walkthrough_with_handlers.ipynb
while preserving the core functionality.

Co-authored-by: Ona <no-reply@ona.com>
