### 2025-09-26 d5e66fe
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
