### 2025-09-26 cc33fd5
Fix PklHandler to handle missing keys from fallback pickling
Make PklHandler robust to missing model_data keys that may be
removed during fallback pickling when full pickling fails.

Set missing attributes to None instead of raising KeyError.
This allows the handler to work with metadata-only pickle files
while maintaining backward compatibility.

Fixes KeyError: 'model' in PyTau_Walkthrough_with_handlers.ipynb.

Co-authored-by: Ona <no-reply@ona.com>
