### 2025-09-26 9e1e676
Handle None tau_array in _tau class initialization
Add check for None tau_array in _tau.__init__ to handle cases where
tau_array is missing from fallback pickling. Set all tau-related
attributes to None and return early to avoid TypeError when trying
to vectorize None values.

This fixes TypeError: int() argument must be a string, a bytes-like
object or a real number, not 'NoneType' in PyTau_Walkthrough_with_handlers.ipynb.

Co-authored-by: Ona <no-reply@ona.com>
