### 2025-09-26 6f3295c
Handle None tau attributes in _firing class initialization
Add None checks in _firing class to handle cases where tau attributes
(raw_mode_tau, scaled_mode_tau) are None due to fallback pickling.

Set all firing-related attributes to None when tau data is unavailable
instead of trying to call analysis functions with None values.

This fixes AttributeError when accessing tau attributes that are None
in PyTau_Walkthrough_with_handlers.ipynb.

Co-authored-by: Ona <no-reply@ona.com>
