### 2025-09-26 49a0db0
Enhance fallback pickling to preserve approx.hist for ELBO plotting
When full pickling fails, attempt to save a simplified approx object
containing only the hist attribute needed for plot_elbo_history().

Create SimpleApprox class with just the hist attribute to avoid
pickling the full approx object which contains unpicklable functions.

This fixes KeyError: 'approx' in PyTau_Walkthrough_with_handlers.ipynb
while preserving ELBO plotting functionality.

Co-authored-by: Ona <no-reply@ona.com>
