+++ pytau/changepoint_io.py
@@ -394,7 +394,7 @@ class FitHandler:
                     "tau_array": picklable_dict.get("model_data", {}).get("tau_array"),
                     "processed_spikes": picklable_dict.get("model_data", {}).get("processed_spikes"),
                 }
-                
+
                 # Try to save approx.hist for ELBO plotting if available
                 approx_obj = picklable_dict.get("model_data", {}).get("approx")
                 if approx_obj and hasattr(approx_obj, 'hist'):
@@ -403,11 +403,12 @@ class FitHandler:
                         class SimpleApprox:
                             def __init__(self, hist):
                                 self.hist = hist
-                        model_data_fallback["approx"] = SimpleApprox(approx_obj.hist)
+                        model_data_fallback["approx"] = SimpleApprox(
+                            approx_obj.hist)
                     except Exception:
                         # If even hist fails to pickle, skip it
                         pass
-                
+
                 metadata_only_dict = {
                     "metadata": picklable_dict.get("metadata", {}),
                     "model_data": model_data_fallback
