### 2025-09-26 4304c76
Fix SimpleApprox class definition for proper pickling
Move SimpleApprox class to module level to make it picklable.
Classes defined inside functions cannot be pickled properly.

This fixes the AttributeError during fallback pickling when trying
to save the simplified approx object with hist attribute.

Co-authored-by: Ona <no-reply@ona.com>
