### 2025-09-28 f9e24eb
refactor(changepoint): switch to cloudpickle for serialization
