+++ pytau/changepoint_io.py
@@ -4,7 +4,7 @@ Pipeline to handle model fitting from data extraction to saving results
 
 import json
 import os
-import pickle
+import cloudpickle as pickle
 import shutil
 import uuid
 from datetime import date, datetime
@@ -384,42 +384,47 @@ class FitHandler:
             self.run_inference()
         out_dict = self._return_fit_output()
 
-        # Create a copy without the model to avoid pickling issues with PyMC5
-        picklable_dict = out_dict.copy()
-        if "model_data" in picklable_dict and "model" in picklable_dict["model_data"]:
-            picklable_model_data = picklable_dict["model_data"].copy()
-            # Remove the model object as it contains unpicklable local functions in PyMC5
-            picklable_model_data.pop("model", None)
-            picklable_dict["model_data"] = picklable_model_data
-
+        # Save output to pkl file
         with open(self.database_handler.model_save_path + ".pkl", "wb") as buff:
-            try:
-                pickle.dump(picklable_dict, buff)
-            except (TypeError, AttributeError) as e:
-                print(
-                    f"Warning: Full pickling failed ({e}). Saving metadata-only version.")
-                # If pickling fails, save only metadata and basic info
-                model_data_fallback = {
-                    "tau_array": picklable_dict.get("model_data", {}).get("tau_array"),
-                    "processed_spikes": picklable_dict.get("model_data", {}).get("processed_spikes"),
-                }
+            pickle.dump(out_dict, buff)
+        print(f"Saved full output to {self.database_handler.model_save_path}.pkl")
 
-                # Try to save approx.hist for ELBO plotting if available
-                approx_obj = picklable_dict.get("model_data", {}).get("approx")
-                if approx_obj and hasattr(approx_obj, 'hist'):
-                    try:
-                        # Create a simple object with just the hist attribute
-                        model_data_fallback["approx"] = SimpleApprox(
-                            approx_obj.hist)
-                    except Exception:
-                        # If even hist fails to pickle, skip it
-                        pass
-
-                metadata_only_dict = {
-                    "metadata": picklable_dict.get("metadata", {}),
-                    "model_data": model_data_fallback
-                }
-                pickle.dump(metadata_only_dict, buff)
+        # # Create a copy without the model to avoid pickling issues with PyMC5
+        # picklable_dict = out_dict.copy()
+        # if "model_data" in picklable_dict and "model" in picklable_dict["model_data"]:
+        #     picklable_model_data = picklable_dict["model_data"].copy()
+        #     # Remove the model object as it contains unpicklable local functions in PyMC5
+        #     picklable_model_data.pop("model", None)
+        #     picklable_dict["model_data"] = picklable_model_data
+        #
+        # with open(self.database_handler.model_save_path + ".pkl", "wb") as buff:
+        #     try:
+        #         pickle.dump(picklable_dict, buff)
+        #     except (TypeError, AttributeError) as e:
+        #         print(
+        #             f"Warning: Full pickling failed ({e}). Saving metadata-only version.")
+        #         # If pickling fails, save only metadata and basic info
+        #         model_data_fallback = {
+        #             "tau_array": picklable_dict.get("model_data", {}).get("tau_array"),
+        #             "processed_spikes": picklable_dict.get("model_data", {}).get("processed_spikes"),
+        #         }
+        #
+        #         # Try to save approx.hist for ELBO plotting if available
+        #         approx_obj = picklable_dict.get("model_data", {}).get("approx")
+        #         if approx_obj and hasattr(approx_obj, 'hist'):
+        #             try:
+        #                 # Create a simple object with just the hist attribute
+        #                 model_data_fallback["approx"] = SimpleApprox(
[…]
