### 2025-09-28 4b2312c
refactor(changepoint_analysis): validate input types in _firing and _tau classes
