### 2025-09-28 dd500ad
refactor(scripts): update base directory handling and data specifications
