### 2025-09-28 894d882
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
