### 2025-09-28 2579fa2
Add test results summary table to GitHub Actions and README
- Modified pytest workflow to generate JUnit XML output
- Added test-summary job to process test results and create markdown tables
- Created Python scripts to parse test results and generate summaries
- Updated README.md with test results table section
- Added GitHub step summary integration for better CI visibility

Fixes #141

Co-authored-by: Ona <no-reply@ona.com>
