+++ .github/workflows/pytest_workflow.yml
@@ -26,13 +26,73 @@ jobs:
         pip install -e .[dev]
     - name: Run tests
       run: |
-        pytest --cov=pytau -v
+        pytest --cov=pytau -v --tb=short --junitxml=test-results-${{ matrix.os }}-${{ matrix.python-version }}.xml
     - name: Run notebook tests
       run: |
-        pytest --nbmake pytau/how_to/model_notebooks/ -v
+        pytest --nbmake pytau/how_to/model_notebooks/ -v --tb=short --junitxml=notebook-test-results-${{ matrix.os }}-${{ matrix.python-version }}.xml
+    - name: Upload test results
+      uses: actions/upload-artifact@v4
+      if: always()
+      with:
+        name: test-results-${{ matrix.os }}-${{ matrix.python-version }}
+        path: |
+          test-results-${{ matrix.os }}-${{ matrix.python-version }}.xml
+          notebook-test-results-${{ matrix.os }}-${{ matrix.python-version }}.xml
+
+  test-summary:
+    needs: test
+    runs-on: ubuntu-latest
+    if: always()
+    steps:
+    - uses: actions/checkout@v4
+    - name: Download test results
+      uses: actions/download-artifact@v4
+      with:
+        path: test-results
+        pattern: test-results-*
+        merge-multiple: true
+    - name: Set up Python 3.11
+      uses: actions/setup-python@v4
+      with:
+        python-version: '3.11'
+    - name: Install dependencies for test summary
+      run: |
+        python -m pip install --upgrade pip
+        pip install junitparser
+    - name: Generate test summary
+      run: |
+        python .github/scripts/generate_test_summary.py
+    - name: Update README with test results
+      run: |
+        python .github/scripts/update_readme_tests.py
+    - name: Upload test summary
+      uses: actions/upload-artifact@v4
+      with:
+        name: test-summary
+        path: |
+          test_summary.md
+          test_results_table.md
+    - name: Commit updated README (main branch only)
+      if: github.event_name == 'push' && github.ref == 'refs/heads/main'
+      run: |
+        git config --local user.email "action@github.com"
+        git config --local user.name "GitHub Action"
+        git add README.md
+        if git diff --staged --quiet; then
+          echo "No changes to README - test results already up to date"
+        else
+          echo "Committing updated README with latest test results"
+          git commit -m "Auto-update README with latest test results
+
+          - Updated test results table with latest CI run
+          - Generated by GitHub Actions workflow
+          
+          [skip ci]"
+          git push
+        fi
 
   execute-notebooks:
-    needs: test
+    needs: test-summary
     runs-on: ubuntu-latest
     steps:
     - uses: actions/checkout@v4
