+++ .github/scripts/generate_test_summary.py
@@ -3,27 +3,29 @@
 Generate test summary from JUnit XML files and create GitHub step summary.
 """
 
-import os
 import glob
-from pathlib import Path
-from junitparser import JUnitXml
-from collections import defaultdict
+import os
 import sys
+from collections import defaultdict
+from pathlib import Path
+
+from junitparser import JUnitXml
+
 
 def parse_test_results():
     """Parse all JUnit XML files and extract test results."""
     results = defaultdict(lambda: defaultdict(dict))
-    
+
     # Find all XML files in test-results directory
     xml_files = glob.glob("test-results/*.xml")
-    
+
     if not xml_files:
         print("No test result files found!")
         return results
-    
+
     for xml_file in xml_files:
         print(f"Processing {xml_file}")
-        
+
         # Extract OS and Python version from filename
         filename = Path(xml_file).stem
         if filename.startswith('test-results-'):
@@ -40,23 +42,23 @@ def parse_test_results():
                 test_type = "notebook"
         else:
             continue
-            
+
         try:
             xml = JUnitXml.fromfile(xml_file)
-            
+
             total_tests = 0
             passed_tests = 0
             failed_tests = 0
             skipped_tests = 0
             errors = 0
-            
+
             for suite in xml:
                 total_tests += suite.tests
                 passed_tests += suite.tests - suite.failures - suite.errors - suite.skipped
                 failed_tests += suite.failures
                 errors += suite.errors
                 skipped_tests += suite.skipped
-            
+
             results[f"{os_name}-{python_version}"][test_type] = {
                 'total': total_tests,
                 'passed': passed_tests,
@@ -65,32 +67,33 @@ def parse_test_results():
                 'skipped': skipped_tests,
                 'success_rate': (passed_tests / total_tests * 100) if total_tests > 0 else 0
             }
-            
+
         except Exception as e:
             print(f"Error parsing {xml_file}: {e}")
             continue
-    
+
     return results
+
 
 def generate_markdown_table(results):
     """Generate a markdown table from test results."""
     if not results:
         return "No test results available."
-    
+
     # Table header
     table = "| Platform | Python | Unit Tests | Notebook Tests | Overall Status |\n"
     table += "|----------|--------|------------|----------------|----------------|\n"
-    
+
     for platform_python, test_types in sorted(results.items()):
         os_name, python_version = platform_python.split('-', 1)
-        
+
         # Format OS name
         os_display = {
             'ubuntu-latest': '🐧 Ubuntu',
-            'macos-latest': '🍎 macOS', 
+            'macos-latest': '🍎 macOS',
             'windows-latest': '🪟 Windows'
         }.get(os_name, os_name)
-        
+
         # Unit tests summary
         unit_summary = "N/A"
         if 'unit' in test_types:
@@ -98,7 +101,7 @@ def generate_markdown_table(results):
             unit_summary = f"✅ {unit['passed']}/{unit['total']} ({unit['success_rate']:.1f}%)"
             if unit['failed'] > 0 or unit['errors'] > 0:
                 unit_summary = f"❌ {unit['passed']}/{unit['total']} ({unit['success_rate']:.1f}%)"
-        
+
         # Notebook tests summary
         notebook_summary = "N/A"
         if 'notebook' in test_types:
@@ -106,30 +109,31 @@ def generate_markdown_table(results):
             notebook_summary = f"✅ {notebook['passed']}/{notebook['total']} ({notebook['success_rate']:.1f}%)"
             if notebook['failed'] > 0 or notebook['errors'] > 0:
[…]
