+++ .github/workflows/pytest_workflow.yml
@@ -86,7 +86,7 @@ jobs:
 
           - Updated test results table with latest CI run
           - Generated by GitHub Actions workflow
-          
+
           [skip ci]"
           git push
         fi
