### 2025-09-28 4827672
Restrict test-summary and execute-notebooks to main branch only
- test-summary job now only runs on push to main branch
- execute-notebooks job now only runs on push to main branch
- Keeps test matrix running on all PRs for validation
- Reduces CI resource usage on feature branches
- README updates only happen on main branch merges

Co-authored-by: Ona <no-reply@ona.com>
