### 2025-09-28 2e4ae97
Split test-summary workflow into two separate jobs
- test-summary: Always runs to generate test summary artifacts
- update-readme: Only runs on main branch pushes to update README
- Separates test summary generation from README updates
- Allows test summaries to be available for all PRs and pushes
- README updates only happen on merge to main

Co-authored-by: Ona <no-reply@ona.com>
