### 2025-09-29 8a63fdd
Delete citation section from README
Removed citation section from README.md
