### 2025-09-29 654413a
Change condition to always run update-readme job
