### 2025-09-29 843db63
Fix parameter name from 'sd' to 'sigma' in model
