+++ pytau/changepoint_model.py
@@ -469,7 +469,7 @@ class TDistributionChangepointModel(ChangepointModel):
         length = idx.max() + 1
 
         with pm.Model() as model:
-            mu = pm.Normal("mu", mu=mean_vals, sd=1, shape=(y_dim, n_states))
+            mu = pm.Normal("mu", mu=mean_vals, sigma=1, shape=(y_dim, n_states))
             sigma = pm.HalfCauchy("sigma", 3.0, shape=(y_dim, n_states))
             # Degrees of freedom
             nu = pm.Exponential("nu", 1 / 30, shape=(y_dim, n_states))
@@ -497,7 +497,7 @@ class TDistributionChangepointModel(ChangepointModel):
             mu_latent = mu.dot(weight_stack)
             sigma_latent = sigma.dot(weight_stack)
             observation = pm.StudentT(
-                "obs", nu=nu, mu=mu_latent, sd=sigma_latent, observed=data_array
+                "obs", nu=nu, mu=mu_latent, sigma=sigma_latent, observed=data_array
             )
 
         return model
