### 2025-09-29 fa47847
Replace nnet.sigmoid with math.sigmoid
