+++ pytau/changepoint_model.py
@@ -486,7 +486,7 @@ class TDistributionChangepointModel(ChangepointModel):
             tau = pm.Deterministic(
                 "tau", idx.min() + (idx.max() - idx.min()) * tau_latent)
 
-            weight_stack = tt.nnet.sigmoid(
+            weight_stack = tt.math.sigmoid(
                 idx[np.newaxis, :] - tau[:, np.newaxis])
             weight_stack = tt.concatenate(
                 [np.ones((1, length)), weight_stack], axis=0)
