### 2025-10-02 75967be
feat: Add elbow method option to find_best_states function
- Add piece-wise linear regression method to find optimal number of states
- New 'method' parameter allows choosing between 'min_elbo' (default) and 'elbow'
- Elbow method finds inflection point in ELBO curve using piecewise linear fit
- Add scikit-learn dependency for LinearRegression
- Maintains backward compatibility with existing min_elbo behavior

Resolves #146

Co-authored-by: Ona <no-reply@ona.com>
