+++ pyproject.toml
@@ -31,6 +31,7 @@ dependencies = [
     "graphviz==0.20.3",
     "easygui",
     "tables",
+    "scikit-learn",
 ]
 
 [project.optional-dependencies]
