### 2025-10-02 1ff7efd
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
