### 2025-10-03 11762e8
Add composable changepoint models framework
Implements issue #21 by creating a new composable submodule that allows
flexible combination of:
- Changepoint types (fixed vs Dirichlet process)
- Change types (sigmoid vs step transitions)
- Likelihood types (Gaussian vs Poisson)
- Emission models

Key components:
- Base classes for changepoint, emission, and likelihood components
- Concrete implementations for common model types
- ComposableModel class that combines components
- Convenience functions for common model combinations
- Comprehensive test suite
- Documentation and examples

This mainta
[…]
