### 2025-10-03 c7c5553
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
