### 2025-10-13 56bbc71
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/pycqa/isort: 6.0.1 → 7.0.0](https://github.com/pycqa/isort/compare/6.0.1...7.0.0)
