+++ .pre-commit-config.yaml
@@ -22,7 +22,7 @@ repos:
 
 
 -   repo: https://github.com/pycqa/isort
-    rev: 6.0.1
+    rev: 7.0.0
     hooks:
     -   id: isort
         args: ['--profile', 'black']
