### 2025-10-16 ddadde9
chore(docs): update Python version in README and remove outdated documentation
- Updated the Python version from 3.6.13 to 3.10 in the README.md to reflect the latest dependencies.
- Removed the outdated README.txt from the  directory as it contained obsolete information.
