### 2025-10-24 #8509 @fboehm
@TomDonoghue @justinshenk @alexander-pastukhov - I'd like to invite you to participate as a reviewer in this JOSS review. Our reviews are checklist-driven, so you would create a checklist as you review the submission. Thank you for considering this invitation.
