### 2025-10-28 b22ae85
Add GaussianChangepointMeanTrialSwitch model for multi-trial Gaussian data
- Implement new model class for detecting changepoints in multi-trial Gaussian data
- Model supports trial-to-trial emission changes with fixed changepoint distribution
- Add test case for the new model in test suite
- Update README documentation with new model description
- Follows same pattern as SingleTastePoissonTrialSwitch but for Gaussian likelihood

Fixes #157

Co-authored-by: Ona <no-reply@ona.com>
