+++ tests/test_changepoint_model.py
@@ -13,6 +13,7 @@ from pytau.changepoint_model import (
     ChangepointModel,
     GaussianChangepointMean2D,
     GaussianChangepointMeanDirichlet,
+    GaussianChangepointMeanTrialSwitch,
     GaussianChangepointMeanVar2D,
     PoissonChangepoint1D,
     SingleTastePoisson,
@@ -67,6 +68,8 @@ def test_gen_test_array():
         (GaussianChangepointMeanVar2D, (10, 100), 3, {}),
         (GaussianChangepointMeanDirichlet, (10, 100), None, {"max_states": 5}),
         (GaussianChangepointMean2D, (10, 100), 3, {}),
+        (GaussianChangepointMeanTrialSwitch,
+         (5, 10, 100), 3, {"switch_components": 2}),
         (SingleTastePoissonDirichlet, (5, 10, 100), None, {"max_states": 5}),
         (SingleTastePoisson, (5, 10, 100), 3, {}),
         # (SingleTastePoissonVarsig, (5, 10, 100), 3, {}),
@@ -90,6 +93,7 @@ def test_model_initialization(model_class, data_shape, n_states, extra_args):
             GaussianChangepointMeanVar2D,
             GaussianChangepointMeanDirichlet,
             GaussianChangepointMean2D,
+            GaussianChangepointMeanTrialSwitch,
         ]
         else "poisson"
     )
