+++ pytau/changepoint_model.py
@@ -460,6 +460,199 @@ def gaussian_changepoint_mean_2d(data_array, n_states, **kwargs):
     return model_class.generate_model()
 
 
+class GaussianChangepointMeanTrialSwitch(ChangepointModel):
+    """
+    Gaussian model with trial switches
+    Assuming only emissions change across trials
+    Changepoint distribution remains constant
+    """
+
+    def __init__(self, data_array, switch_components, n_states, **kwargs):
+        """
+        Args:
+            data_array (3D Numpy array): trials x dimensions x time
+            switch_components (int): Number of trial switch components
+            n_states (int): Number of states to model
+            **kwargs: Additional arguments
+        """
+        super().__init__(**kwargs)
+        self.data_array = data_array
+        self.switch_components = switch_components
+        self.n_states = n_states
+
+    def generate_model(self):
+        """
+        Returns:
+            pymc model: Model class containing graph to run inference on
+        """
+        data_array = self.data_array
+        switch_components = self.switch_components
+        n_states = self.n_states
+
+        trial_num, y_dim, time_bins = data_array.shape
+
+        with pm.Model() as model:
+            # Define Emissions (means)
+
+            # dims
+            dim_mu = pm.Normal("dim_mu", mu=0, sigma=5, shape=(y_dim))
+
+            # dims x switch_comps
+            trial_mu = pm.Normal(
+                "trial_mu",
+                mu=dim_mu.dimshuffle(0, "x"),
+                sigma=5,
+                shape=(y_dim, switch_components),
+            )
+
+            # dims x switch_comps x n_states
+            state_mu = pm.Normal(
+                "state_mu",
+                mu=trial_mu.dimshuffle(0, 1, "x"),
+                sigma=5,
+                shape=(y_dim, switch_components, n_states),
+            )
+
+            # One variance for each dimension
+            sigma = pm.HalfCauchy("sigma", 3.0, shape=(y_dim))
+
+            # Define Changepoints
+            # Assuming distribution of changepoints remains
+            # the same across all trials
+
+            a = pm.HalfCauchy("a_tau", 3.0, shape=n_states - 1)
+            b = pm.HalfCauchy("b_tau", 3.0, shape=n_states - 1)
+
+            even_switches = np.linspace(0, 1, n_states + 1)[1:-1]
+            tau_latent = pm.Beta(
+                "tau_latent", a, b,
+                shape=(trial_num, n_states - 1)
+            ).sort(axis=-1)
+
+            # Trials x Changepoints
+            tau = pm.Deterministic("tau", time_bins * tau_latent)
+
+            # Define trial switches
+            # Will have same structure as regular changepoints
+
+            even_trial_switches = np.linspace(
+                0, 1, switch_components + 1)[1:-1]
+            tau_trial_latent = pm.Beta(
+                "tau_trial_latent",
+                1,
+                1,
+                initval=even_trial_switches,
+                shape=(switch_components - 1),
+            ).sort(axis=-1)
+
+            # Trial_changepoints
+            tau_trial = pm.Deterministic(
+                "tau_trial", trial_num * tau_trial_latent)
+
+            trial_idx = np.arange(trial_num)
+            trial_selector = tt.math.sigmoid(
+                trial_idx[np.newaxis, :] - tau_trial.dimshuffle(0, "x")
+            )
+
+            trial_selector = tt.concatenate(
+                [np.ones((1, trial_num)), trial_selector], axis=0)
+            inverse_trial_selector = 1 - trial_selector[1:, :]
+            inverse_trial_selector = tt.concatenate(
+                [inverse_trial_selector, np.ones((1, trial_num))], axis=0
+            )
+
[…]
