### 2025-10-28 85885ed
Add GaussianChangepointMean3D model for multi-trial Gaussian data
- Implement new model class for multi-trial Gaussian changepoint detection
- Model handles 3D data (trials x dimensions x time) without trial switches
- Similar to SingleTastePoisson but uses Gaussian likelihood
- Detects changes in mean only, with shared variance per dimension
- Add test case for the new model in test suite
- Update README documentation with new model description

Co-authored-by: Ona <no-reply@ona.com>
