+++ pytau/changepoint_model.py
@@ -653,6 +653,123 @@ def gaussian_changepoint_mean_trial_switch(data_array, switch_components, n_stat
     return model_class.generate_model()
 
 
+class GaussianChangepointMean3D(ChangepointModel):
+    """Model for Gaussian changepoint on multi-trial data
+    
+    Detects changes in mean only across trials.
+    Similar to SingleTastePoisson but for Gaussian data.
+    No hierarchical structure for emissions.
+    """
+
+    def __init__(self, data_array, n_states, **kwargs):
+        """
+        Args:
+            data_array (3D Numpy array): trials x dimensions x time
+            n_states (int): Number of states to model
+            **kwargs: Additional arguments
+        """
+        super().__init__(**kwargs)
+        self.data_array = data_array
+        self.n_states = n_states
+
+    def generate_model(self):
+        """
+        Returns:
+            pymc model: Model class containing graph to run inference on
+        """
+        data_array = self.data_array
+        n_states = self.n_states
+
+        mean_vals = np.array(
+            [np.mean(x, axis=-1)
+             for x in np.array_split(data_array, n_states, axis=-1)]
+        ).T
+        mean_vals = np.mean(mean_vals, axis=1)
+        mean_vals += 0.01  # To avoid zero starting prob
+
+        y_dim = data_array.shape[1]
+        trials = data_array.shape[0]
+        idx = np.arange(data_array.shape[-1])
+        length = idx.max() + 1
+
+        with pm.Model() as model:
+            # Emissions: mean for each dimension and state
+            mu = pm.Normal("mu", mu=mean_vals, sigma=5,
+                          shape=(y_dim, n_states))
+            
+            # One variance for each dimension
+            sigma = pm.HalfCauchy("sigma", 3.0, shape=(y_dim))
+
+            # Changepoints
+            a_tau = pm.HalfCauchy("a_tau", 3.0, shape=n_states - 1)
+            b_tau = pm.HalfCauchy("b_tau", 3.0, shape=n_states - 1)
+
+            even_switches = np.linspace(0, 1, n_states + 1)[1:-1]
+            tau_latent = pm.Beta(
+                "tau_latent",
+                a_tau,
+                b_tau,
+                shape=(trials, n_states - 1),
+            ).sort(axis=-1)
+
+            tau = pm.Deterministic(
+                "tau", idx.min() + (idx.max() - idx.min()) * tau_latent)
+
+            # Create weight stack for changepoints
+            weight_stack = tt.math.sigmoid(
+                idx[np.newaxis, :] - tau[:, :, np.newaxis])
+            weight_stack = tt.concatenate(
+                [np.ones((trials, 1, length)), weight_stack], axis=1)
+            inverse_stack = 1 - weight_stack[:, 1:]
+            inverse_stack = tt.concatenate(
+                [inverse_stack, np.ones((trials, 1, length))], axis=1)
+            weight_stack = np.multiply(weight_stack, inverse_stack)
+
+            # Compute latent means: trials x dims x time
+            mu_latent = tt.tensordot(weight_stack, mu, [
+                                     1, 1]).swapaxes(1, 2)
+            
+            # Broadcast sigma to match data shape
+            sigma_latent = sigma.dimshuffle("x", 0, "x")
+            
+            observation = pm.Normal("obs", mu=mu_latent, sigma=sigma_latent, 
+                                   observed=data_array)
+
+        return model
+
+    def test(self):
+        """Test the model with synthetic data"""
+        # Generate test data
+        test_data = gen_test_array(
+            (5, 10, 100), n_states=self.n_states, type="normal")
+
+        # Create model with test data
+        test_model = GaussianChangepointMean3D(test_data, self.n_states)
+        model = test_model.generate_model()
+
[…]
