### 2025-10-28 dbaf701
Add Student's t-distribution changepoint models for robust outlier handling
- Implement TChangepointMean2D for 2D data (dimensions x time)
- Implement TChangepointMean3D for 3D multi-trial data (trials x dimensions x time)
- Both models use Student's t-distribution instead of Gaussian for robustness to outliers
- Include degrees of freedom parameter (nu) with Gamma prior
- Add test cases for both models in test suite
- Update README documentation with new model descriptions
- Follow same pattern as Gaussian models but with t-distribution likelihood

Fixes #93

Co-authored-by: Ona <no-reply@ona
[…]
