### 2025-11-03 e441d6e
docs: Add PyMC concepts guide for non-PyMC users
- Create comprehensive pymc_concepts.md explaining Model, Samples, Trace, and Posterior Predictive
- Include practical examples and code snippets from PyTau
- Add troubleshooting section and links to PyMC documentation
- Update mkdocs.yml navigation to include new concepts page

Closes #155

Co-authored-by: Ona <no-reply@ona.com>
