+++ docs/pymc_concepts.md
@@ -1,0 +1,421 @@
+# Understanding PyMC Concepts in PyTau
+
+This guide explains the key PyMC concepts used in PyTau for users who may not be familiar with Bayesian modeling or PyMC. For more detailed information, refer to the [official PyMC documentation](https://www.pymc.io/welcome.html).
+
+## Overview
+
+PyTau uses [PyMC](https://www.pymc.io/), a probabilistic programming library, to perform Bayesian inference on changepoint models. This approach allows us to:
+
+- Quantify uncertainty in changepoint locations
+- Incorporate prior knowledge about neural dynamics
+- Obtain full posterior distributions rather than point estimates
+
+## Core Concepts
+
+### 1. Model
+
+A **model** in PyMC is a probabilistic representation of your data and assumptions. It consists of:
+
+- **Prior distributions**: What we believe about parameters before seeing data
+- **Likelihood**: How the data is generated given the parameters
+- **Observed data**: The actual measurements we want to explain
+
+In PyTau, models are defined as classes that inherit from `ChangepointModel` and implement a `generate_model()` method.
+
+#### Example: Single Taste Poisson Model
+
+```python
+from pytau.changepoint_model import SingleTastePoisson
+
+# Create a model for spike train data
+# spike_array shape: (trials, neurons, time)
+model_obj = SingleTastePoisson(
+    spike_array=spike_array,
+    states=3,  # Number of states to detect
+    fit_type='advi'
+)
+
+# Generate the PyMC model
+model = model_obj.generate_model()
+```
+
+**What happens inside the model:**
+
+1. **Priors on changepoint times (tau)**: Where state transitions might occur
+2. **Priors on emission rates (lambda)**: Firing rates in each state
+3. **Likelihood**: Poisson distribution connecting rates to observed spike counts
+
+The model structure looks like this:
+
+```python
+with pm.Model() as model:
+    # Prior on changepoint times (sorted)
+    tau = pm.Uniform('tau', lower=0, upper=time_bins, shape=states-1)
+    tau_sorted = tt.sort(tau)
+    
+    # Prior on emission rates for each state
+    lambda_latent = pm.Exponential('lambda', lam=1.0, shape=states)
+    
+    # Assign rates to time bins based on changepoints
+    # (simplified - actual implementation more complex)
+    lambda_t = assign_rates_by_changepoints(tau_sorted, lambda_latent)
+    
+    # Likelihood: observed spikes follow Poisson distribution
+    obs = pm.Poisson('obs', mu=lambda_t, observed=spike_array)
+```
+
+For more details on PyMC models, see the [PyMC Model Building Guide](https://www.pymc.io/projects/docs/en/stable/learn/core_notebooks/pymc_overview.html).
+
+---
+
+### 2. Inference: Samples and Trace
+
+Once we have a model, we need to perform **inference** to learn about the parameters from the data. PyMC offers two main approaches:
+
+#### Variational Inference (ADVI)
+
+**ADVI** (Automatic Differentiation Variational Inference) is a fast approximate inference method that:
+
+- Finds a simpler distribution that approximates the true posterior
+- Is much faster than MCMC but less accurate
+- Good for initial exploration or large datasets
+
+```python
+from pytau.changepoint_model import advi_fit
+
+# Fit the model using ADVI
+model, approx, lambda_stack, tau_samples, data = advi_fit(
+    model,
+    fit=10000,      # Number of optimization iterations
[…]
