### 2025-11-03 83c3863
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
