### 2025-11-03 342d8c0
fix: Correct model parameter names in PyMC concepts docs
- Remove non-existent fit_type parameter
- Change spike_array to data_array (correct parameter name)
- Change states to n_states (correct parameter name)
- Fix best_model attribute access (n_states instead of kwargs['states'])

All examples now match actual PyTau API

Co-authored-by: Ona <no-reply@ona.com>
