### 2025-11-11 562e8fc
fix: Separate check_mismatched_paths into two functions
Refactored check_mismatched_paths() into two separate functions:
- check_mismatched_files(): Checks for pkl files without database entries
- check_mismatched_rows(): Checks for database rows without pkl files

The original check_mismatched_paths() now calls both functions to maintain
backward compatibility with existing code.

Fixes #162

Co-authored-by: Ona <no-reply@ona.com>
