### 2025-11-11 24df00a
feat: Return row indices from check_mismatched_rows
Updated check_mismatched_rows() to return both the boolean list and the
actual dataframe row indices of mismatched rows. This provides more
useful information for debugging and selective row operations.

Updated check_mismatched_paths() to handle the new return signature
while maintaining backward compatibility.

Co-authored-by: Ona <no-reply@ona.com>
