### 2025-11-11 fb79933
fix: Add elbo_hist property to PklHandler for easy ELBO access
Fixes #154

The MeanField approximation object does not support direct indexing,
which caused confusion when users tried to access ELBO values using
approx[-1]. This commit adds a convenient elbo_hist property to
PklHandler that provides direct access to the ELBO history array.

Changes:
- Add elbo_hist property to PklHandler class that returns the hist
  attribute from the approximation object
- Add comprehensive tests for the new property
- Property returns None when approx is not available

Users can now access ELBO values using
[…]
