### 2025-11-13 #8509 @gillescourtand
@fboehm Thank you for the invitation, I'm sorry, this topic is not really my area of expertise, so I will have to decline the offer.
