### 2025-11-13 7c05c2f
fix: Implement sharp sigmoid for changepoint transitions
Replace default sigmoid with custom sharp_sigmoid function that uses
a steeper slope (100) to enforce sharper state changes while maintaining
smoothness for VI and HMC. This addresses issues with large binning
where the default exponential timescale can be problematic.

Fixes #165

Co-authored-by: Ona <no-reply@ona.com>
