### 2025-11-13 fe3bca0
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
