### 2025-11-13 abuzarmahmood/pytau#167 @HectorMozo3110
README Suggestions (JOSS Review )
👋 Hi @abuzarmahmood  - after reviewing the README, it has a good foundation, but a few adjustments could help make it clearer and easier for new users to follow. Below I’ve outlined several points for consideration, and we can use this thread to go through each of them until everything related to the README is addressed.

#### 1. Improve CI Transparency
The README currently includes a static operating-system test table. Because this table does not update automatically and may become inconsistent with the current CI status, I recommend removing it entirely. A dynamic CI badge is sufficient and provides accurate, real-time information about the health of the repository.

#### 2. Strengthen the Introductory Context
The “What is PyTau?” section would benefit from an expanded description explaining what the software does, what scientific problem it addresses, and who the intended users are. Adding a brief contextual paragraph will help new readers understand the purpose and relevance of the package more effectively.

#### 3. Installation Section Improvements
In the Quick Start section, the only installation path shown involves cloning the repository and installing all dependencies manually using `pip install -e .`. I did not see a direct PyPI installation command, nor any alternative installation method aside from cloning. It would improve clarity to explicitly provide both a PyPI installation option and a GitHub-based installation option. Additionally, the only version currently visible on PyPI appears to be version 0.1.1 from 2022, whereas a newer release is present on GitHub, so noting this distinction in the README would help users understand how to obtain the most up-to-date version of the software.

#### 4. Provide Local Documentation Build Instructions

It would be helpful to include a brief note on how to build the documentation locally, or to direct users to where these instructions can be found. This would assist collaborators who want to make changes or contribute, as well as anyone who needs to copy or review the documentation locally.

#### 5. Add Basic Testing Instructions
It would be useful to include brief instructions on how to run the test suite and how to generate coverage reports. This information can be added either in the README or in the documentation, wherever it fits best, but having it available is important for anyone who needs to verify local test coverage or contribute changes.

#### 6. Keep Only High-Level Feature Summaries
It would be clearer if the README included only high-level summaries of the main features, model categories, inference approaches, data formats, and the general pipeline. At the moment, this is not clearly distinguished, as the README mixes concise overviews with highly detailed technical descriptions. Keeping the summaries brief and moving detailed explanations to the documentation would make the README easier to read and more effective as an introduction.

#### 7. Move Detailed Technical Content to the Documentation
Extended lists of model classes, long inference examples, internal storage architecture details, advanced pipeline descriptions, and parallelization notes should be moved from the README into the main documentation. The README should remain compact and act as an entry point that directs users to more detailed pages.

#### 8. Reference, but Do Not Include, the Minimal Example

Rather than embedding long code blocks directly in the README, it would be more effective to include a brief sentence that points users to the minimal usage example in the documentation or in an example notebook. This helps keep the README focused and easy to navigate.

You may still choose to include a very short example in the README and show its output, but for full details, linking directly to the documentation is preferable. This approach keeps the README from becoming overloaded while still guiding users to more complete information when needed.

#### 9. Contribution Section Context

The contribution section in the README is fine as it is, and I noticed that this is where the pytest and coverage instructions are located. Mentioning this briefly in the README or adding a small testing note there would help address the testing clarity mentioned earlier in ***point 5.***

---

**These are my initial observations. I will continue updating this thread with any further suggestions as I identify them. If you need clarification on any point, feel free to ask here. Once everything has been addressed, I will consider this issue complete.**

---
**Note:** This observation is part of the ongoing JOSS review (see https://github.com/openjournals/jossreviews/issues/https://github.com/openjournals/joss-reviews/issues/8509)
