### 2025-11-19 35bf2cb
docs: Improve README clarity and structure per JOSS review
Address issue #167 with comprehensive README improvements:

1. Remove static test results table - CI badge provides real-time status
2. Expand 'What is PyTau?' with context on scientific problem and users
3. Improve installation section with PyPI and GitHub options
4. Add note about PyPI version (0.1.1) vs latest GitHub version
5. Add local documentation build instructions
6. Add testing instructions (pytest and coverage)
7. Condense feature summaries and move details to documentation
8. Move detailed technical content to documentation
[…]
