+++ README.md
@@ -8,32 +8,36 @@
   [![Pytest](https://github.com/abuzarmahmood/pytau/actions/workflows/pytest_workflow.yml/badge.svg)](https://github.com/abuzarmahmood/pytau/actions/workflows/pytest_workflow.yml)
 </div>
 
-## 🧪 Test Results
-
-| Platform | Python | Unit Tests | Notebook Tests | Overall Status |
-|----------|--------|------------|----------------|----------------|
-| 🐧 Ubuntu | 3.10 | ✅ Passing | ✅ Passing | ✅ Pass |
-| 🐧 Ubuntu | 3.11 | ✅ Passing | ✅ Passing | ✅ Pass |
-| 🍎 macOS | 3.10 | ✅ Passing | ✅ Passing | ✅ Pass |
-| 🍎 macOS | 3.11 | ✅ Passing | ✅ Passing | ✅ Pass |
-| 🪟 Windows | 3.10 | ✅ Passing | ✅ Passing | ✅ Pass |
-| 🪟 Windows | 3.11 | ✅ Passing | ✅ Passing | ✅ Pass |
-
-*Test results are automatically updated by CI/CD pipeline*
-
 ## 🚀 What is PyTau?
 
-PyTau is a specialized Python package for detecting state changes in neural data using Bayesian changepoint models. It provides:
+PyTau is a specialized Python package for detecting state changes in neural data using Bayesian changepoint models. Understanding when neural activity transitions between different states is crucial for analyzing how the brain processes information, particularly in sensory neuroscience and behavioral studies. PyTau addresses this challenge by providing a comprehensive toolkit for identifying these state transitions in neural firing patterns.
 
-- **Streamlined batch inference** on PyMC3-based changepoint models
-- **Robust detection** of state transitions in neural firing patterns
-- **Comprehensive visualization tools** for model results
+The package is designed for neuroscientists and computational researchers working with spike train data from electrophysiological recordings. It enables users to:
+
+- Detect when neurons change their firing patterns in response to stimuli or behavioral events
+- Identify the number and timing of distinct neural states in an experiment
+- Compare state transitions across different experimental conditions, brain regions, or subjects
+- Quantify uncertainty in state transition timing using Bayesian inference
+
+PyTau builds on PyMC3 to provide flexible, extensible models while offering a streamlined interface for batch processing and analysis of large-scale neural datasets.
 
 ## 📚 Documentation
 
 [**Full API Documentation**](https://abuzarmahmood.github.io/pytau/)
 
-## ⚡ Quick Start
+For detailed usage examples, model descriptions, and advanced features, please refer to the [complete documentation](https://abuzarmahmood.github.io/pytau/).
+
+## ⚡ Installation
+
+### Install from PyPI (Stable Release)
+
+```bash
+pip install pytau
+```
+
+**Note:** The PyPI version (0.1.1) is from 2022 and may not include the latest features. For the most up-to-date version, install from GitHub (see below).
+
+### Install from GitHub (Latest Version)
 
 ```bash
 # Create and activate conda environment
@@ -42,165 +46,76 @@ conda activate pytau_env
 
 # Clone repository
 git clone https://github.com/abuzarmahmood/pytau.git
-
-# Install package
 cd pytau
-pip install -e .  # Install in development mode
 
-# Alternative: Install from requirements file
-# pip install -r requirements.txt
+# Install in development mode
+pip install -e .
+```
 
-# Download test data and run example notebook
+### Quick Start Example
+
+After installation, download test data and explore the example notebooks:
+
+```bash
 cd pytau/how_to
 bash scripts/download_test_data.sh
 cd notebooks
 jupyter notebook
 ```
 
[…]
