### 2025-11-20 e0046ea
docs: Add comprehensive documentation pages for removed README content
Add four new documentation pages to retain detailed information that was
removed from README for clarity:

- models.md: Complete guide to all available models with use cases
- inference.md: Detailed inference methods (ADVI, MCMC, model selection)
- data_formats.md: Data format specifications and preprocessing guide
- pipeline.md: Pipeline architecture and batch processing workflows

Update mkdocs.yml navigation to organize docs into User Guide section
with logical flow from models → inference → data → pipeline.

These pages
[…]
