+++ docs/pipeline.md
@@ -1,0 +1,580 @@
+# Pipeline Architecture
+
+PyTau provides a modular pipeline architecture for reproducible, large-scale changepoint analysis. This page describes the pipeline components, batch processing capabilities, and best practices for managing large-scale analyses.
+
+## Overview
+
+PyTau's pipeline consists of four main components:
+
+1. **Model Generation**: Creates and fits Bayesian changepoint models
+2. **Data Preprocessing**: Prepares raw neural data for modeling
+3. **I/O Management**: Handles data loading, model storage, and database management
+4. **Batch Processing**: Enables large-scale model fitting across datasets
+
+## Pipeline Components
+
+### 1. Model Generation
+
+**Purpose**: Creates and fits Bayesian changepoint models
+
+**Key Module**: `pytau.changepoint_model`
+
+**Components**:
+- Model classes for different data types (Poisson, Gaussian, Categorical)
+- Inference functions (ADVI, MCMC)
+- Model selection utilities
+
+**Example**:
+```python
+from pytau.changepoint_model import single_taste_poisson, advi_fit
+
+# Create model
+model = single_taste_poisson(data, states=3)
+
+# Fit model
+model, approx = advi_fit(model, fit=10000, samples=5000)
+```
+
+**Input**: Processed spike trains and model parameters
+
+**Output**: Fitted model with posterior samples
+
+### 2. Data Preprocessing
+
+**Purpose**: Prepares raw neural data for modeling
+
+**Key Module**: `pytau.changepoint_preprocess`
+
+**Features**:
+- Spike train binning
+- Trial shuffling for null distributions
+- Synthetic data generation
+- Data validation and cleaning
+
+**Example**:
+```python
+from pytau.changepoint_preprocess import preprocess_spike_data
+
+# Preprocess data
+processed_data = preprocess_spike_data(
+    raw_spikes,
+    bin_width=0.05,
+    time_lims=(0, 2.0),
+    shuffle=False
+)
+```
+
+**Input**: Raw spike trains with preprocessing parameters
+
+**Output**: Binned, processed spike count data
+
+### 3. I/O Management
+
+**Purpose**: Handles data loading, model storage, and database management
+
+**Key Module**: `pytau.changepoint_io`
+
+**Components**:
+- `FitHandler`: Manages model fitting pipeline
+- `DatabaseHandler`: Manages model database
+- HDF5 data loading utilities
+- Model serialization
+
+**Example**:
+```python
+from pytau.changepoint_io import FitHandler
+
+# Create handler
+handler = FitHandler(
+    data_dir='/path/to/data',
+    save_dir='/path/to/models',
+    animal_name='mouse1',
+    session_date='2024-01-15',
+    taste_name='sucrose',
+    region_name='GC'
+)
+
+# Fit and save model
+handler.fit_and_save(
+    model_type='single_taste_poisson',
+    states=3,
+    fit_steps=10000
+)
+```
+
+**Operations**:
+- HDF5 data loading
+- Metadata tracking
+- Database integration
+- Automatic model retrieval or fitting
+
+### 4. Batch Processing
+
+**Purpose**: Enables large-scale model fitting across datasets
+
+**Key Module**: `pytau.utils.batch_utils`
+
+**Features**:
+- Parameter grid iteration
+- Parallel processing support
+- Progress tracking
+- Error handling and logging
+
+**Example**:
+```python
+from pytau.utils.batch_utils import batch_fit_models
+
+# Define parameter grid
+param_grid = {
+    'states': [2, 3, 4, 5],
+    'bin_width': [0.025, 0.05, 0.1],
+    'model_type': ['single_taste_poisson', 'single_taste_poisson_varsig']
+}
+
+# Batch fit
+results = batch_fit_models(
+    data_dir='/path/to/data',
+    param_grid=param_grid,
[…]
