+++ docs/models.md
@@ -1,0 +1,234 @@
+# Available Models
+
+PyTau provides a comprehensive suite of changepoint detection models designed for different types of neural data and experimental designs. This page details all available models and their use cases.
+
+## Model Selection Guide
+
+| Data Type | Recommended Models |
+|-----------|-------------------|
+| Single neuron time series | `PoissonChangepoint1D` |
+| Multiple trials, single taste | `SingleTastePoisson*` |
+| Multiple tastes | `AllTastePoisson*` |
+| Continuous neural data | `GaussianChangepoint*` |
+| Categorical behavioral data | `CategoricalChangepoint2D` |
+| Unknown state number | `*Dirichlet` models |
+
+## Poisson Models
+
+Poisson models are designed for spike count data from neural recordings. They model the firing rate of neurons and detect when these rates change over time.
+
+### PoissonChangepoint1D
+
+**Purpose**: 1D time series changepoint detection
+
+**Use Case**: Single neuron or single trial analysis where you want to detect changes in firing rate over time.
+
+**Data Shape**: `(time,)`
+
+**Key Features**:
+- Simplest Poisson model
+- Ideal for exploratory analysis
+- Fast inference
+
+### SingleTastePoisson
+
+**Purpose**: Basic single-taste model with fixed sigmoid transitions
+
+**Use Case**: Multi-trial experiments with a single stimulus (taste) where you want to detect consistent changepoints across trials.
+
+**Data Shape**: `(trials, neurons, time)` or `(trials, time)`
+
+**Key Features**:
+- Fixed sigmoid transition shape
+- Assumes consistent changepoint timing across trials
+- Hierarchical structure across neurons
+
+### SingleTastePoissonVarsig
+
+**Purpose**: Variable sigmoid slope inferred from data
+
+**Use Case**: When transition sharpness is unknown and should be learned from data.
+
+**Data Shape**: `(trials, neurons, time)` or `(trials, time)`
+
+**Key Features**:
+- Infers transition sharpness from data
+- More flexible than fixed sigmoid
+- Longer inference time
+
+### SingleTastePoissonVarsigFixed
+
+**Purpose**: Fixed sigmoid slope with configurable sharpness
+
+**Use Case**: When you want to specify transition sharpness based on prior knowledge.
+
+**Data Shape**: `(trials, neurons, time)` or `(trials, time)`
+
+**Key Features**:
+- User-specified transition sharpness
+- Faster than variable sigmoid
+- Good for batch processing with consistent parameters
+
+### SingleTastePoissonDirichlet
+
+**Purpose**: Automatic state number detection using Dirichlet process
+
+**Use Case**: When the number of states is unknown and should be inferred from data.
+
+**Data Shape**: `(trials, neurons, time)` or `(trials, time)`
+
+**Key Features**:
+- Automatically determines optimal number of states
+- Uses Dirichlet process prior
+- Requires MCMC sampling (no ADVI support)
+- Longer inference time but more flexible
+
+### SingleTastePoissonTrialSwitch
+
+**Purpose**: Models trial-to-trial emission changes
+
+**Use Case**: When changepoint times are consistent but firing rates vary across trials.
+
+**Data Shape**: `(trials, neurons, time)` or `(trials, time)`
+
+**Key Features**:
+- Fixed changepoint times across trials
+- Trial-specific emission rates
+- Useful for detecting trial-to-trial variability
+
+### AllTastePoisson
+
+**Purpose**: Hierarchical model across multiple tastes
+
+**Use Case**: Multi-taste experiments where you want to detect both taste-specific and shared changepoints.
+
+**Data Shape**: `(tastes, trials, neurons, time)`
+
+**Key Features**:
+- Hierarchical structure across tastes
[…]
