### 2025-11-20 28e9be3
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
