### 2025-11-21 88a4148
fix: Ensure preprocessed_data is saved when advi_fit returns None
Fixes #170

When advi_fit returns None for the data field (to avoid PyMC5
compatibility issues), the FitHandler now uses the preprocessed_data
that was actually used for inference. This ensures that when PklHandler
loads the saved pkl file, it has the actual spike data available instead
of None, preventing the AssertionError in _firing initialization.

Co-authored-by: Ona <no-reply@ona.com>
