### 2025-11-21 07b1171
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
