### 2025-11-21 eb6437e
fix: Use .get() to safely check for 'data' key
Prevents KeyError when inference_outs doesn't have 'data' key
(e.g., in tests with mock inference functions)

Co-authored-by: Ona <no-reply@ona.com>
