### 2025-11-21 9a12f15
fix: Ensure raw_mode_tau has correct shape for multi-trial data
Apply the same shape handling logic to raw_mode_tau that was already
used for scaled_mode_tau. This ensures raw_mode_tau is a 2D array
(n_trials, n_changepoints) when n_trials is provided, which is required
by get_state_firing and other functions that expect this shape.

Fixes the IndexError: tuple index out of range that occurred when
get_state_firing tried to access tau_array.shape[-1].

Co-authored-by: Ona <no-reply@ona.com>
