### 2025-11-21 7e03153
ci: Skip Bayesian_Changepoint_Model.ipynb in workflow tests
Update the GitHub Actions workflow to skip the deprecated notebook
that requires pymc3.

Co-authored-by: Ona <no-reply@ona.com>
