### 2025-11-21 4058605
ci: Allow no tests collected for examples directory
Since we're ignoring the only notebook in the examples directory,
pytest returns exit code 5 (no tests collected). Allow this to pass.

Co-authored-by: Ona <no-reply@ona.com>
