+++ .github/workflows/pytest_workflow.yml
@@ -41,7 +41,8 @@ jobs:
 
         echo "Testing example notebooks..."
         # Skip Bayesian_Changepoint_Model.ipynb as it requires pymc3 (deprecated)
-        pytest --nbmake pytau/how_to/examples/ -v --tb=short --ignore=pytau/how_to/examples/Bayesian_Changepoint_Model.ipynb
+        # Allow no tests collected since we're ignoring the only notebook
+        pytest --nbmake pytau/how_to/examples/ -v --tb=short --ignore=pytau/how_to/examples/Bayesian_Changepoint_Model.ipynb || [ $? -eq 5 ]
 
         echo "Notebook testing completed"
       shell: bash
