### 2025-11-29 abuzarmahmood/pytau#167 @HectorMozo3110
Thank you for raising this,@duodenum96  - I completely agree with your point.

@abuzarmahmood, what are your thoughts on this?

And just doing a quick follow-up on what we discussed earlier, how is it going so far?
