### 2025-11-30 #8509 @HectorMozo3110
Hi 👋 @fboehm

Regarding the reviews: I'm currently waiting for the author's response to the points I raised in [Issue #167](https://github.com/abuzarmahmood/pytau/issues/167), which focus on the README improvements. Those points need to be clarified before I can continue with the rest of the review, especially since they cover important parts of the documentation and installation process.

Once the author replies to [Issue #167](https://github.com/abuzarmahmood/pytau/issues/167), I’ll proceed with the remaining sections. For now, I’m just waiting for their feedback so I can move forward.
