### 2025-12-03 a82f726
chore(ci): remove deprecated notebook tests
- Removed testing for example notebooks that include deprecated dependencies.
- Updated CI workflow to reflect the removal of Bayesian_Changepoint_Model.ipynb testing.
- Simplified the workflow by eliminating unnecessary test commands.
