+++ tox.ini
@@ -20,8 +20,6 @@ commands =
     # Run with: tox -- --nbmake-timeout=300 to increase timeout for slow notebooks
     - pytest --nbmake pytau/how_to/model_notebooks/ --nbmake-timeout=300
     - pytest --nbmake pytau/how_to/notebooks/ --nbmake-timeout=300
-    # Skip Bayesian_Changepoint_Model.ipynb as it requires pymc3 (deprecated)
-    - pytest --nbmake pytau/how_to/examples/ --nbmake-timeout=300 --ignore=pytau/how_to/examples/Bayesian_Changepoint_Model.ipynb
 
 [pytest]
 testpaths = tests
