### 2025-12-03 2addd7a
ci: reorganize matrix tests with primary job first
Co-authored-by: aider (anthropic/claude-sonnet-4-20250514) <aider@aider.chat>
