+++ .github/workflows/pytest_workflow.yml
@@ -7,12 +7,54 @@ on:
     branches: [ '**' ]
 
 jobs:
-  test:
+  test-primary:
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v4
+    - name: Set up Python 3.11
+      uses: actions/setup-python@v4
+      with:
+        python-version: '3.11'
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install -e .[dev]
+    - name: Run tests
+      run: |
+        pytest --cov=pytau -v --tb=short --junitxml=test-results-ubuntu-latest-3.11.xml
+    - name: Run notebook tests
+      run: |
+        # Try to run notebooks with nbmake (should now work with dummy matplotlib)
+        echo "Testing model notebooks..."
+        pytest --nbmake pytau/how_to/model_notebooks/ -v --tb=short --junitxml=notebook-test-results-ubuntu-latest-3.11.xml
+
+        # Test notebook structure without execution
+        pytest tests/test_how_to_notebooks.py -v
+
+        echo "Testing how-to notebooks..."
+        pytest --nbmake pytau/how_to/notebooks/ -v --tb=short
+
+        echo "Notebook testing completed"
+      shell: bash
+    - name: Upload test results
+      uses: actions/upload-artifact@v4
+      if: always()
+      with:
+        name: test-results-ubuntu-latest-3.11
+        path: |
+          test-results-ubuntu-latest-3.11.xml
+          notebook-test-results-ubuntu-latest-3.11.xml
+
+  test-matrix:
+    needs: test-primary
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
         os: [ubuntu-latest, macos-latest, windows-latest]
         python-version: ['3.10', '3.11']
+        exclude:
+          - os: ubuntu-latest
+            python-version: '3.11'
       fail-fast: false
     steps:
     - uses: actions/checkout@v4
@@ -51,7 +93,7 @@ jobs:
           notebook-test-results-${{ matrix.os }}-${{ matrix.python-version }}.xml
 
   test-summary:
-    needs: test
+    needs: [test-primary, test-matrix]
     runs-on: ubuntu-latest
     if: always()
     steps:
