### 2025-12-05 abuzarmahmood/pytau#173 @abuzarmahmood
Automate pypi release build
Package and upload automatically with a new release

https://www.paigeniedringhaus.com/blog/automatically-publish-a-repo-as-a-py-pi-library-with-git-hub-actions
