### 2025-12-05 bcae25e
chore(docs): update documentation build instructions
- Modified `README.md` to change the installation command for documentation dependencies.
- Added `requirements-docs.txt` to list specific packages required for building the documentation.
- This change promotes better dependency management and consistency in the documentation setup process.
