### 2025-12-05 c15eb93
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
