### 2025-12-05 604d888
ci: add workflow to test documentation build on specific file changes
Co-authored-by: aider (anthropic/claude-sonnet-4-20250514) <aider@aider.chat>
