### 2025-12-05 af8c84a
fix(ci): remove strict mode from mkdocs build in workflow
- Removed the `--strict` flag from the `mkdocs build` command in the GitHub Actions workflow to prevent build failures due to strict mode checks.
