+++ .github/workflows/docs_test.yml
@@ -27,7 +27,7 @@ jobs:
         pip install -r requirements-docs.txt
     - name: Test documentation build
       run: |
-        mkdocs build --strict
+        mkdocs build
     - name: Upload built docs as artifact
       uses: actions/upload-artifact@v4
       with:
