### 2025-12-05 79de345
ci: add path filters to workflow to run only on Python/notebook changes
Co-authored-by: aider (anthropic/claude-sonnet-4-20250514) <aider@aider.chat>
