+++ .github/workflows/pytest_workflow.yml
@@ -3,8 +3,22 @@ name: Pytest
 on:
   push:
     branches: [ main ]
+    paths:
+      - '**.py'
+      - '**.ipynb'
+      - 'pyproject.toml'
+      - 'setup.py'
+      - 'requirements*.txt'
+      - '.github/workflows/pytest_workflow.yml'
   pull_request:
     branches: [ '**' ]
+    paths:
+      - '**.py'
+      - '**.ipynb'
+      - 'pyproject.toml'
+      - 'setup.py'
+      - 'requirements*.txt'
+      - '.github/workflows/pytest_workflow.yml'
 
 jobs:
   test-primary:
