### 2025-12-05 cc42a43
chore(ci): update pytest workflow trigger configuration
- Remove branch and path filters for push events to trigger the workflow more broadly.
- Simplify and focus pull request triggers by removing branch specification.
