+++ .github/workflows/pytest_workflow.yml
@@ -1,17 +1,7 @@
 name: Pytest
 
 on:
-  push:
-    branches: [ main ]
-    paths:
-      - '**.py'
-      - '**.ipynb'
-      - 'pyproject.toml'
-      - 'setup.py'
-      - 'requirements*.txt'
-      - '.github/workflows/pytest_workflow.yml'
   pull_request:
-    branches: [ '**' ]
     paths:
       - '**.py'
       - '**.ipynb'
