### 2025-12-05 7659802
chore(workflows): update documentation test trigger
- Remove push event triggers for specific documentation files.
- Continue triggering on pull requests that update documentation-related files.
