### 2025-12-05 a182b41
Remove half-baked pipeline architecture page for now
