### 2025-12-05 abuzarmahmood/pytau#175 @abuzarmahmood
Update JOSS manuscript
https://github.com/openjournals/joss-reviews/issues/8509#issuecomment-3530323485

> While reviewing the paper, I noticed that the text is considerably longer than what JOSS allows, and several sections could be made clearer and more concise. Since the manuscript needs to be reduced to 1000 words or fewer, I’m outlining below the kind of content that each section should ideally contain so the structure remains solid while keeping only what is necessary in the paper.
>
> Summary:
> A short and accessible description of what the software does and its overall purpose, written so that non-specialists can understand it.
>
> Statement of Need:
> A concise explanation of the problem the software addresses, who the main users are, and why the tool is useful. Any prior motivation or related work can be mentioned briefly here.
>
> State of the Field:
> A brief comparison with similar tools or approaches, highlighting how this software differs from or complements existing solutions.
>
> Quality of Writing:
> The text should remain clear and to the point. Longer explanations, technical details, or extended examples should be moved to the documentation rather than included in the manuscript.
>
> References:
> All citations should be complete, properly formatted, and correspond to all works mentioned in the text.
>
> Everything beyond this essential content - full examples, detailed workflows, extended comparisons, or long formulas -should be placed in the documentation and linked from the paper when needed. Including one or two figures in the manuscript is perfectly fine, but if the number exceeds that, I would recommend placing the additional figures in the documentation under the examples section. This keeps the paper within JOSS limits while still giving readers access to the full technical details.
>
> These are my initial observations regarding the paper; as mentioned earlier, the content is generally solid and covers the necessary elements. The main remaining task is to organize the sections more concisely and bring the manuscript down to 1000 words or fewer. With those adjustments, the paper should align well with JOSS requirements. I will continue reviewing as needed.
>
> Here you can review the JOSS paper guide for more details:
> https://joss.readthedocs.io/en/latest/paper.html
