### 2025-12-06 abuzarmahmood/pytau#167 @HectorMozo3110
> [@HectorMozo3110](https://github.com/HectorMozo3110) [@duodenum96](https://github.com/duodenum96) Thank you for the feedback. I certainly agree that the applicability of the framework is very general. Even the naming (related to taste) comes from my own work in taste neuroscience but can be changed to reflect this general applicability. For example, `single_taste` can become `single_stimulus` and `all_taste` can become `hierarchical_mutli_stimulus`. I can definitely include this as part of the discussion of functionality and include examples where the model CAN be applied, but building "real-world" examples might take longer. I was thinking to demonstrate generalizability, I can apply the models in pytau to some of the exmaples shown here: https://www.lancaster.ac.uk/~romano/teaching/2425MATH337/1_intro_cusum.html#introduction-to-changepoints (all except the EEG data) Please let me know what y'all think. Thank you!

That approach sounds good, and it should be sufficient for now.
