### 2025-12-06 abuzarmahmood/pytau#167 @duodenum96
I agree with @HectorMozo3110, I think we can close this issue.
