### 2025-12-08 e8fc43b
Add mkdocs gh-deploy to docs workflow
Deploy documentation to GitHub Pages when changes are pushed to main.
Adds deploy-docs job that runs after test-docs passes.

Fixes #176

Co-authored-by: Ona <no-reply@ona.com>
