+++ .github/workflows/docs_test.yml
@@ -6,6 +6,13 @@ on:
       - 'mkdocs.yml'
       - 'docs/**'
       - 'requirements-docs.txt'
+  push:
+    branches:
+      - main
+    paths:
+      - 'mkdocs.yml'
+      - 'docs/**'
+      - 'requirements-docs.txt'
 
 jobs:
   test-docs:
@@ -29,3 +36,29 @@ jobs:
         name: built-docs
         path: site/
         retention-days: 7
+
+  deploy-docs:
+    runs-on: ubuntu-latest
+    if: github.event_name == 'push' && github.ref == 'refs/heads/main'
+    needs: test-docs
+    permissions:
+      contents: write
+    steps:
+    - uses: actions/checkout@v4
+      with:
+        fetch-depth: 0
+    - name: Set up Python 3.11
+      uses: actions/setup-python@v4
+      with:
+        python-version: '3.11'
+    - name: Install documentation dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install -r requirements-docs.txt
+    - name: Configure git
+      run: |
+        git config user.name "github-actions[bot]"
+        git config user.email "github-actions[bot]@users.noreply.github.com"
+    - name: Deploy documentation to GitHub Pages
+      run: |
+        mkdocs gh-deploy --force
