### 2025-12-08 bfc868d
Run mkdocs gh-deploy in PR workflow instead of push to master
Master is protected and never pushed to directly. Deploy docs during PR testing.

Co-authored-by: Ona <no-reply@ona.com>
