### 2025-12-08 37d61a1
Separate test and deploy: test on PR, deploy on merge to master
- test-docs job runs on PR open/sync/reopen
- deploy-docs job runs only when PR is merged to master
- Deployment happens after merge, not during PR

Co-authored-by: Ona <no-reply@ona.com>
