+++ .github/workflows/docs_test.yml
@@ -6,16 +6,14 @@ on:
       - 'mkdocs.yml'
       - 'docs/**'
       - 'requirements-docs.txt'
+    types: [opened, synchronize, reopened, closed]
 
 jobs:
-  test-and-deploy-docs:
+  test-docs:
     runs-on: ubuntu-latest
-    permissions:
-      contents: write
+    if: github.event.action != 'closed'
     steps:
     - uses: actions/checkout@v4
-      with:
-        fetch-depth: 0
     - name: Set up Python 3.11
       uses: actions/setup-python@v4
       with:
@@ -33,6 +31,24 @@ jobs:
         name: built-docs
         path: site/
         retention-days: 7
+
+  deploy-docs:
+    runs-on: ubuntu-latest
+    if: github.event.pull_request.merged == true
+    permissions:
+      contents: write
+    steps:
+    - uses: actions/checkout@v4
+      with:
+        fetch-depth: 0
+    - name: Set up Python 3.11
+      uses: actions/setup-python@v4
+      with:
+        python-version: '3.11'
+    - name: Install documentation dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install -r requirements-docs.txt
     - name: Configure git
       run: |
         git config user.name "github-actions[bot]"
