### 2025-12-10 abuzarmahmood/pytau#178 @duodenum96
Problems with installation script in Windows
I tried to run the quick start example. The bash script gave the following errors:

```
scripts/download_test_data.sh: line 1: pip: command not found
mkdir: cannot create directory ‘test_data\r’: File exists
scripts/download_test_data.sh: line 5: gdown: command not found
.ZIP.or test_data.zipr open test_data.zip
rm: cannot remove 'test_data.zip'$'\r': No such file or directory
```

That is probably due to the windows terminal though. I activated the conda environment in which the package is installed, using the following commands directly:

```
pip install gdown
mkdir test_data
cd test_data
gdown 1s1svxg4CvyUi9eeMzqUbcpiwCtNc170z -O test_data.zip
unzip test_data.zip
rm test_data.zip
```

And there were no problems. Might be useful to put a warning for windows users.

This issue is a part of the [review for JOSS](https://github.com/openjournals/joss-reviews/issues/8509).
