### 2025-12-15 abuzarmahmood/pytau#178 @abuzarmahmood
Thank you for catching this. I'll add a note for sure, but also see if I can find a more robust way to include test data. I believe I may be able to add it to the repo as a tracked file
