### 2025-12-15 2e271da
Add GitHub Actions workflow for automated PyPI releases
Automates package building and publishing to PyPI when a new release is created.
Uses trusted publishing (OIDC) for secure authentication without API tokens.

Closes #173

Co-authored-by: Ona <no-reply@ona.com>
