### 2025-12-15 e2ecc24
Add Windows installation workaround for test data download
Addresses issue where bash script fails on Windows due to line endings
and command availability. Updates both README and docs to guide Windows
users to manual download commands.

Fixes #178

Co-authored-by: Ona <no-reply@ona.com>
