### 2025-12-15 3d873ac
Update Python version to 3.10 in docs installation instructions
Aligns with pyproject.toml requirement (>=3.10,<3.12)

Co-authored-by: Ona <no-reply@ona.com>
