+++ docs/index.md
@@ -40,7 +40,7 @@ To set up PyTau, follow these steps:
 
 ```bash
 # Create and activate conda environment
-conda create -n "pytau_env" python=3.6.13 ipython notebook -y
+conda create -n "pytau_env" python=3.10 ipython notebook -y
 conda activate pytau_env
 
 # Clone repository
