### 2025-12-15 86b337f
Condense JOSS manuscript to meet word limit
Reduced manuscript from 1955 to 816 words to comply with JOSS requirements.
Moved detailed content to documentation (docs/examples.md) including:
- Extended code examples and tutorials
- Model types and features
- Statistical analysis tools
- Visualization examples
- Batch processing workflows

Updated mkdocs.yml to include new examples page.

Fixes #175

Co-authored-by: Ona <no-reply@ona.com>
