+++ JOSS_submission/paper.md
@@ -21,96 +21,50 @@ bibliography: paper.bib
 
 # Summary
 
-Analyzing complex biological data, particularly time-series data from neuroscience experiments, often requires sophisticated statistical modeling to identify significant changes in system dynamics. Several decades of research has emphasized that the dynamics of neural activity may show sharp changes accurately captured by models detecting state transitions such as Hidden aMarkov Models and changepoint models [@Seidemann1994; @Jones2007; @Saravani2019]. `pytau` is a Python software package designed to perform streamlined, batched inference for changepoint models across different parameter grids and datasets. It provides tools to efficiently query and analyze the results from sets of fitted models, facilitating the study of dynamic processes in biological systems, such as neural ensemble activity in response to stimuli. The package integrates with PyMC3 for Bayesian inference of these models (providing estimates of uncertainty in inference which are critical for noisy datasets usually with small sample sizes and low channel counts common in neuroscience) and provides utilities for data preprocessing, model fitting, and result visualization. The package has been successfully used in published research [@Mahmood2023; @Flores2023] and is currently being utilized in several ongoing studies [@MazzioInPrep; @BaasThomasInPrep; @MahmoodInPrep; @CaliaBoganInPrep].
+Neural activity often exhibits sharp transitions between distinct states, captured by changepoint models [@Seidemann1994; @Jones2007; @Saravani2019]. `pytau` is a Python package for batched Bayesian changepoint inference across parameter grids and datasets. It integrates with PyMC3 to provide uncertainty estimates critical for noisy neuroscience data with small sample sizes and low channel counts, and includes tools for preprocessing, model fitting, result visualization, and statistical analysis. The package has been successfully applied in published research examining taste processing [@Mahmood2023] and taste aversion learning [@Flores2023], and is currently being utilized in several ongoing neuroscience studies [@MazzioInPrep; @BaasThomasInPrep; @MahmoodInPrep; @CaliaBoganInPrep].
 
 # Statement of need
 
-Understanding how neural populations encode information often involves analyzing activity changes over time, potentially across different experimental conditions, parameters, or subjects. Fitting and comparing complex models like Bayesian changepoint models across numerous datasets or parameter settings can be computationally intensive and logistically challenging. There is a need for tools that streamline this process, enabling researchers to efficiently apply these models in batch, manage the results, and compare outcomes across conditions. `pytau` aims to fill this gap by providing a modularized pipeline specifically for fitting and analyzing changepoint models applied to neuroscience data, enabling efficient comparisons and analysis. This need is demonstrated by the tool's adoption in recent studies examining neural dynamics in taste processing [@Mahmood2023] and taste aversion learning [@Flores2023].
+Understanding how neural populations encode information often involves analyzing activity changes over time across different experimental conditions, parameters, or subjects. Fitting and comparing Bayesian changepoint models across numerous datasets or parameter settings is computationally intensive and logistically challenging. Existing changepoint detection tools lack the specialized functionality needed for neuroscience applications, particularly for handling multi-trial, multi-neuron spike train data.
 
-The package offers several key advantages:
+`pytau` addresses this gap by providing a modularized pipeline specifically designed for neuroscience data. The package offers several key advantages:
 
[…]
