+++ docs/examples.md
@@ -1,0 +1,335 @@
+# Examples and Tutorials
+
+This page provides detailed examples and tutorials for using PyTau to analyze neural data with changepoint models.
+
+## Complete Analysis Workflow
+
+Below is a comprehensive example demonstrating the full workflow from data loading to visualization:
+
+```python
+from pytau.changepoint_io import FitHandler
+from pytau.changepoint_analysis import PklHandler
+import matplotlib.pyplot as plt
+from pytau.utils.plotting import plot_changepoint_raster, plot_state_firing_rates
+
+# Initialize fit handler
+fh = FitHandler(
+    data_dir='/path/to/data',
+    taste_num=1,
+    region_name='GC',
+    experiment_name='example_experiment'
+)
+
+# Set preprocessing parameters
+fh.set_preprocess_params(
+    time_lims=[0, 2000],  # Time window in ms
+    bin_width=10,         # Bin width in ms
+    data_transform=None   # No transformation
+)
+
+# Set model parameters
+fh.set_model_params(
+    states=3,             # Number of states to fit
+    fit=5000,             # ADVI iterations
+    samples=1000,         # Number of posterior samples
+    model_kwargs={}       # Additional model parameters
+)
+
+# Run the full pipeline
+fh.load_spike_trains()
+fh.preprocess_data()
+fh.create_model()
+fh.run_inference()
+fh.save_fit_output()
+```
+
+## Analyzing Fitted Models
+
+After fitting, analyze the results using the `PklHandler` class:
+
+```python
+from pytau.changepoint_analysis import PklHandler
+
+# Load fitted model
+pkl_handler = PklHandler('/path/to/saved/model.pkl')
+
+# Access model components
+tau = pkl_handler.tau  # Changepoint times
+firing = pkl_handler.firing  # Firing rate analysis
+
+# Analyze significant neurons
+significant_neurons = firing.anova_significant_neurons
+
+# Access transition analysis data
+transition_snippets = firing.transition_snips
+pairwise_significant = firing.pairwise_significant_neurons
+```
+
+## Visualization Examples
+
+### Spike Rasters with Changepoint Overlays
+
+```python
+import matplotlib.pyplot as plt
+from pytau.utils.plotting import plot_changepoint_raster
+
+# Plot spike rasters with changepoint overlays
+fig, ax = plt.subplots(figsize=(10, 6))
+plot_changepoint_raster(pkl_handler.processed_spikes, 
+                        pkl_handler.tau.scaled_mode_tau,
+                        plot_lims=[0, 2000])
+plt.title('Spike Rasters with Changepoint Overlays')
+plt.show()
+```
+
+### State-Dependent Firing Rates
+
+```python
+from pytau.utils.plotting import plot_state_firing_rates
+
+# Plot state-dependent firing rates
+plot_state_firing_rates(pkl_handler.processed_spikes, 
+                        pkl_handler.tau.scaled_mode_tau)
+plt.title('State-Specific Neuron Activity')
+plt.show()
+```
+
+### Transition-Aligned Activity
+
+```python
+from pytau.utils.plotting import plot_aligned_state_firing
+
+# Plot transition-aligned activity
+plot_aligned_state_firing(pkl_handler.processed_spikes, 
+                          pkl_handler.tau.scaled_mode_tau, 
+                          window_radius=300)
+plt.title('Transition-Aligned Neural Activity')
+plt.show()
+```
+
+### State Timing Overview
+
+```python
+from pytau.utils.plotting import plot_changepoint_overview
+
+# Plot state timing overview
+plot_changepoint_overview(pkl_handler.tau.scaled_mode_tau, 
+                         plot_lims=[0, 2000])
+plt.title('State Timing Overview')
[…]
