### 2025-12-15 edb4819
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
