### 2025-12-15 c716f50
Update citations with correct DOIs and add Maigler et al
- Updated Mahmood 2023 to Mahmood 2025 with new bioRxiv preprint
- Updated Flores 2023 with correct title and full author name
- Updated Baas-Thomas from in-prep to published bioRxiv preprint
- Added Maigler et al 2024 citation

Co-authored-by: Ona <no-reply@ona.com>
