+++ JOSS_submission/paper.bib
@@ -32,27 +32,26 @@
 }
 
 @article{Mahmood2023,
-  author = {Mahmood, Abuzar and Steindler, Jessica and Germaine, Hannah and Miller, Paul and Katz, Donald B.},
-  title = {Coupled Dynamics of Stimulus-Evoked Gustatory Cortical and Basolateral Amygdalar Activity},
-  volume = {43},
-  number = {3},
-  pages = {386--404},
-  year = {2023},
-  doi = {10.1523/JNEUROSCI.1412-22.2022},
-  publisher = {Society for Neuroscience},
-  issn = {0270-6474},
-  URL = {https://www.jneurosci.org/content/43/3/386},
-  journal = {Journal of Neuroscience}
+  author = {Mahmood, Abuzar and Steindler, Jessica R. and Germaine, Hannah and Miller, Paul and Katz, Donald B.},
+  title = {Sensory and Palatability Coding of Taste Stimuli in Cortex Involves Dynamic and Asymmetric Cortico-Amygdalar Interactions},
+  elocation-id = {2025.07.01.662567},
+  year = {2025},
+  month = {Jul},
+  doi = {10.1101/2025.07.01.662567},
+  publisher = {Cold Spring Harbor Laboratory},
+  URL = {https://doi.org/10.1101/2025.07.01.662567},
+  journal = {bioRxiv}
 }
 
 @article{Flores2023,
-  author = {Flores, Veronica Lee and Lin, Jian-You},
-  title = {Taste Experience Enhances Cortical Response Reliability during Latent Enhancement of Taste Aversion Learning},
+  author = {Flores, Veronica L. and Lin, Jian-You},
+  title = {Benign Taste Experience Enhances Cortical Response Reliability during Aversion Conditioning Towards a Novel Taste},
   elocation-id = {2023.12.19.572413},
   year = {2023},
+  month = {Dec},
   doi = {10.1101/2023.12.19.572413},
   publisher = {Cold Spring Harbor Laboratory},
-  URL = {https://www.biorxiv.org/content/10.1101/2023.12.19.572413},
+  URL = {https://doi.org/10.1101/2023.12.19.572413},
   journal = {bioRxiv}
 }
 
@@ -70,11 +69,16 @@
   note = {In preparation}
 }
 
-@unpublished{BaasThomasInPrep,
+@article{BaasThomasInPrep,
   author = {Baas-Thomas, Natasha and Mahmood, Abuzar and Katz, Donald B.},
-  title = {Investigating the Neural Signals Driving the Consummatory Response in Rats},
+  title = {The Ingestive Response Reflects Neural Dynamics in Gustatory Cortex},
+  elocation-id = {2025.10.01.679845},
   year = {2025},
-  note = {In preparation}
+  month = {Oct},
+  doi = {10.1101/2025.10.01.679845},
+  publisher = {Cold Spring Harbor Laboratory},
+  URL = {https://doi.org/10.1101/2025.10.01.679845},
+  journal = {bioRxiv}
 }
 
 @unpublished{MahmoodInPrep,
@@ -89,6 +93,18 @@
   title = {Taste-evoked intra-state dynamics in the Gustatory Cortex},
   year = {2025},
   note = {In preparation}
+}
+
+@article{Maigler2024,
+  author = {Maigler, Kathleen C. and Mahmood, Abuzar and Osman, Bradly T. and Katz, Donald B.},
+  title = {Individual Taste Preferences Predict Cortical Taste Dynamics but Are Modified by Experience},
+  elocation-id = {2024.04.02.587752},
+  year = {2024},
+  month = {Apr},
+  doi = {10.1101/2024.04.02.587752},
[…]
