### 2025-12-15 1b0e9a8
Separate Mahmood2023 and Mahmood2025 citations, remove MahmoodInPrep
- Restored Mahmood2023 as the original Journal of Neuroscience paper
- Created Mahmood2025 as the new bioRxiv preprint
- Removed MahmoodInPrep (now published as Mahmood2025)
- Updated paper.md to cite both Mahmood papers and Maigler2024
- Updated Statement of Need to include BaasThomasInPrep

Co-authored-by: Ona <no-reply@ona.com>
