+++ JOSS_submission/paper.bib
@@ -32,6 +32,20 @@
 }
 
 @article{Mahmood2023,
+  author = {Mahmood, Abuzar and Steindler, Jessica and Germaine, Hannah and Miller, Paul and Katz, Donald B.},
+  title = {Coupled Dynamics of Stimulus-Evoked Gustatory Cortical and Basolateral Amygdalar Activity},
+  volume = {43},
+  number = {3},
+  pages = {386--404},
+  year = {2023},
+  doi = {10.1523/JNEUROSCI.1412-22.2022},
+  publisher = {Society for Neuroscience},
+  issn = {0270-6474},
+  URL = {https://www.jneurosci.org/content/43/3/386},
+  journal = {Journal of Neuroscience}
+}
+
+@article{Mahmood2025,
   author = {Mahmood, Abuzar and Steindler, Jessica R. and Germaine, Hannah and Miller, Paul and Katz, Donald B.},
   title = {Sensory and Palatability Coding of Taste Stimuli in Cortex Involves Dynamic and Asymmetric Cortico-Amygdalar Interactions},
   elocation-id = {2025.07.01.662567},
@@ -81,12 +95,7 @@
   journal = {bioRxiv}
 }
 
-@unpublished{MahmoodInPrep,
-  author = {Mahmood, Abuzar and Steindler, Jessica R. and Katz, Donald B.},
-  title = {Basolateral Amygdala and Gustatory Cortex interact asymmetrically during taste processing in rodents},
-  year = {2025},
-  note = {In preparation}
-}
+
 
 @unpublished{CaliaBoganInPrep,
   author = {Calia-Bogan, Vincent and Mahmood, Abuzar and Steindler, Jessica R. and Katz, Donald B.},
