### 2025-12-15 460b7e1
Update Mahmood2025 citation with correct author list and URL
- Changed to abbreviated author names (A., J.R., D.B.)
- Updated to lowercase title per bioRxiv format
- Changed URL to bioRxiv content URL
- Added eprint field for full PDF

Co-authored-by: Ona <no-reply@ona.com>
