+++ JOSS_submission/paper.bib
@@ -46,14 +46,14 @@
 }
 
 @article{Mahmood2025,
-  author = {Mahmood, Abuzar and Steindler, Jessica R. and Germaine, Hannah and Miller, Paul and Katz, Donald B.},
-  title = {Sensory and Palatability Coding of Taste Stimuli in Cortex Involves Dynamic and Asymmetric Cortico-Amygdalar Interactions},
+  author = {Mahmood, A. and Steindler, J.R. and Katz, D.B.},
+  title = {Sensory and palatability coding of taste stimuli in cortex involves dynamic and asymmetric cortico-amygdalar interactions},
   elocation-id = {2025.07.01.662567},
   year = {2025},
-  month = {Jul},
   doi = {10.1101/2025.07.01.662567},
   publisher = {Cold Spring Harbor Laboratory},
-  URL = {https://doi.org/10.1101/2025.07.01.662567},
+  URL = {https://www.biorxiv.org/content/early/2025/10/08/2025.07.01.662567},
+  eprint = {https://www.biorxiv.org/content/early/2025/10/08/2025.07.01.662567.full.pdf},
   journal = {bioRxiv}
 }
 
