### 2025-12-15 c93bb34
Add documentation references throughout the paper
Added links to detailed documentation sections:
- Statement of Need: Link to main documentation
- Implementation: Links to Available Models and Data Formats
- Implementation: Link to Inference Methods
- Example Usage: Link to Examples and Tutorials

Word count: 843 words (still under 1000 limit)

Co-authored-by: Ona <no-reply@ona.com>
