### 2025-12-15 4191e82
Add citations for bayesloop and PyChange
- Added bayesloop citation (Mark et al., Nature Communications 2018)
- Added PyChange citation (Gumbsch, GitHub repository)
- Updated State of the Field section to cite both tools

Word count: 845 words (still under 1000 limit)

Co-authored-by: Ona <no-reply@ona.com>
