+++ JOSS_submission/paper.bib
@@ -137,6 +137,26 @@
   doi = {10.48550/arXiv.1801.00826}
 }
 
+@article{bayesloop,
+  title = {Bayesian model selection for complex dynamic systems},
+  author = {Mark, Christoph and Metzner, Claus and Lautscham, Lena and Strissel, Pamela L. and Strick, Reiner and Fabry, Ben},
+  journal = {Nature Communications},
+  volume = {9},
+  number = {1},
+  pages = {1803},
+  year = {2018},
+  doi = {10.1038/s41467-018-04241-5},
+  url = {https://www.nature.com/articles/s41467-018-04241-5}
+}
+
+@misc{pychange,
+  title = {PyChange: Multiple change detection with Python},
+  author = {Gumbsch, Thomas},
+  year = {2017},
+  url = {https://github.com/BorgwardtLab/PyChange},
+  note = {GitHub repository}
+}
+
 @article{numpy,
   title = {Array programming with {NumPy}},
   author = {Charles R. Harris and K. Jarrod Millman and St{\'{e}}fan J. van der Walt and Ralf Gommers and Pauli Virtanen and David Cournapeau and Eric Wieser and Julian Taylor and Sebastian Berg and Nathaniel J. Smith and Robert Kern and Matti Picus and Stephan Hoyer and Marten H. van Kerkwijk and Matthew Brett and Allan Haldane and Jaime Fern{\'{a}}ndez del R{\'{i}}o and Mark Wiebe and Pearu Peterson and Pierre G{\'{e}}rard-Marchant and Kevin Sheppard and Tyler Reddy and Warren Weckesser and Hameer Abbasi and Christoph Gohlke and Travis E. Oliphant},
