+++ JOSS_submission/paper.md
@@ -102,7 +102,7 @@ The package includes visualization tools for examining neural activity with over
 
 # State of the field
 
-Several tools exist for changepoint detection, including `ruptures` [@ruptures] for offline change point detection, `bayesloop` for probabilistic time series analysis, `PyChange` for general time series changepoint detection, and Bayesian online changepoint detection methods [@adams2007bayesian; @fearnhead2007line]. While these general-purpose tools are valuable, `pytau` differs by focusing specifically on neuroscience applications. It provides specialized functionality for handling multi-trial, multi-neuron spike train data, batch processing across parameter grids, database management for model comparison, and neuroscience-specific visualization and statistical analysis tools. This specialization makes `pytau` particularly suited for researchers analyzing state transitions in neural population activity across different experimental paradigms.
+Several tools exist for changepoint detection, including `ruptures` [@ruptures] for offline change point detection, `bayesloop` [@bayesloop] for probabilistic time series analysis, `PyChange` [@pychange] for general time series changepoint detection, and Bayesian online changepoint detection methods [@adams2007bayesian; @fearnhead2007line]. While these general-purpose tools are valuable, `pytau` differs by focusing specifically on neuroscience applications. It provides specialized functionality for handling multi-trial, multi-neuron spike train data, batch processing across parameter grids, database management for model comparison, and neuroscience-specific visualization and statistical analysis tools. This specialization makes `pytau` particularly suited for researchers analyzing state transitions in neural population activity across different experimental paradigms.
 
 # Acknowledgements
 
