### 2025-12-15 d183537
Fix examples.md to match actual repository contents
Removed incorrect references:
- Removed DatabaseHandler.add_fit() and FitHandler.get_metadata() (not in API)
- Removed custom prior parameters that aren't exposed in model_kwargs
- Updated notebook list to match actual files in repo
- Updated script list to match actual files in repo
- Removed references to non-existent notebooks

All examples now reference actual functions and files in the repository.

Co-authored-by: Ona <no-reply@ona.com>
