### 2025-12-15 26c7d99
Remove redundant content from examples.md and add cross-references
Removed duplicate content that's already covered in other docs:
- Model types and features (now in models.md)
- Detailed statistical analysis descriptions (in API docs)
- Custom model parameters (not exposed in current API)
- Data transformation details (covered elsewhere)

Added cross-references to:
- Available Models documentation
- Inference Methods documentation
- API documentation

The examples.md now focuses on practical code examples and workflows,
with links to detailed documentation for each topic.

Co-authored-by: Ona
[…]
