### 2025-12-16 1b8255f
Update JOSS manuscript to highlight general applicability
- Add paragraph in Summary about broad applicability beyond neuroscience
- Mention classic changepoint examples (coal mining, temperature)
- Add GeneralChangepointExamples notebook to tutorials section
- Note applicability to economic, environmental, and quality control data

Addresses #174

Co-authored-by: Ona <no-reply@ona.com>
