### 2025-12-16 a0d984c
Remove update-readme workflow step and script
- Remove update-readme job from pytest_workflow.yml
- Delete update_readme_tests.py script
- Update execute-notebooks to depend on test-summary instead
- Fixes detached HEAD issue causing workflow failures

The update-readme job was causing failures due to detached HEAD state
when trying to push commits. Test results can be viewed in the
workflow artifacts instead.

Co-authored-by: Ona <no-reply@ona.com>
