+++ .github/workflows/pytest_workflow.yml
@@ -127,44 +127,8 @@ jobs:
           test_summary.md
           test_results_table.md
 
-  update-readme:
-    needs: test-summary
-    runs-on: ubuntu-latest
-    if: always()
-    steps:
-    - uses: actions/checkout@v4
-    - name: Download test summary
-      uses: actions/download-artifact@v4
-      with:
-        name: test-summary
-        path: .
-    - name: Set up Python 3.11
-      uses: actions/setup-python@v4
-      with:
-        python-version: '3.11'
-    - name: Update README with test results
-      run: |
-        python .github/scripts/update_readme_tests.py
-    - name: Commit updated README
-      run: |
-        git config --local user.email "action@github.com"
-        git config --local user.name "GitHub Action"
-        git add README.md
-        if git diff --staged --quiet; then
-          echo "No changes to README - test results already up to date"
-        else
-          echo "Committing updated README with latest test results"
-          git commit -m "Auto-update README with latest test results
-
-          - Updated test results table with latest CI run
-          - Generated by GitHub Actions workflow
-
-          [skip ci]"
-          git push
-        fi
-
   execute-notebooks:
-    needs: update-readme
+    needs: test-summary
     runs-on: ubuntu-latest
     if: github.event_name == 'push' && github.ref == 'refs/heads/main'
     steps:
