### 2025-12-16 abdac4c
Move execute-notebooks to independent workflow
- Create new .github/workflows/execute_notebooks.yml
- Triggers on PR merge to master (similar to deploy-docs)
- Remove execute-notebooks job from pytest_workflow.yml
- Uses github-actions[bot] for commits (consistent with deploy-docs)
- Runs independently without blocking test workflow

This separates notebook execution from testing, allowing:
- Faster test feedback (no waiting for notebook execution)
- Cleaner workflow organization
- Notebook execution only on actual merges to master

Co-authored-by: Ona <no-reply@ona.com>
