+++ .github/workflows/pytest_workflow.yml
@@ -126,54 +126,3 @@ jobs:
         path: |
           test_summary.md
           test_results_table.md
-
-  execute-notebooks:
-    needs: test-summary
-    runs-on: ubuntu-latest
-    if: github.event_name == 'push' && github.ref == 'refs/heads/main'
-    steps:
-    - uses: actions/checkout@v4
-      with:
-        token: ${{ secrets.GITHUB_TOKEN }}
-        fetch-depth: 0
-    - name: Set up Python 3.11
-      uses: actions/setup-python@v4
-      with:
-        python-version: '3.11'
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        pip install -e .[dev]
-    - name: Execute notebooks and save outputs
-      run: |
-        echo "Executing model notebooks with outputs..."
-        for notebook in pytau/how_to/model_notebooks/*.ipynb; do
-          echo "Executing $(basename "$notebook")..."
-          jupyter nbconvert --to notebook --execute "$notebook" --inplace --ExecutePreprocessor.timeout=300
-        done
-        echo "All notebooks executed successfully!"
-    - name: Upload executed notebooks as artifacts
-      uses: actions/upload-artifact@v4
-      with:
-        name: executed-notebooks-${{ github.sha }}
-        path: pytau/how_to/model_notebooks/*.ipynb
-        retention-days: 30
-    - name: Commit executed notebooks (main branch only)
-      if: github.event_name == 'push' && github.ref == 'refs/heads/main'
-      run: |
-        git config --local user.email "action@github.com"
-        git config --local user.name "GitHub Action"
-        git add pytau/how_to/model_notebooks/*.ipynb
-        if git diff --staged --quiet; then
-          echo "No changes to commit - notebooks already up to date"
-        else
-          echo "Committing updated notebooks with execution outputs"
-          git commit -m "Auto-update model notebooks with execution outputs
-
-          - Executed all model notebooks with latest code
-          - Preserved outputs for documentation and review
-          - Generated by GitHub Actions workflow
-
-          [skip ci]"
-          git push
-        fi
