+++ .github/workflows/execute_notebooks.yml
@@ -1,0 +1,64 @@
+name: Execute Notebooks
+
+on:
+  pull_request:
+    paths:
+      - '**.ipynb'
+      - 'pytau/**/*.py'
+      - 'pyproject.toml'
+      - 'requirements*.txt'
+    types: [closed]
+
+jobs:
+  execute-notebooks:
+    runs-on: ubuntu-latest
+    if: github.event.pull_request.merged == true && github.event.pull_request.base.ref == 'master'
+    permissions:
+      contents: write
+    steps:
+    - uses: actions/checkout@v4
+      with:
+        ref: master
+        fetch-depth: 0
+    - name: Set up Python 3.11
+      uses: actions/setup-python@v4
+      with:
+        python-version: '3.11'
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install -e .[dev]
+    - name: Execute notebooks and save outputs
+      run: |
+        echo "Executing model notebooks with outputs..."
+        for notebook in pytau/how_to/model_notebooks/*.ipynb; do
+          echo "Executing $(basename "$notebook")..."
+          jupyter nbconvert --to notebook --execute "$notebook" --inplace --ExecutePreprocessor.timeout=300
+        done
+        echo "All notebooks executed successfully!"
+    - name: Upload executed notebooks as artifacts
+      uses: actions/upload-artifact@v4
+      with:
+        name: executed-notebooks-${{ github.sha }}
+        path: pytau/how_to/model_notebooks/*.ipynb
+        retention-days: 30
+    - name: Configure git
+      run: |
+        git config user.name "github-actions[bot]"
+        git config user.email "github-actions[bot]@users.noreply.github.com"
+    - name: Commit executed notebooks
+      run: |
+        git add pytau/how_to/model_notebooks/*.ipynb
+        if git diff --staged --quiet; then
+          echo "No changes to commit - notebooks already up to date"
+        else
+          echo "Committing updated notebooks with execution outputs"
+          git commit -m "Auto-update model notebooks with execution outputs
+
+          - Executed all model notebooks with latest code
+          - Preserved outputs for documentation and review
+          - Generated by GitHub Actions workflow
+
+          [skip ci]"
+          git push
+        fi
