### 2025-12-16 342f9d4
Fix execute-notebooks workflow to only run on push to master
- Change trigger from pull_request[closed] to push[master]
- Remove conditional check (no longer needed)
- Remove explicit ref checkout (already on master)
- Prevents accidental execution on PR updates

This matches the pattern used by deploy-docs workflow and ensures
notebooks are only executed when changes are actually merged to master.

Co-authored-by: Ona <no-reply@ona.com>
