### 2025-12-17 12a3e8b
Minor manuscript adjustments (JOSS review)
