### 2025-12-17 #8509 @HectorMozo3110
> [@fboehm](https://github.com/fboehm) , Thank you for checking in. I have addressed most of the comments from [@HectorMozo3110](https://github.com/HectorMozo3110) and [@duodenum96](https://github.com/duodenum96) from [issue # 167](https://github.com/abuzarmahmood/pytau/issues/167) regarding improvements to the README.
>
> I have also updated the manuscript per [@HectorMozo3110](https://github.com/HectorMozo3110)'s [comment](https://github.com/openjournals/joss-reviews/issues/8509#issuecomment-3530323485), making the manuscript < 1000 words and moving lengthier technical discussion to the docs with references in the paper to the appropriate sections in the docs, and updated the documentation to address an [issue ](https://github.com/abuzarmahmood/pytau/issues/178)raised by [@duodenum96](https://github.com/duodenum96) regarding downloading test_data.
>
> A partially addressed issue was highlighting the general applicability of the package to changepoint detection beyond neural data. Pytau [PR 184](https://github.com/abuzarmahmood/pytau/pull/184) attempts this by performing changepoint inference on simulated data but I have left the issue open for now for future work.
>
> I am (as of now) unable to update the package on pypi because I am recovering access to my account but this is also currently noted as an [issue](https://github.com/abuzarmahmood/pytau/issues/173).
>
> Other than the pypi package, I believe I have incorporated and addressed all other suggestions and issues. Please let me know if I have missed anything.
>
> Thank you again everyone for taking the time to review this submission and provide feedback!!

@abuzarmahmood I’ve tested the project locally, including the examples, core code, and test suite, and everything looks good to me overall.

From a packaging perspective, the `pyproject.toml` also looks solid. I only have a few minor observations that might help improve consistency and ease of use. In particular, there is a small mismatch between the declared Python version support (`requires-python = ">=3.10,<3.12"`) and the Python 3.12 classifier, so it might be worth aligning those. Adding a `[project.urls]` section (e.g., Homepage, Documentation, Repository, Issues) would also help point users to the relevant resources.

I see that dependencies are already well organized through separate `requirements.txt`, `requirements-dev.txt`, and `requirements-docs.txt` files, which works well. As a complementary option, it could be helpful to also expose some of these groups as optional extras in `pyproject.toml` (e.g., `test`, `docs`, `dev`, or a `gui` extra for `easygui`). This would give users a more flexible installation path via `pip install pytau[extra]`, while keeping the existing workflow intact. Similarly, using automatic package discovery (e.g., including all `pytau*` subpackages) would help ensure everything is consistently packaged.

Along similar lines, `easygui` can be useful on systems where GUI interaction is available, but it may require system-level `tkinter`, which is not always present in headless or minimal environments. Making this dependency optional (for example via a `gui` extra) could help avoid installation issues in those cases, while still allowing users who need GUI functionality to enable it explicitly.

Aside from these small points, the project looks well structured and in good shape.
