### 2025-12-17 #8509 @HectorMozo3110
>I am (as of now) unable to update the package on pypi because I am recovering access to my account but this is also >currently noted as an https://github.com/abuzarmahmood/pytau/issues/173.

>Other than the pypi package, I believe I have incorporated and addressed all other suggestions and issues. Please let me >know if I have missed anything.

Thanks for the update. That’s fine for the moment , from my side everything looks good, and aside from the PyPI update (which is already noted), I don’t see any remaining issues. In my view, this should not affect the submission at this stage. The previous comment covers my remaining observations.
