### 2025-12-17 abuzarmahmood/pytau#186 @duodenum96
ELBO values in CategoricalChangepoint2D.ipynb in the tutorial notebooks end up in infinity
The ELBO values in CategoricalChangepoint2D.ipynb in the jupyter notebook tutorials are ending up in infinity:

This is a part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/8509).
