### 2025-12-17 abuzarmahmood/pytau#187 @abuzarmahmood
pyproject.toml improvements (as part of JOSS review)
- Fix python version classification
- Add [project.urls] for Homepage, Documentation, Repository, Issues
- Add dynamic requirements to pyproject.toml
- Separate out easygui as it may require tkinter (which may not be available on all systems)

https://github.com/openjournals/joss-reviews/issues/8509#issuecomment-3663389800
>From a packaging perspective, the pyproject.toml also looks solid. I only have a few minor observations that might help improve consistency and ease of use. In particular, there is a small mismatch between the declared Python version support (requires-python = ">=3.10,<3.12") and the Python 3.12 classifier, so it might be worth aligning those. Adding a [project.urls] section (e.g., Homepage, Documentation, Repository, Issues) would also help point users to the relevant resources.

>I see that dependencies are already well organized through separate requirements.txt, requirements-dev.txt, and requirements-docs.txt files, which works well. As a complementary option, it could be helpful to also expose some of these groups as optional extras in pyproject.toml (e.g., test, docs, dev, or a gui extra for easygui). This would give users a more flexible installation path via pip install pytau[extra], while keeping the existing workflow intact. Similarly, using automatic package discovery (e.g., including all pytau* subpackages) would help ensure everything is consistently packaged.

>Along similar lines, easygui can be useful on systems where GUI interaction is available, but it may require system-level tkinter, which is not always present in headless or minimal environments. Making this dependency optional (for example via a gui extra) could help avoid installation issues in those cases, while still allowing users who need GUI functionality to enable it explicitly.
