### 2025-12-17 3b6e513
chore(dependencies): update Python version consistency in configuration
- Remove Python 3.12 support from `pyproject.toml` classifiers to maintain consistent version support.
- Update `tox.ini` to utilize Python 3.11, aligning the environment configurations with current support standards.
