### 2025-12-17 e69a885
refactor(dependencies): move dependencies to separate requirements files
- Updated `pyproject.toml` to use dynamic dependencies loading from external files.
- Renamed and reorganized requirements files into the `requirements` directory.
- Separated dependencies into distinct files for `dev`, `docs`, and `gui` optional dependencies.
- Removed hardcoded dependencies from `pyproject.toml` and `requirements.txt`.
- Added a new `requirements-gui.txt` file for GUI-specific dependencies.
