### 2025-12-17 4d2a370
ci: add installation test workflow for package dependencies
Co-authored-by: aider (anthropic/claude-sonnet-4-20250514) <aider@aider.chat>
