+++ .github/workflows/install_test.yml
@@ -1,0 +1,56 @@
+name: Installation Test
+
+on:
+  pull_request:
+    paths:
+      - 'requirements/**'
+      - 'pyproject.toml'
+      - 'setup.py'
+      - 'setup.cfg'
+  push:
+    branches:
+      - main
+    paths:
+      - 'requirements/**'
+      - 'pyproject.toml'
+      - 'setup.py'
+      - 'setup.cfg'
+
+jobs:
+  test-install:
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: ['3.10', '3.11']
+    
+    steps:
+    - uses: actions/checkout@v4
+    
+    - name: Set up Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v4
+      with:
+        python-version: ${{ matrix.python-version }}
+    
+    - name: Install package
+      run: |
+        python -m pip install --upgrade pip
+        pip install -e .
+    
+    - name: Test basic import
+      run: |
+        python -c "import pytau; print('pytau imported successfully')"
+    
+    - name: Install dev dependencies and test
+      run: |
+        pip install -e .[dev]
+        python -c "import pytest; print('dev dependencies installed successfully')"
+    
+    - name: Install docs dependencies and test
+      run: |
+        pip install -e .[docs]
+        python -c "print('docs dependencies installed successfully')"
+    
+    - name: Install gui dependencies and test
+      run: |
+        pip install -e .[gui]
+        python -c "print('gui dependencies installed successfully')"
