### 2025-12-17 09e2f2c
ci(workflow): update install_test workflow configuration
- Removed the push event trigger from the workflow configuration to focus on specific paths for the pull request events only.
- Added a new step to install all dependencies and verify by printing a success message.
