### 2025-12-17 dffd34a
docs: update installation instructions and add optional dependencies
Co-authored-by: aider (anthropic/claude-sonnet-4-20250514) <aider@aider.chat>
