### 2025-12-17 1f5ac05
fix(workflow): correct path for requirements-docs.txt
- Updated the path to `requirements-docs.txt` to reflect its new location in the `requirements` directory.
