### 2025-12-17 9ed649f
ci: update GitHub Actions to latest versions and installation procedure
Co-authored-by: aider (anthropic/claude-sonnet-4-20250514) <aider@aider.chat>
