### 2025-12-17 abuzarmahmood/pytau#186 @abuzarmahmood
Thank you for pointing this out. This indicates some issue with the model.
I believe the problem would be subtle at best because we just used this model for a manuscript (https://www.biorxiv.org/content/10.1101/2025.10.01.679845v1.full) and it performed as expected.

For now I will add a warning printout after fitting to say that an `inf` indicates an issue with fitting and the user should try to track the model paramters as specified in (https://www.pymc.io/projects/examples/en/2022.01.0/variational_inference/variational_api_quickstart.html#tracking-parameters)

I will also add checks prior to fitting to make sure the data have no nans and infs.

However, I will keep this issue open to indicate that a more rigorous assessment of the model structure is needed.
