### 2025-12-17 5c3ea2c
ci(build): add GUI dependencies to pytest workflow
Update GitHub Actions workflow to include GUI dependencies to ensure `pytest` runs successfully. The addition of `gui` extra is necessary because `pytest` relies on `easygui` due to `ephysdata` dependencies.
