+++ .github/workflows/pytest_workflow.yml
@@ -22,7 +22,7 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install -e .[dev]
+        pip install -e '.[dev, gui]'
     - name: Run tests
       run: |
         pytest --cov=pytau -v --tb=short --junitxml=test-results-ubuntu-latest-3.11.xml
@@ -69,7 +69,7 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install -e .[dev]
+        pip install -e '.[dev, gui]'
     - name: Run tests
       run: |
         pytest --cov=pytau -v --tb=short --junitxml=test-results-${{ matrix.os }}-${{ matrix.python-version }}.xml
