### 2025-12-17 7543799
feat: add data quality check function to detect and warn about NaNs and infs in input data
Co-authored-by: aider (anthropic/claude-sonnet-4-20250514) <aider@aider.chat>
