+++ pytau/changepoint_model.py
@@ -42,6 +42,47 @@ class ChangepointModel:
 ############################################################
 
 
+def check_data_quality(data_array, array_name="data_array"):
+    """
+    Check input data for infs and nans that could cause fitting issues.
+    
+    Args:
+        data_array (numpy.ndarray): Input data array to check
+        array_name (str): Name of the array for error messages
+        
+    Returns:
+        bool: True if data is clean, False if issues found
+    """
+    data_array = np.asarray(data_array)
+    
+    has_nan = np.any(np.isnan(data_array))
+    has_inf = np.any(np.isinf(data_array))
+    
+    if has_nan or has_inf:
+        print("=" * 60)
+        print("⚠️  WARNING: DATA QUALITY ISSUES DETECTED ⚠️")
+        print("=" * 60)
+        
+        if has_nan:
+            nan_count = np.sum(np.isnan(data_array))
+            print(f"❌ Found {nan_count} NaN values in {array_name}")
+            
+        if has_inf:
+            inf_count = np.sum(np.isinf(data_array))
+            print(f"❌ Found {inf_count} infinite values in {array_name}")
+            
+        print("\n🚨 MODEL FITTING MAY FAIL OR PRODUCE UNRELIABLE RESULTS!")
+        print("\nRecommended actions:")
+        print("  • Remove or interpolate NaN values")
+        print("  • Replace infinite values with finite numbers")
+        print("  • Check data preprocessing pipeline")
+        print("=" * 60)
+        
+        return False
+    
+    return True
+
+
 def gen_test_array(array_size, n_states, type="poisson"):
     """
     Generate test array for model fitting
@@ -144,6 +185,7 @@ class GaussianChangepointMeanVar2D(ChangepointModel):
             **kwargs: Additional arguments
         """
         super().__init__(**kwargs)
+        check_data_quality(data_array, "data_array")
         self.data_array = data_array
         self.n_states = n_states
 
@@ -249,6 +291,7 @@ class GaussianChangepointMeanDirichlet(ChangepointModel):
             **kwargs: Additional arguments
         """
         super().__init__(**kwargs)
+        check_data_quality(data_array, "data_array")
         self.data_array = data_array
         self.max_states = max_states
 
@@ -373,6 +416,7 @@ class GaussianChangepointMean2D(ChangepointModel):
             **kwargs: Additional arguments
         """
         super().__init__(**kwargs)
+        check_data_quality(data_array, "data_array")
         self.data_array = data_array
         self.n_states = n_states
 
@@ -484,6 +528,7 @@ class SingleTastePoissonDirichlet(ChangepointModel):
             **kwargs: Additional arguments
         """
         super().__init__(**kwargs)
+        check_data_quality(data_array, "data_array")
         self.data_array = data_array
         self.max_states = max_states
 
@@ -613,6 +658,7 @@ class SingleTastePoisson(ChangepointModel):
             **kwargs: Additional arguments
         """
         super().__init__(**kwargs)
+        check_data_quality(data_array, "data_array")
         self.data_array = data_array
         self.n_states = n_states
 
@@ -734,6 +780,7 @@ class SingleTastePoissonVarsig(ChangepointModel):
             **kwargs: Additional arguments
         """
         super().__init__(**kwargs)
+        check_data_quality(data_array, "data_array")
         self.data_array = data_array
         self.n_states = n_states
 
@@ -888,6 +935,7 @@ class SingleTastePoissonVarsigFixed(ChangepointModel):
             **kwargs: Additional arguments
         """
         super().__init__(**kwargs)
+        check_data_quality(data_array, "data_array")
         self.data_array = data_array
         self.n_states = n_states
         self.inds_span = inds_span
[…]
