### 2025-12-17 31bb20b
feat: add warning for inf/nan values in ADVI ELBO history
Co-authored-by: aider (anthropic/claude-sonnet-4-20250514) <aider@aider.chat>
