### 2025-12-17 1b3a87e
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
