### 2025-12-17 3f10163
build(dependencies): add 'tables' to project dependencies
- Added 'tables' to the requirements.txt file to include it as a project dependency.
