### 2025-12-17 6650e0a
docs: add known ELBO infinity issue for CategoricalChangepoint2D model
Co-authored-by: aider (anthropic/claude-sonnet-4-20250514) <aider@aider.chat>
