### 2025-12-18 #8509 @HectorMozo3110
Hi 👋 @abuzarmahmood , thanks for the updates, everything looks good from my side.

Regarding *automatic package discovery (`pytau*`)*, I was referring to using setuptools’ built-in package discovery so that any `pytau.*` subpackages are included automatically, rather than listing only `packages = ["pytau"]`. This is optional and mainly a suggestion to improve robustness as the project evolves. Reference: https://setuptools.pypa.io/en/latest/userguide/package_discovery.html

With that said, the submission is in good shape, and I recommend it for acceptance.

@fboehm From my side, this is ready to proceed.
