### 2025-12-18 abuzarmahmood/pytau#190 @abuzarmahmood
Update setup for automatic package discovery 
https://github.com/openjournals/joss-reviews/issues/8509#issuecomment-3667915280

> Regarding automatic package discovery (pytau*), I was referring to using setuptools’ built-in package discovery so that any pytau.* subpackages are included automatically, rather than listing only packages = ["pytau"]. This is optional and mainly a suggestion to improve robustness as the project evolves. Reference: https://setuptools.pypa.io/en/latest/userguide/package_discovery.html
