### 2025-12-19 #8509 @fboehm
@abuzarmahmood - I think it's fine if they don't yet have dois; the way you've listed them is fine.
