### 2025-12-23 3a66427
fix(bibliography): correct preparation status in citations
- Updated titles in `paper.bib` to reflect the correct "In Preparation" status.
- Adjusted references in `paper.md` to match the updated citations in the bibliography.
