+++ JOSS_submission/paper.bib
@@ -71,14 +71,14 @@
 
 @unpublished{MartinInPrep,
   author = {Raymond, Martin and Lin, Jian-You and Katz, Donald B.},
-  title = {A Comparison of Cortical Taste Representations Arising from Free Licking and Intraoral Delivery},
+  title = {A Comparison of Cortical Taste Representations Arising from Free Licking and Intraoral Delivery (In Preparataion)},
   year = {2025},
   note = {In preparation}
 }
 
 @unpublished{MazzioInPrep,
   author = {Mazzio, Christina and Germaine, Hannah and Lin, Jian-You and Katz, Donald B.},
-  title = {Cortical Dynamics Underlying Learned and Non-Learned Aversive Behavior},
+  title = {Cortical Dynamics Underlying Learned and Non-Learned Aversive Behavior (In Preparation)},
   year = {2025},
   note = {In preparation}
 }
@@ -99,7 +99,7 @@
 
 @unpublished{CaliaBoganInPrep,
   author = {Calia-Bogan, Vincent and Mahmood, Abuzar and Steindler, Jessica R. and Katz, Donald B.},
-  title = {Taste-evoked intra-state dynamics in the Gustatory Cortex},
+  title = {Taste-evoked intra-state dynamics in the Gustatory Cortex (In Preparation)},
   year = {2025},
   note = {In preparation}
 }
