### 2026-01-06 a730a04
Add get_state_snippets function to extract raw state activity
Implements get_state_snippets to return neural activity for each state
without averaging, enabling downstream analysis on raw state data.
Refactors get_state_firing to use get_state_snippets internally.

Resolves #196

Co-authored-by: Ona <no-reply@ona.com>
