### 2026-01-06 aaf7450
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
