### 2026-01-09 22e9405
Fixes for JOSS submission
https://github.com/openjournals/joss-reviews/issues/8509#issuecomment-3720758673
