### 2026-01-09 0023f84
Add references for SciPy and Matplotlib articles
