### 2026-01-09 9a61499
Apply proper title case capitalization to all bibliography entries in paper.bib
Co-authored-by: abuzarmahmood <12436309+abuzarmahmood@users.noreply.github.com>
