+++ JOSS_submission/paper.bib
@@ -1,5 +1,5 @@
 @article{Saravani2019,
-  title={Dynamic Brain Interactions during picture naming},
+  title={Dynamic Brain Interactions During Picture Naming},
   volume={6},
   DOI={10.1523/eneuro.0472-18.2019},
   number={4},
@@ -9,7 +9,7 @@
   month={Jun}
 }
 @article{Jones2007,
-  title={Natural stimuli evoke dynamic sequences of states in sensory cortical ensembles},
+  title={Natural Stimuli Evoke Dynamic Sequences of States in Sensory Cortical Ensembles},
   volume={104},
   DOI={10.1073/pnas.0705546104},
   number={47},
@@ -20,7 +20,7 @@
   pages={18772–18777}
 }
 @article{Seidemann1994,
-  title={Simultaneously recorded single units in the frontal cortex go through sequences of discrete and stable states in monkeys performing a delayed localization task},
+  title={Simultaneously Recorded Single Units in the Frontal Cortex Go Through Sequences of Discrete and Stable States in Monkeys Performing a Delayed Localization Task},
   volume={16},
   DOI={10.1523/jneurosci.16-02-00752.1996},
   number={2},
@@ -47,7 +47,7 @@
 
 @article{Mahmood2025,
   author = {Mahmood, A. and Steindler, J.R. and Katz, D.B.},
-  title = {Sensory and palatability coding of taste stimuli in cortex involves dynamic and asymmetric cortico-amygdalar interactions},
+  title = {Sensory and Palatability Coding of Taste Stimuli in Cortex Involves Dynamic and Asymmetric Cortico-Amygdalar Interactions},
   elocation-id = {2025.07.01.662567},
   year = {2025},
   doi = {10.1101/2025.07.01.662567},
@@ -59,7 +59,7 @@
 
 @article{Flores2023,
   author = {Flores, Veronica L. and Lin, Jian-You},
-  title = {Benign Taste Experience Enhances Cortical Response Reliability during Aversion Conditioning Towards a Novel Taste},
+  title = {Benign Taste Experience Enhances Cortical Response Reliability During Aversion Conditioning Towards a Novel Taste},
   elocation-id = {2023.12.19.572413},
   year = {2023},
   month = {Dec},
@@ -99,14 +99,14 @@
 
 @unpublished{CaliaBoganInPrep,
   author = {Calia-Bogan, Vincent and Mahmood, Abuzar and Steindler, Jessica R. and Katz, Donald B.},
-  title = {Taste-evoked intra-state dynamics in the Gustatory Cortex (In Preparation)},
+  title = {Taste-Evoked Intra-State Dynamics in the Gustatory Cortex (In Preparation)},
   year = {2025},
   note = {In preparation}
 }
 
 @article{Maigler2024,
   author = {Maigler, Kathleen C. and Mahmood, Abuzar and Osman, Bradly T. and Katz, Donald B.},
-  title = {Individual Taste Preferences Predict Cortical Taste Dynamics but Are Modified by Experience},
+  title = {Individual Taste Preferences Predict Cortical Taste Dynamics But Are Modified by Experience},
   elocation-id = {2024.04.02.587752},
   year = {2024},
   month = {Apr},
@@ -118,7 +118,7 @@
 
 @article{pymc3,
   author = {Salvatier, John and Wiecki, Thomas V and Fonnesbeck, Christopher},
-  title = {Probabilistic programming in Python using PyMC3},
+  title = {Probabilistic Programming in Python Using PyMC3},
   journal = {PeerJ Computer Science},
   volume = {2},
   pages = {e55},
@@ -129,7 +129,7 @@
 }
 
 @article{ruptures,
[…]
