+++ JOSS_submission/paper.bib
@@ -59,7 +59,7 @@
 
 @article{Flores2023,
   author = {Flores, Veronica L. and Lin, Jian-You},
-  title = {Benign Taste Experience Enhances Cortical Response Reliability During Aversion Conditioning Towards a Novel Taste},
+  title = {Benign Taste Experience Enhances Cortical Response Reliability During Aversion Conditioning towards a Novel Taste},
   elocation-id = {2023.12.19.572413},
   year = {2023},
   month = {Dec},
@@ -106,7 +106,7 @@
 
 @article{Maigler2024,
   author = {Maigler, Kathleen C. and Mahmood, Abuzar and Osman, Bradly T. and Katz, Donald B.},
-  title = {Individual Taste Preferences Predict Cortical Taste Dynamics But Are Modified by Experience},
+  title = {Individual Taste Preferences Predict Cortical Taste Dynamics but Are Modified by Experience},
   elocation-id = {2024.04.02.587752},
   year = {2024},
   month = {Apr},
@@ -150,7 +150,7 @@
 }
 
 @misc{pychange,
-  title = {PyChange: Multiple Change Detection With Python},
+  title = {PyChange: Multiple Change Detection with Python},
   author = {Gumbsch, Thomas},
   year = {2017},
   url = {https://github.com/BorgwardtLab/PyChange},
@@ -158,7 +158,7 @@
 }
 
 @article{numpy,
-  title = {Array Programming With {NumPy}},
+  title = {Array Programming with {NumPy}},
   author = {Charles R. Harris and K. Jarrod Millman and St{\'{e}}fan J. van der Walt and Ralf Gommers and Pauli Virtanen and David Cournapeau and Eric Wieser and Julian Taylor and Sebastian Berg and Nathaniel J. Smith and Robert Kern and Matti Picus and Stephan Hoyer and Marten H. van Kerkwijk and Matthew Brett and Allan Haldane and Jaime Fern{\'{a}}ndez del R{\'{i}}o and Mark Wiebe and Pearu Peterson and Pierre G{\'{e}}rard-Marchant and Kevin Sheppard and Tyler Reddy and Warren Weckesser and Hameer Abbasi and Christoph Gohlke and Travis E. Oliphant},
   journal = {Nature},
   year = {2020},
