+++ JOSS_submission/paper.bib
@@ -59,7 +59,7 @@
 
 @article{Flores2023,
   author = {Flores, Veronica L. and Lin, Jian-You},
-  title = {Benign Taste Experience Enhances Cortical Response Reliability During Aversion Conditioning towards a Novel Taste},
+  title = {Benign Taste Experience Enhances Cortical Response Reliability During Aversion Conditioning Towards a Novel Taste},
   elocation-id = {2023.12.19.572413},
   year = {2023},
   month = {Dec},
@@ -71,7 +71,7 @@
 
 @unpublished{MartinInPrep,
   author = {Raymond, Martin and Lin, Jian-You and Katz, Donald B.},
-  title = {A Comparison of Cortical Taste Representations Arising from Free Licking and Intraoral Delivery (In Preparataion)},
+  title = {A Comparison of Cortical Taste Representations Arising from Free Licking and Intraoral Delivery (In Preparation)},
   year = {2025},
   note = {In preparation}
 }
