### 2026-01-09 7605ae4
Protect proper nouns with curly brackets to force capitalization
Co-authored-by: abuzarmahmood <12436309+abuzarmahmood@users.noreply.github.com>
