+++ JOSS_submission/paper.bib
@@ -118,7 +118,7 @@
 
 @article{pymc3,
   author = {Salvatier, John and Wiecki, Thomas V and Fonnesbeck, Christopher},
-  title = {Probabilistic Programming in Python Using PyMC3},
+  title = {Probabilistic Programming in {Python} Using {PyMC3}},
   journal = {PeerJ Computer Science},
   volume = {2},
   pages = {e55},
@@ -129,7 +129,7 @@
 }
 
 @article{ruptures,
-  title = {Ruptures: Change Point Detection in Python},
+  title = {{Ruptures}: Change Point Detection in {Python}},
   author = {Truong, Charles and Oudre, Laurent and Vayatis, Nicolas},
   journal = {arXiv preprint arXiv:1801.00826},
   year = {2018},
@@ -138,7 +138,7 @@
 }
 
 @article{bayesloop,
-  title = {Bayesian Model Selection for Complex Dynamic Systems},
+  title = {{Bayesian} Model Selection for Complex Dynamic Systems},
   author = {Mark, Christoph and Metzner, Claus and Lautscham, Lena and Strissel, Pamela L. and Strick, Reiner and Fabry, Ben},
   journal = {Nature Communications},
   volume = {9},
@@ -150,7 +150,7 @@
 }
 
 @misc{pychange,
-  title = {PyChange: Multiple Change Detection with Python},
+  title = {{PyChange}: Multiple Change Detection with {Python}},
   author = {Gumbsch, Thomas},
   year = {2017},
   url = {https://github.com/BorgwardtLab/PyChange},
@@ -171,7 +171,7 @@
 }
 
 @article{adams2007bayesian,
-  title={Bayesian Online Changepoint Detection},
+  title={{Bayesian} Online Changepoint Detection},
   author={Adams, Ryan Prescott and MacKay, David JC},
   journal={arXiv preprint arXiv:0710.3742},
   year={2007},
@@ -204,7 +204,7 @@
 }
 
 @article{hoffman2011no,
-  title={The No-U-Turn Sampler: Adaptively Setting Path Lengths in Hamiltonian Monte Carlo},
+  title={The {No-U-Turn Sampler}: Adaptively Setting Path Lengths in {Hamiltonian} {Monte} {Carlo}},
   author={Hoffman, Matthew D. and Gelman, Andrew},
   journal={Journal of Machine Learning Research},
   volume={15},
@@ -229,7 +229,7 @@
             Ribeiro, Ant{\^o}nio H. and Pedregosa, Fabian and
             {van Mulbregt}, Paul and {SciPy 1.0 Contributors}},
   title   = {{{SciPy} 1.0: Fundamental Algorithms for Scientific
-            Computing in Python}},
+            Computing in {Python}}},
   journal = {Nature Methods},
   year    = {2020},
   volume  = {17},
@@ -240,12 +240,12 @@
 
 @article{matplotlib,
   Author    = {Hunter, J. D.},
-  Title     = {Matplotlib: A 2D Graphics Environment},
+  Title     = {{Matplotlib}: A 2D Graphics Environment},
   Journal   = {Computing in Science \& Engineering},
   Volume    = {9},
   Number    = {3},
   Pages     = {90--95},
-  abstract  = {Matplotlib is a 2D graphics package used for Python for
+  abstract  = {{Matplotlib} is a 2D graphics package used for {Python} for
   application development, interactive scripting, and publication-quality
   image generation across user interfaces and operating systems.},
   publisher = {IEEE COMPUTER SOC},
