### 2026-01-09 4488ad1
Protect Hamiltonian Monte Carlo as complete algorithm name
Co-authored-by: abuzarmahmood <12436309+abuzarmahmood@users.noreply.github.com>
