+++ JOSS_submission/paper.bib
@@ -204,7 +204,7 @@
 }
 
 @article{hoffman2011no,
-  title={The {No-U-Turn Sampler}: Adaptively Setting Path Lengths in {Hamiltonian} {Monte} {Carlo}},
+  title={The {No-U-Turn Sampler}: Adaptively Setting Path Lengths in {Hamiltonian Monte Carlo}},
   author={Hoffman, Matthew D. and Gelman, Andrew},
   journal={Journal of Machine Learning Research},
   volume={15},
