### 2026-01-12 #8509 @danielskatz
The DOI resolves for me. However, the paper itself (https://www.theoj.org/joss-papers/joss.08509/10.21105.joss.08509.pdf) does not currently exist.
