### 2026-01-12 #8509 @arfon
Hrm. The file exists in `master`: https://github.com/openjournals/joss-papers/blob/master/joss.08509/10.21105.joss.08509.pdf

I wonder if the GitHub Pages build failed... (Checking)
